Network.AWS.S3.AbortMultipartUpload

Description

Aborts a multipart upload.

To verify that all parts have been removed, so you don't get charged for the part storage, you should call the List Parts operation and ensure the parts list is empty.

http://docs.aws.amazon.com/AmazonS3/latest/API/AbortMultipartUpload.html

Request

data AbortMultipartUpload

Request constructor

abortMultipartUpload

Arguments

:: Text
amuBucket

-> Text
amuKey

-> Text
amuUploadId

-> AbortMultipartUpload

AbortMultipartUpload constructor.

The fields accessible through corresponding lenses are:

amuBucket :: Text

amuKey :: Text

amuRequestPayer :: Maybe RequestPayer

amuUploadId :: Text

Request lenses

amuBucket :: Lens' AbortMultipartUpload Text

amuKey :: Lens' AbortMultipartUpload Text

amuRequestPayer :: Lens' AbortMultipartUpload (Maybe RequestPayer)

amuUploadId :: Lens' AbortMultipartUpload Text

Response

data AbortMultipartUploadResponse

Response constructor

abortMultipartUploadResponse :: AbortMultipartUploadResponse

AbortMultipartUploadResponse constructor.

The fields accessible through corresponding lenses are:

amurRequestCharged :: Maybe RequestCharged

Response lenses

amurRequestCharged :: Lens' AbortMultipartUploadResponse (Maybe RequestCharged)