Network.AWS.S3.CompleteMultipartUpload

Description

Completes a multipart upload by assembling previously uploaded parts.

Request

data CompleteMultipartUpload

Request constructor

completeMultipartUpload

Arguments

:: Text
cmuBucket

-> Text
cmuKey

-> Text
cmuUploadId

-> CompleteMultipartUpload

CompleteMultipartUpload constructor.

The fields accessible through corresponding lenses are:

cmuBucket :: Text

cmuKey :: Text

cmuMultipartUpload :: Maybe CompletedMultipartUpload

cmuUploadId :: Text

Request lenses

cmuBucket :: Lens' CompleteMultipartUpload Text

cmuKey :: Lens' CompleteMultipartUpload Text

cmuMultipartUpload :: Lens' CompleteMultipartUpload (Maybe CompletedMultipartUpload)

cmuUploadId :: Lens' CompleteMultipartUpload Text

Response

data CompleteMultipartUploadOutput

Response constructor

completeMultipartUploadOutput :: CompleteMultipartUploadOutput

CompleteMultipartUploadOutput constructor.

The fields accessible through corresponding lenses are:

cmuoBucket :: Maybe Text

cmuoETag :: Maybe Text

cmuoExpiration :: Maybe UTCTime

cmuoKey :: Maybe Text

cmuoLocation :: Maybe Text

cmuoServerSideEncryption :: Maybe Text

cmuoVersionId :: Maybe Text

Response lenses

cmuoBucket :: Lens' CompleteMultipartUploadOutput (Maybe Text)

cmuoETag :: Lens' CompleteMultipartUploadOutput (Maybe Text)

Entity tag of the object.

cmuoExpiration :: Lens' CompleteMultipartUploadOutput (Maybe UTCTime)

If the object expiration is configured, this will contain the expiration date (expiry-date) and rule ID (rule-id). The value of rule-id is URL encoded.

cmuoKey :: Lens' CompleteMultipartUploadOutput (Maybe Text)

cmuoLocation :: Lens' CompleteMultipartUploadOutput (Maybe Text)

cmuoServerSideEncryption :: Lens' CompleteMultipartUploadOutput (Maybe Text)

The Server-side encryption algorithm used when storing this object in S3.

cmuoVersionId :: Lens' CompleteMultipartUploadOutput (Maybe Text)

Version of the object.