Network.AWS.S3.CopyObject

Description

Creates a copy of an object that is already stored in Amazon S3.

Request

data CopyObject

Request constructor

copyObject

Arguments

:: Text
corBucket

-> Text
corCopySource

-> Text
corKey

-> CopyObject

CopyObject constructor.

The fields accessible through corresponding lenses are:

corACL :: Maybe Text

corBucket :: Text

corCacheControl :: Maybe Text

corContentDisposition :: Maybe Text

corContentEncoding :: Maybe Text

corContentLanguage :: Maybe Text

corContentType :: Maybe Text

corCopySource :: Text

corCopySourceIfMatch :: Maybe Text

corCopySourceIfModifiedSince :: Maybe UTCTime

corCopySourceIfNoneMatch :: Maybe Text

corCopySourceIfUnmodifiedSince :: Maybe UTCTime

corCopySourceSSECustomerAlgorithm :: Maybe Text

corCopySourceSSECustomerKey :: Maybe Text

corCopySourceSSECustomerKeyMD5 :: Maybe Text

corExpires :: Maybe UTCTime

corGrantFullControl :: Maybe Text

corGrantRead :: Maybe Text

corGrantReadACP :: Maybe Text

corGrantWriteACP :: Maybe Text

corKey :: Text

corMetadata :: HashMap Text Text

corMetadataDirective :: Maybe Text

corSSECustomerAlgorithm :: Maybe Text

corSSECustomerKey :: Maybe Text

corSSECustomerKeyMD5 :: Maybe Text

corServerSideEncryption :: Maybe Text

corStorageClass :: Maybe Text

corWebsiteRedirectLocation :: Maybe Text

Request lenses

corACL :: Lens' CopyObject (Maybe Text)

The canned ACL to apply to the object.

corBucket :: Lens' CopyObject Text

corCacheControl :: Lens' CopyObject (Maybe Text)

Specifies caching behavior along the request/reply chain.

corContentDisposition :: Lens' CopyObject (Maybe Text)

Specifies presentational information for the object.

corContentEncoding :: Lens' CopyObject (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

corContentLanguage :: Lens' CopyObject (Maybe Text)

The language the content is in.

corContentType :: Lens' CopyObject (Maybe Text)

A standard MIME type describing the format of the object data.

corCopySource :: Lens' CopyObject Text

The name of the source bucket and key name of the source object, separated by a slash (/). Must be URL-encoded.

corCopySourceIfMatch :: Lens' CopyObject (Maybe Text)

Copies the object if its entity tag (ETag) matches the specified tag.

corCopySourceIfModifiedSince :: Lens' CopyObject (Maybe UTCTime)

Copies the object if it has been modified since the specified time.

corCopySourceIfNoneMatch :: Lens' CopyObject (Maybe Text)

Copies the object if its entity tag (ETag) is different than the specified ETag.

corCopySourceIfUnmodifiedSince :: Lens' CopyObject (Maybe UTCTime)

Copies the object if it hasn't been modified since the specified time.

corCopySourceSSECustomerAlgorithm :: Lens' CopyObject (Maybe Text)

Specifies the algorithm to use when decrypting the source object (e.g., AES256).

corCopySourceSSECustomerKey :: Lens' CopyObject (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

corCopySourceSSECustomerKeyMD5 :: Lens' CopyObject (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

corExpires :: Lens' CopyObject (Maybe UTCTime)

The date and time at which the object is no longer cacheable.

corGrantFullControl :: Lens' CopyObject (Maybe Text)

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

corGrantRead :: Lens' CopyObject (Maybe Text)

Allows grantee to read the object data and its metadata.

corGrantReadACP :: Lens' CopyObject (Maybe Text)

Allows grantee to read the object ACL.

corGrantWriteACP :: Lens' CopyObject (Maybe Text)

Allows grantee to write the ACL for the applicable object.

corKey :: Lens' CopyObject Text

corMetadata :: Lens' CopyObject (HashMap Text Text)

A map of metadata to store with the object in S3.

corMetadataDirective :: Lens' CopyObject (Maybe Text)

Specifies whether the metadata is copied from the source object or replaced with metadata provided in the request.

corSSECustomerAlgorithm :: Lens' CopyObject (Maybe Text)

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

corSSECustomerKey :: Lens' CopyObject (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

corSSECustomerKeyMD5 :: Lens' CopyObject (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

corServerSideEncryption :: Lens' CopyObject (Maybe Text)

The Server-side encryption algorithm used when storing this object in S3.

corStorageClass :: Lens' CopyObject (Maybe Text)

The type of storage to use for the object. Defaults to STANDARD.

corWebsiteRedirectLocation :: Lens' CopyObject (Maybe Text)

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

Response

data CopyObjectOutput

Response constructor

copyObjectOutput :: CopyObjectOutput

CopyObjectOutput constructor.

The fields accessible through corresponding lenses are:

cooCopyObjectResult :: Maybe CopyObjectResult

cooCopySourceVersionId :: Maybe Text

cooExpiration :: Maybe UTCTime

cooSSECustomerAlgorithm :: Maybe Text

cooSSECustomerKeyMD5 :: Maybe Text

cooServerSideEncryption :: Maybe Text

Response lenses

cooCopyObjectResult :: Lens' CopyObjectOutput (Maybe CopyObjectResult)

cooCopySourceVersionId :: Lens' CopyObjectOutput (Maybe Text)

cooExpiration :: Lens' CopyObjectOutput (Maybe UTCTime)

If the object expiration is configured, the response includes this header.

cooSSECustomerAlgorithm :: Lens' CopyObjectOutput (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

cooSSECustomerKeyMD5 :: Lens' CopyObjectOutput (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

cooServerSideEncryption :: Lens' CopyObjectOutput (Maybe Text)

The Server-side encryption algorithm used when storing this object in S3.