Network.AWS.S3.CreateBucket

Description

Creates a new bucket.

http://docs.aws.amazon.com/AmazonS3/latest/API/CreateBucket.html

Request

data CreateBucket

Request constructor

createBucket

Arguments

:: Text
cbBucket

-> CreateBucket

CreateBucket constructor.

The fields accessible through corresponding lenses are:

cbACL :: Maybe BucketCannedACL

cbBucket :: Text

cbCreateBucketConfiguration :: Maybe CreateBucketConfiguration

cbGrantFullControl :: Maybe Text

cbGrantRead :: Maybe Text

cbGrantReadACP :: Maybe Text

cbGrantWrite :: Maybe Text

cbGrantWriteACP :: Maybe Text

Request lenses

cbACL :: Lens' CreateBucket (Maybe BucketCannedACL)

The canned ACL to apply to the bucket.

cbBucket :: Lens' CreateBucket Text

cbCreateBucketConfiguration :: Lens' CreateBucket (Maybe CreateBucketConfiguration)

cbGrantFullControl :: Lens' CreateBucket (Maybe Text)

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

cbGrantRead :: Lens' CreateBucket (Maybe Text)

Allows grantee to list the objects in the bucket.

cbGrantReadACP :: Lens' CreateBucket (Maybe Text)

Allows grantee to read the bucket ACL.

cbGrantWrite :: Lens' CreateBucket (Maybe Text)

Allows grantee to create, overwrite, and delete any object in the bucket.

cbGrantWriteACP :: Lens' CreateBucket (Maybe Text)

Allows grantee to write the ACL for the applicable bucket.

Response

data CreateBucketResponse

Response constructor

createBucketResponse :: CreateBucketResponse

CreateBucketResponse constructor.

The fields accessible through corresponding lenses are:

cbrLocation :: Maybe Text

Response lenses

cbrLocation :: Lens' CreateBucketResponse (Maybe Text)