Network.AWS.S3.CreateMultipartUpload

Description

Initiates a multipart upload and returns an upload ID. Note: After you initiate multipart upload and upload one or more parts, you must either complete or abort multipart upload in order to stop getting charged for storage of the uploaded parts. Only after you either complete or abort multipart upload, Amazon S3 frees up the parts storage and stops charging you for the parts storage.

Request

data CreateMultipartUpload

Request constructor

createMultipartUpload

Arguments

:: Text
cmur1Bucket

-> Text
cmur1Key

-> CreateMultipartUpload

CreateMultipartUpload constructor.

The fields accessible through corresponding lenses are:

cmur1ACL :: Maybe Text

cmur1Bucket :: Text

cmur1CacheControl :: Maybe Text

cmur1ContentDisposition :: Maybe Text

cmur1ContentEncoding :: Maybe Text

cmur1ContentLanguage :: Maybe Text

cmur1ContentType :: Maybe Text

cmur1Expires :: Maybe UTCTime

cmur1GrantFullControl :: Maybe Text

cmur1GrantRead :: Maybe Text

cmur1GrantReadACP :: Maybe Text

cmur1GrantWriteACP :: Maybe Text

cmur1Key :: Text

cmur1Metadata :: HashMap Text Text

cmur1SSECustomerAlgorithm :: Maybe Text

cmur1SSECustomerKey :: Maybe Text

cmur1SSECustomerKeyMD5 :: Maybe Text

cmur1ServerSideEncryption :: Maybe Text

cmur1StorageClass :: Maybe Text

cmur1WebsiteRedirectLocation :: Maybe Text

Request lenses

cmur1ACL :: Lens' CreateMultipartUpload (Maybe Text)

The canned ACL to apply to the object.

cmur1Bucket :: Lens' CreateMultipartUpload Text

cmur1CacheControl :: Lens' CreateMultipartUpload (Maybe Text)

Specifies caching behavior along the request/reply chain.

cmur1ContentDisposition :: Lens' CreateMultipartUpload (Maybe Text)

Specifies presentational information for the object.

cmur1ContentEncoding :: Lens' CreateMultipartUpload (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

cmur1ContentLanguage :: Lens' CreateMultipartUpload (Maybe Text)

The language the content is in.

cmur1ContentType :: Lens' CreateMultipartUpload (Maybe Text)

A standard MIME type describing the format of the object data.

cmur1Expires :: Lens' CreateMultipartUpload (Maybe UTCTime)

The date and time at which the object is no longer cacheable.

cmur1GrantFullControl :: Lens' CreateMultipartUpload (Maybe Text)

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

cmur1GrantRead :: Lens' CreateMultipartUpload (Maybe Text)

Allows grantee to read the object data and its metadata.

cmur1GrantReadACP :: Lens' CreateMultipartUpload (Maybe Text)

Allows grantee to read the object ACL.

cmur1GrantWriteACP :: Lens' CreateMultipartUpload (Maybe Text)

Allows grantee to write the ACL for the applicable object.

cmur1Key :: Lens' CreateMultipartUpload Text

cmur1Metadata :: Lens' CreateMultipartUpload (HashMap Text Text)

A map of metadata to store with the object in S3.

cmur1SSECustomerAlgorithm :: Lens' CreateMultipartUpload (Maybe Text)

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

cmur1SSECustomerKey :: Lens' CreateMultipartUpload (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

cmur1SSECustomerKeyMD5 :: Lens' CreateMultipartUpload (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

cmur1ServerSideEncryption :: Lens' CreateMultipartUpload (Maybe Text)

The Server-side encryption algorithm used when storing this object in S3.

cmur1StorageClass :: Lens' CreateMultipartUpload (Maybe Text)

The type of storage to use for the object. Defaults to STANDARD.

cmur1WebsiteRedirectLocation :: Lens' CreateMultipartUpload (Maybe Text)

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

Response

data CreateMultipartUploadOutput

Response constructor

createMultipartUploadOutput :: CreateMultipartUploadOutput

CreateMultipartUploadOutput constructor.

The fields accessible through corresponding lenses are:

cmuo1Bucket :: Maybe Text

cmuo1Key :: Maybe Text

cmuo1SSECustomerAlgorithm :: Maybe Text

cmuo1SSECustomerKeyMD5 :: Maybe Text

cmuo1ServerSideEncryption :: Maybe Text

cmuo1UploadId :: Maybe Text

Response lenses

cmuo1Bucket :: Lens' CreateMultipartUploadOutput (Maybe Text)

Name of the bucket to which the multipart upload was initiated.

cmuo1Key :: Lens' CreateMultipartUploadOutput (Maybe Text)

Object key for which the multipart upload was initiated.

cmuo1SSECustomerAlgorithm :: Lens' CreateMultipartUploadOutput (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

cmuo1SSECustomerKeyMD5 :: Lens' CreateMultipartUploadOutput (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

cmuo1ServerSideEncryption :: Lens' CreateMultipartUploadOutput (Maybe Text)

The Server-side encryption algorithm used when storing this object in S3.

cmuo1UploadId :: Lens' CreateMultipartUploadOutput (Maybe Text)

ID for the initiated multipart upload.