Network.AWS.S3.DeleteBucket

Description

Deletes the bucket. All objects (including all object versions and Delete Markers) in the bucket must be deleted before the bucket itself can be deleted.

Creating a Request

deleteBucket

Arguments

:: BucketName
dbBucket

-> DeleteBucket

Creates a value of DeleteBucket with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbBucket - Undocumented member.

data DeleteBucket

See: deleteBucket smart constructor.

Request Lenses

dbBucket :: Lens' DeleteBucket BucketName

Undocumented member.

Destructuring the Response

deleteBucketResponse :: DeleteBucketResponse

Creates a value of DeleteBucketResponse with the minimum fields required to make a request.

data DeleteBucketResponse

See: deleteBucketResponse smart constructor.