Network.AWS.S3.DeleteBucketCORS

Description

Deletes the cors configuration information set for the bucket.

Creating a Request

deleteBucketCORS

Arguments

:: BucketName
dbcBucket

-> DeleteBucketCORS

Creates a value of DeleteBucketCORS with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbcBucket - Undocumented member.

data DeleteBucketCORS

See: deleteBucketCORS smart constructor.

Request Lenses

dbcBucket :: Lens' DeleteBucketCORS BucketName

Undocumented member.

Destructuring the Response

deleteBucketCORSResponse :: DeleteBucketCORSResponse

Creates a value of DeleteBucketCORSResponse with the minimum fields required to make a request.

data DeleteBucketCORSResponse

See: deleteBucketCORSResponse smart constructor.