Network.AWS.S3.DeleteBucketInventoryConfiguration

Description

Deletes an inventory configuration (identified by the inventory ID) from the bucket.

Creating a Request

deleteBucketInventoryConfiguration

Arguments

:: BucketName
dbicBucket

-> Text
dbicId

-> DeleteBucketInventoryConfiguration

Creates a value of DeleteBucketInventoryConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbicBucket - The name of the bucket containing the inventory configuration to delete.

dbicId - The ID used to identify the inventory configuration.

data DeleteBucketInventoryConfiguration

See: deleteBucketInventoryConfiguration smart constructor.

Request Lenses

dbicBucket :: Lens' DeleteBucketInventoryConfiguration BucketName

The name of the bucket containing the inventory configuration to delete.

dbicId :: Lens' DeleteBucketInventoryConfiguration Text

The ID used to identify the inventory configuration.

Destructuring the Response

deleteBucketInventoryConfigurationResponse :: DeleteBucketInventoryConfigurationResponse

Creates a value of DeleteBucketInventoryConfigurationResponse with the minimum fields required to make a request.

data DeleteBucketInventoryConfigurationResponse

See: deleteBucketInventoryConfigurationResponse smart constructor.