Network.AWS.S3.DeleteBucketLifecycle

Description

Deletes the lifecycle configuration from the bucket.

Creating a Request

deleteBucketLifecycle

Arguments

:: BucketName
dblBucket

-> DeleteBucketLifecycle

Creates a value of DeleteBucketLifecycle with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dblBucket - Undocumented member.

data DeleteBucketLifecycle

See: deleteBucketLifecycle smart constructor.

Request Lenses

dblBucket :: Lens' DeleteBucketLifecycle BucketName

Undocumented member.

Destructuring the Response

deleteBucketLifecycleResponse :: DeleteBucketLifecycleResponse

Creates a value of DeleteBucketLifecycleResponse with the minimum fields required to make a request.

data DeleteBucketLifecycleResponse

See: deleteBucketLifecycleResponse smart constructor.