Network.AWS.S3.DeleteBucketMetricsConfiguration

Description

Deletes a metrics configuration (specified by the metrics configuration ID) from the bucket.

Creating a Request

deleteBucketMetricsConfiguration

Arguments

:: BucketName
dbmcBucket

-> Text
dbmcId

-> DeleteBucketMetricsConfiguration

Creates a value of DeleteBucketMetricsConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbmcBucket - The name of the bucket containing the metrics configuration to delete.

dbmcId - The ID used to identify the metrics configuration.

data DeleteBucketMetricsConfiguration

See: deleteBucketMetricsConfiguration smart constructor.

Request Lenses

dbmcBucket :: Lens' DeleteBucketMetricsConfiguration BucketName

The name of the bucket containing the metrics configuration to delete.

dbmcId :: Lens' DeleteBucketMetricsConfiguration Text

The ID used to identify the metrics configuration.

Destructuring the Response

deleteBucketMetricsConfigurationResponse :: DeleteBucketMetricsConfigurationResponse

Creates a value of DeleteBucketMetricsConfigurationResponse with the minimum fields required to make a request.

data DeleteBucketMetricsConfigurationResponse

See: deleteBucketMetricsConfigurationResponse smart constructor.