Network.AWS.S3.DeleteBucketPolicy

Description

Deletes the policy from the bucket.

Creating a Request

deleteBucketPolicy

Arguments

:: BucketName
dbpBucket

-> DeleteBucketPolicy

Creates a value of DeleteBucketPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dbpBucket - Undocumented member.

data DeleteBucketPolicy

See: deleteBucketPolicy smart constructor.

Request Lenses

dbpBucket :: Lens' DeleteBucketPolicy BucketName

Undocumented member.

Destructuring the Response

deleteBucketPolicyResponse :: DeleteBucketPolicyResponse

Creates a value of DeleteBucketPolicyResponse with the minimum fields required to make a request.

data DeleteBucketPolicyResponse

See: deleteBucketPolicyResponse smart constructor.