Network.AWS.S3.DeleteObject

Description

Removes the null version (if there is one) of an object and inserts a delete marker, which becomes the latest version of the object. If there isn't a null version, Amazon S3 does not remove any objects.

Request

data DeleteObject

Request constructor

deleteObject

Arguments

:: Text
do1Bucket

-> Text
do1Key

-> DeleteObject

DeleteObject constructor.

The fields accessible through corresponding lenses are:

do1Bucket :: Text

do1Key :: Text

do1MFA :: Maybe Text

do1VersionId :: Maybe Text

Request lenses

do1Bucket :: Lens' DeleteObject Text

do1Key :: Lens' DeleteObject Text

do1MFA :: Lens' DeleteObject (Maybe Text)

The concatenation of the authentication device's serial number, a space, and the value that is displayed on your authentication device.

do1VersionId :: Lens' DeleteObject (Maybe Text)

VersionId used to reference a specific version of the object.

Response

data DeleteObjectOutput

Response constructor

deleteObjectOutput :: DeleteObjectOutput

DeleteObjectOutput constructor.

The fields accessible through corresponding lenses are:

dooDeleteMarker :: Maybe Bool

dooVersionId :: Maybe Text

Response lenses

dooDeleteMarker :: Lens' DeleteObjectOutput (Maybe Bool)

Specifies whether the versioned object that was permanently deleted was (true) or was not (false) a delete marker.

dooVersionId :: Lens' DeleteObjectOutput (Maybe Text)

Returns the version ID of the delete marker created as a result of the DELETE operation.