Network.AWS.S3.DeleteObjects

Description

This operation enables you to delete multiple objects from a bucket using a single HTTP request. You may specify up to 1000 keys.

Request

data DeleteObjects

Request constructor

deleteObjects

Arguments

:: Text
dorBucket

-> Delete
dorDelete

-> DeleteObjects

DeleteObjects constructor.

The fields accessible through corresponding lenses are:

dorBucket :: Text

dorDelete :: Delete

dorMFA :: Maybe Text

Request lenses

dorBucket :: Lens' DeleteObjects Text

dorDelete :: Lens' DeleteObjects Delete

dorMFA :: Lens' DeleteObjects (Maybe Text)

The concatenation of the authentication device's serial number, a space, and the value that is displayed on your authentication device.

Response

data DeleteObjectsOutput

Response constructor

deleteObjectsOutput :: DeleteObjectsOutput

DeleteObjectsOutput constructor.

The fields accessible through corresponding lenses are:

dooDeleted :: '[S3ServiceError]'

dooErrors :: '[S3ServiceError]'

Response lenses

dooDeleted :: Lens' DeleteObjectsOutput [S3ServiceError]

dooErrors :: Lens' DeleteObjectsOutput [S3ServiceError]