Network.AWS.S3.GetBucketACL

Description

Gets the access control policy for the bucket.

Creating a Request

getBucketACL

Arguments

:: BucketName
gbaBucket

-> GetBucketACL

Creates a value of GetBucketACL with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbaBucket - Undocumented member.

data GetBucketACL

See: getBucketACL smart constructor.

Request Lenses

gbaBucket :: Lens' GetBucketACL BucketName

Undocumented member.

Destructuring the Response

getBucketACLResponse

Arguments

:: Int
gbarsResponseStatus

-> GetBucketACLResponse

Creates a value of GetBucketACLResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbarsGrants - A list of grants.

gbarsOwner - Undocumented member.

gbarsResponseStatus - -- | The response status code.

data GetBucketACLResponse

See: getBucketACLResponse smart constructor.

Response Lenses

gbarsGrants :: Lens' GetBucketACLResponse [Grant]

A list of grants.

gbarsOwner :: Lens' GetBucketACLResponse (Maybe Owner)

Undocumented member.

gbarsResponseStatus :: Lens' GetBucketACLResponse Int

- | The response status code.