Network.AWS.S3.GetBucketAnalyticsConfiguration

Description

Gets an analytics configuration for the bucket (specified by the analytics configuration ID).

Creating a Request

getBucketAnalyticsConfiguration

Arguments

:: BucketName
getBucket

-> Text
getId

-> GetBucketAnalyticsConfiguration

Creates a value of GetBucketAnalyticsConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

getBucket - The name of the bucket from which an analytics configuration is retrieved.

getId - The identifier used to represent an analytics configuration.

data GetBucketAnalyticsConfiguration

See: getBucketAnalyticsConfiguration smart constructor.

Request Lenses

getBucket :: Lens' GetBucketAnalyticsConfiguration BucketName

The name of the bucket from which an analytics configuration is retrieved.

getId :: Lens' GetBucketAnalyticsConfiguration Text

The identifier used to represent an analytics configuration.

Destructuring the Response

getBucketAnalyticsConfigurationResponse

Arguments

:: Int
gbacrsResponseStatus

-> GetBucketAnalyticsConfigurationResponse

Creates a value of GetBucketAnalyticsConfigurationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbacrsAnalyticsConfiguration - The configuration and any analyses for the analytics filter.

gbacrsResponseStatus - -- | The response status code.

data GetBucketAnalyticsConfigurationResponse

See: getBucketAnalyticsConfigurationResponse smart constructor.

Response Lenses

gbacrsAnalyticsConfiguration :: Lens' GetBucketAnalyticsConfigurationResponse (Maybe AnalyticsConfiguration)

The configuration and any analyses for the analytics filter.

gbacrsResponseStatus :: Lens' GetBucketAnalyticsConfigurationResponse Int

- | The response status code.