Network.AWS.S3.GetBucketCORS

Description

Returns the cors configuration for the bucket.

Creating a Request

getBucketCORS

Arguments

:: BucketName
gbcBucket

-> GetBucketCORS

Creates a value of GetBucketCORS with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbcBucket - Undocumented member.

data GetBucketCORS

See: getBucketCORS smart constructor.

Request Lenses

gbcBucket :: Lens' GetBucketCORS BucketName

Undocumented member.

Destructuring the Response

getBucketCORSResponse

Arguments

:: Int
gbcrsResponseStatus

-> GetBucketCORSResponse

Creates a value of GetBucketCORSResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbcrsCORSRules - Undocumented member.

gbcrsResponseStatus - -- | The response status code.

data GetBucketCORSResponse

See: getBucketCORSResponse smart constructor.

Response Lenses

gbcrsCORSRules :: Lens' GetBucketCORSResponse [CORSRule]

Undocumented member.

gbcrsResponseStatus :: Lens' GetBucketCORSResponse Int

- | The response status code.