Network.AWS.S3.GetBucketInventoryConfiguration

Description

Returns an inventory configuration (identified by the inventory ID) from the bucket.

Creating a Request

getBucketInventoryConfiguration

Arguments

:: BucketName
gbicBucket

-> Text
gbicId

-> GetBucketInventoryConfiguration

Creates a value of GetBucketInventoryConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbicBucket - The name of the bucket containing the inventory configuration to retrieve.

gbicId - The ID used to identify the inventory configuration.

data GetBucketInventoryConfiguration

See: getBucketInventoryConfiguration smart constructor.

Request Lenses

gbicBucket :: Lens' GetBucketInventoryConfiguration BucketName

The name of the bucket containing the inventory configuration to retrieve.

gbicId :: Lens' GetBucketInventoryConfiguration Text

The ID used to identify the inventory configuration.

Destructuring the Response

getBucketInventoryConfigurationResponse

Arguments

:: Int
gbicrsResponseStatus

-> GetBucketInventoryConfigurationResponse

Creates a value of GetBucketInventoryConfigurationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbicrsInventoryConfiguration - Specifies the inventory configuration.

gbicrsResponseStatus - -- | The response status code.

data GetBucketInventoryConfigurationResponse

See: getBucketInventoryConfigurationResponse smart constructor.

Response Lenses

gbicrsInventoryConfiguration :: Lens' GetBucketInventoryConfigurationResponse (Maybe InventoryConfiguration)

Specifies the inventory configuration.

gbicrsResponseStatus :: Lens' GetBucketInventoryConfigurationResponse Int

- | The response status code.