Network.AWS.S3.GetBucketLogging

Description

Returns the logging status of a bucket and the permissions users have to view and modify that status. To use GET, you must be the bucket owner.

Request

data GetBucketLogging

Request constructor

mkGetBucketLogging

Arguments

:: BucketName
gbl2Bucket

-> GetBucketLogging

Smart constructor for the minimum required parameters to construct a valid GetBucketLogging request.

The fields accessible through corresponding lenses are:

Bucket :: BucketName

Request lenses

gbl2Bucket :: Lens' GetBucketLogging BucketName

Response

data GetBucketLoggingResponse

Response constructor

mkGetBucketLoggingResponse :: GetBucketLoggingResponse

Smart constructor for the minimum required parameters to construct a valid GetBucketLoggingResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

LoggingEnabled :: Maybe LoggingEnabled

Response lenses

gblr1LoggingEnabled :: Lens' GetBucketLoggingResponse (Maybe LoggingEnabled)