Network.AWS.S3.GetBucketMetricsConfiguration

Description

Gets a metrics configuration (specified by the metrics configuration ID) from the bucket.

Creating a Request

getBucketMetricsConfiguration

Arguments

:: BucketName
gbmcBucket

-> Text
gbmcId

-> GetBucketMetricsConfiguration

Creates a value of GetBucketMetricsConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbmcBucket - The name of the bucket containing the metrics configuration to retrieve.

gbmcId - The ID used to identify the metrics configuration.

data GetBucketMetricsConfiguration

See: getBucketMetricsConfiguration smart constructor.

Request Lenses

gbmcBucket :: Lens' GetBucketMetricsConfiguration BucketName

The name of the bucket containing the metrics configuration to retrieve.

gbmcId :: Lens' GetBucketMetricsConfiguration Text

The ID used to identify the metrics configuration.

Destructuring the Response

getBucketMetricsConfigurationResponse

Arguments

:: Int
gbmcrsResponseStatus

-> GetBucketMetricsConfigurationResponse

Creates a value of GetBucketMetricsConfigurationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbmcrsMetricsConfiguration - Specifies the metrics configuration.

gbmcrsResponseStatus - -- | The response status code.

data GetBucketMetricsConfigurationResponse

See: getBucketMetricsConfigurationResponse smart constructor.

Response Lenses

gbmcrsMetricsConfiguration :: Lens' GetBucketMetricsConfigurationResponse (Maybe MetricsConfiguration)

Specifies the metrics configuration.

gbmcrsResponseStatus :: Lens' GetBucketMetricsConfigurationResponse Int

- | The response status code.