Network.AWS.S3.GetBucketReplication

Description

Returns the replication configuration of a bucket.

Creating a Request

getBucketReplication

Arguments

:: BucketName
gbrBucket

-> GetBucketReplication

Creates a value of GetBucketReplication with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbrBucket - Undocumented member.

data GetBucketReplication

See: getBucketReplication smart constructor.

Request Lenses

gbrBucket :: Lens' GetBucketReplication BucketName

Undocumented member.

Destructuring the Response

getBucketReplicationResponse

Arguments

:: Int
gbrrsResponseStatus

-> GetBucketReplicationResponse

Creates a value of GetBucketReplicationResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbrrsReplicationConfiguration - Undocumented member.

gbrrsResponseStatus - -- | The response status code.

data GetBucketReplicationResponse

See: getBucketReplicationResponse smart constructor.

Response Lenses

gbrrsReplicationConfiguration :: Lens' GetBucketReplicationResponse (Maybe ReplicationConfiguration)

Undocumented member.

gbrrsResponseStatus :: Lens' GetBucketReplicationResponse Int

- | The response status code.