Network.AWS.S3.GetBucketRequestPayment

Description

Returns the request payment configuration of a bucket.

Request

data GetBucketRequestPayment

Request constructor

getBucketRequestPayment

Arguments

:: BucketName
gbrpBucket

-> GetBucketRequestPayment

Smart constructor for the minimum required parameters to construct a valid GetBucketRequestPayment request.

The fields accessible through corresponding lenses are:

Bucket :: BucketName

Request lenses

gbrpBucket :: Lens' GetBucketRequestPayment BucketName

Response

data GetBucketRequestPaymentResponse

Response constructor

getBucketRequestPaymentResponse :: GetBucketRequestPaymentResponse

Smart constructor for the minimum required parameters to construct a valid GetBucketRequestPaymentResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Payer :: Maybe Payer

Response lenses

gbrprPayer :: Lens' GetBucketRequestPaymentResponse (Maybe Payer)

Specifies who pays for the download and request fees.