Network.AWS.S3.GetBucketVersioning

Description

Returns the versioning state of a bucket.

http://docs.aws.amazon.com/AmazonS3/latest/API/GetBucketVersioning.html

Request

data GetBucketVersioning

Request constructor

getBucketVersioning

Arguments

:: Text
gbvBucket

-> GetBucketVersioning

GetBucketVersioning constructor.

The fields accessible through corresponding lenses are:

gbvBucket :: Text

Request lenses

gbvBucket :: Lens' GetBucketVersioning Text

Response

data GetBucketVersioningResponse

Response constructor

getBucketVersioningResponse :: GetBucketVersioningResponse

GetBucketVersioningResponse constructor.

The fields accessible through corresponding lenses are:

gbvrMFADelete :: Maybe MFADeleteStatus

gbvrStatus :: Maybe BucketVersioningStatus

Response lenses

gbvrMFADelete :: Lens' GetBucketVersioningResponse (Maybe MFADeleteStatus)

Specifies whether MFA delete is enabled in the bucket versioning configuration. This element is only returned if the bucket has been configured with MFA delete. If the bucket has never been so configured, this element is not returned.

gbvrStatus :: Lens' GetBucketVersioningResponse (Maybe BucketVersioningStatus)

The versioning state of the bucket.