Network.AWS.S3.GetBucketWebsite

Description

Returns the website configuration for a bucket.

Creating a Request

getBucketWebsite

Arguments

:: BucketName
gbwBucket

-> GetBucketWebsite

Creates a value of GetBucketWebsite with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbwBucket - Undocumented member.

data GetBucketWebsite

See: getBucketWebsite smart constructor.

Request Lenses

gbwBucket :: Lens' GetBucketWebsite BucketName

Undocumented member.

Destructuring the Response

getBucketWebsiteResponse

Arguments

:: Int
gbwrsResponseStatus

-> GetBucketWebsiteResponse

Creates a value of GetBucketWebsiteResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gbwrsRedirectAllRequestsTo - Undocumented member.

gbwrsErrorDocument - Undocumented member.

gbwrsIndexDocument - Undocumented member.

gbwrsRoutingRules - Undocumented member.

gbwrsResponseStatus - -- | The response status code.

data GetBucketWebsiteResponse

See: getBucketWebsiteResponse smart constructor.

Response Lenses

gbwrsRedirectAllRequestsTo :: Lens' GetBucketWebsiteResponse (Maybe RedirectAllRequestsTo)

Undocumented member.

gbwrsErrorDocument :: Lens' GetBucketWebsiteResponse (Maybe ErrorDocument)

Undocumented member.

gbwrsIndexDocument :: Lens' GetBucketWebsiteResponse (Maybe IndexDocument)

Undocumented member.

gbwrsRoutingRules :: Lens' GetBucketWebsiteResponse [RoutingRule]

Undocumented member.

gbwrsResponseStatus :: Lens' GetBucketWebsiteResponse Int

- | The response status code.