Network.AWS.S3.GetObject

Description

Retrieves objects from Amazon S3.

Request

data GetObject

Request constructor

getObject

Arguments

:: Text
gorBucket

-> Text
gorKey

-> GetObject

GetObject constructor.

The fields accessible through corresponding lenses are:

gorBucket :: Text

gorIfMatch :: Maybe Text

gorIfModifiedSince :: Maybe UTCTime

gorIfNoneMatch :: Maybe Text

gorIfUnmodifiedSince :: Maybe UTCTime

gorKey :: Text

gorRange :: Maybe Text

gorResponseCacheControl :: Maybe Text

gorResponseContentDisposition :: Maybe Text

gorResponseContentEncoding :: Maybe Text

gorResponseContentLanguage :: Maybe Text

gorResponseContentType :: Maybe Text

gorResponseExpires :: Maybe UTCTime

gorSSECustomerAlgorithm :: Maybe Text

gorSSECustomerKey :: Maybe Text

gorSSECustomerKeyMD5 :: Maybe Text

gorVersionId :: Maybe Text

Request lenses

gorBucket :: Lens' GetObject Text

gorIfMatch :: Lens' GetObject (Maybe Text)

Return the object only if its entity tag (ETag) is the same as the one specified, otherwise return a 412 (precondition failed).

gorIfModifiedSince :: Lens' GetObject (Maybe UTCTime)

Return the object only if it has been modified since the specified time, otherwise return a 304 (not modified).

gorIfNoneMatch :: Lens' GetObject (Maybe Text)

Return the object only if its entity tag (ETag) is different from the one specified, otherwise return a 304 (not modified).

gorIfUnmodifiedSince :: Lens' GetObject (Maybe UTCTime)

Return the object only if it has not been modified since the specified time, otherwise return a 412 (precondition failed).

gorKey :: Lens' GetObject Text

gorRange :: Lens' GetObject (Maybe Text)

Downloads the specified range bytes of an object. For more information about the HTTP Range header, go to http:/www.w3.orgProtocolsrfc2616rfc2616-sec14.html#sec14.35.

gorResponseCacheControl :: Lens' GetObject (Maybe Text)

Sets the Cache-Control header of the response.

gorResponseContentDisposition :: Lens' GetObject (Maybe Text)

Sets the Content-Disposition header of the response.

gorResponseContentEncoding :: Lens' GetObject (Maybe Text)

Sets the Content-Encoding header of the response.

gorResponseContentLanguage :: Lens' GetObject (Maybe Text)

Sets the Content-Language header of the response.

gorResponseContentType :: Lens' GetObject (Maybe Text)

Sets the Content-Type header of the response.

gorResponseExpires :: Lens' GetObject (Maybe UTCTime)

Sets the Expires header of the response.

gorSSECustomerAlgorithm :: Lens' GetObject (Maybe Text)

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

gorSSECustomerKey :: Lens' GetObject (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

gorSSECustomerKeyMD5 :: Lens' GetObject (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

gorVersionId :: Lens' GetObject (Maybe Text)

VersionId used to reference a specific version of the object.

Response

data GetObjectOutput

Response constructor

getObjectOutput

Arguments

:: RsBody
gooBody

-> GetObjectOutput

GetObjectOutput constructor.

The fields accessible through corresponding lenses are:

gooAcceptRanges :: Maybe Text

gooBody :: RsBody

gooCacheControl :: Maybe Text

gooContentDisposition :: Maybe Text

gooContentEncoding :: Maybe Text

gooContentLanguage :: Maybe Text

gooContentLength :: Maybe Int

gooContentType :: Maybe Text

gooDeleteMarker :: Maybe Bool

gooETag :: Maybe Text

gooExpiration :: Maybe UTCTime

gooExpires :: Maybe UTCTime

gooLastModified :: Maybe UTCTime

gooMetadata :: HashMap Text Text

gooMissingMeta :: Maybe Int

gooRestore :: Maybe Text

gooSSECustomerAlgorithm :: Maybe Text

gooSSECustomerKeyMD5 :: Maybe Text

gooServerSideEncryption :: Maybe Text

gooVersionId :: Maybe Text

gooWebsiteRedirectLocation :: Maybe Text

Response lenses

gooAcceptRanges :: Lens' GetObjectOutput (Maybe Text)

gooBody :: Lens' GetObjectOutput RsBody

Object data.

gooCacheControl :: Lens' GetObjectOutput (Maybe Text)

Specifies caching behavior along the request/reply chain.

gooContentDisposition :: Lens' GetObjectOutput (Maybe Text)

Specifies presentational information for the object.

gooContentEncoding :: Lens' GetObjectOutput (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

gooContentLanguage :: Lens' GetObjectOutput (Maybe Text)

The language the content is in.

gooContentLength :: Lens' GetObjectOutput (Maybe Int)

Size of the body in bytes.

gooContentType :: Lens' GetObjectOutput (Maybe Text)

A standard MIME type describing the format of the object data.

gooDeleteMarker :: Lens' GetObjectOutput (Maybe Bool)

Specifies whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

gooETag :: Lens' GetObjectOutput (Maybe Text)

An ETag is an opaque identifier assigned by a web server to a specific version of a resource found at a URL.

gooExpiration :: Lens' GetObjectOutput (Maybe UTCTime)

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key value pairs providing object expiration information. The value of the rule-id is URL encoded.

gooExpires :: Lens' GetObjectOutput (Maybe UTCTime)

The date and time at which the object is no longer cacheable.

gooLastModified :: Lens' GetObjectOutput (Maybe UTCTime)

Last modified date of the object.

gooMetadata :: Lens' GetObjectOutput (HashMap Text Text)

A map of metadata to store with the object in S3.

gooMissingMeta :: Lens' GetObjectOutput (Maybe Int)

This is set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

gooRestore :: Lens' GetObjectOutput (Maybe Text)

Provides information about object restoration operation and expiration time of the restored object copy.

gooSSECustomerAlgorithm :: Lens' GetObjectOutput (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

gooSSECustomerKeyMD5 :: Lens' GetObjectOutput (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

gooServerSideEncryption :: Lens' GetObjectOutput (Maybe Text)

The Server-side encryption algorithm used when storing this object in S3.

gooVersionId :: Lens' GetObjectOutput (Maybe Text)

Version of the object.

gooWebsiteRedirectLocation :: Lens' GetObjectOutput (Maybe Text)

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.