Network.AWS.S3.GetObjectAcl

Description

Returns the access control list (ACL) of an object.

Request

data GetObjectAcl

Request constructor

getObjectAcl

Arguments

:: BucketName
goaBucket

-> ObjectKey
goaKey

-> GetObjectAcl

Smart constructor for the minimum required parameters to construct a valid GetObjectAcl request.

The fields accessible through corresponding lenses are:

Bucket :: BucketName

Key :: ObjectKey

VersionId :: Maybe ObjectVersionId

Request lenses

goaBucket :: Lens' GetObjectAcl BucketName

goaKey :: Lens' GetObjectAcl ObjectKey

goaVersionId :: Lens' GetObjectAcl (Maybe ObjectVersionId)

VersionId used to reference a specific version of the object.

Response

data GetObjectAclResponse

Response constructor

getObjectAclResponse :: GetObjectAclResponse

Smart constructor for the minimum required parameters to construct a valid GetObjectAclResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Owner :: Maybe Owner

Grants :: [Grant]

Response lenses

goarOwner :: Lens' GetObjectAclResponse (Maybe Owner)

goarGrants :: Lens' GetObjectAclResponse [Grant]

A list of grants.