Network.AWS.S3.GetObjectTorrent

Description

Return torrent files from a bucket.

Creating a Request

getObjectTorrent

Arguments

:: BucketName
gotBucket

-> ObjectKey
gotKey

-> GetObjectTorrent

Creates a value of GetObjectTorrent with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gotRequestPayer - Undocumented member.

gotBucket - Undocumented member.

gotKey - Undocumented member.

data GetObjectTorrent

See: getObjectTorrent smart constructor.

Request Lenses

gotRequestPayer :: Lens' GetObjectTorrent (Maybe RequestPayer)

Undocumented member.

gotBucket :: Lens' GetObjectTorrent BucketName

Undocumented member.

gotKey :: Lens' GetObjectTorrent ObjectKey

Undocumented member.

Destructuring the Response

getObjectTorrentResponse

Arguments

:: Int
getrsResponseStatus

-> RsBody
getrsBody

-> GetObjectTorrentResponse

Creates a value of GetObjectTorrentResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

getrsRequestCharged - Undocumented member.

getrsResponseStatus - -- | The response status code.

getrsBody - Undocumented member.

data GetObjectTorrentResponse

See: getObjectTorrentResponse smart constructor.

Response Lenses

getrsRequestCharged :: Lens' GetObjectTorrentResponse (Maybe RequestCharged)

Undocumented member.

getrsResponseStatus :: Lens' GetObjectTorrentResponse Int

- | The response status code.

getrsBody :: Lens' GetObjectTorrentResponse RsBody

Undocumented member.