Network.AWS.S3.HeadBucket

Description

This operation is useful to determine if a bucket exists and you have permission to access it.

http://docs.aws.amazon.com/AmazonS3/latest/API/HeadBucket.html

Request

data HeadBucket

Request constructor

headBucket

Arguments

:: Text
hbBucket

-> HeadBucket

HeadBucket constructor.

The fields accessible through corresponding lenses are:

hbBucket :: Text

Request lenses

hbBucket :: Lens' HeadBucket Text

Response

data HeadBucketResponse

Response constructor

headBucketResponse :: HeadBucketResponse

HeadBucketResponse constructor.