Network.AWS.S3.HeadObject

Description

The HEAD operation retrieves metadata from an object without returning the object itself. This operation is useful if you're only interested in an object's metadata. To use HEAD, you must have READ access to the object.

http://docs.aws.amazon.com/AmazonS3/latest/API/HeadObject.html

Request

data HeadObject

Request constructor

headObject

Arguments

:: Text
hoBucket

-> Text
hoKey

-> HeadObject

HeadObject constructor.

The fields accessible through corresponding lenses are:

hoBucket :: Text

hoIfMatch :: Maybe Text

hoIfModifiedSince :: Maybe UTCTime

hoIfNoneMatch :: Maybe Text

hoIfUnmodifiedSince :: Maybe UTCTime

hoKey :: Text

hoRange :: Maybe Text

hoRequestPayer :: Maybe RequestPayer

hoSSECustomerAlgorithm :: Maybe Text

hoSSECustomerKey :: Maybe Text

hoSSECustomerKeyMD5 :: Maybe Text

hoVersionId :: Maybe Text

Request lenses

hoBucket :: Lens' HeadObject Text

hoIfMatch :: Lens' HeadObject (Maybe Text)

Return the object only if its entity tag (ETag) is the same as the one specified, otherwise return a 412 (precondition failed).

hoIfModifiedSince :: Lens' HeadObject (Maybe UTCTime)

Return the object only if it has been modified since the specified time, otherwise return a 304 (not modified).

hoIfNoneMatch :: Lens' HeadObject (Maybe Text)

Return the object only if its entity tag (ETag) is different from the one specified, otherwise return a 304 (not modified).

hoIfUnmodifiedSince :: Lens' HeadObject (Maybe UTCTime)

Return the object only if it has not been modified since the specified time, otherwise return a 412 (precondition failed).

hoKey :: Lens' HeadObject Text

hoRange :: Lens' HeadObject (Maybe Text)

Downloads the specified range bytes of an object. For more information about the HTTP Range header, go to http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.35.

hoRequestPayer :: Lens' HeadObject (Maybe RequestPayer)

hoSSECustomerAlgorithm :: Lens' HeadObject (Maybe Text)

Specifies the algorithm to use to when encrypting the object (e.g., AES256, aws:kms).

hoSSECustomerKey :: Lens' HeadObject (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

hoSSECustomerKeyMD5 :: Lens' HeadObject (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

hoVersionId :: Lens' HeadObject (Maybe Text)

VersionId used to reference a specific version of the object.

Response

data HeadObjectResponse

Response constructor

headObjectResponse :: HeadObjectResponse

HeadObjectResponse constructor.

The fields accessible through corresponding lenses are:

horAcceptRanges :: Maybe Text

horCacheControl :: Maybe Text

horContentDisposition :: Maybe Text

horContentEncoding :: Maybe Text

horContentLanguage :: Maybe Text

horContentLength :: Maybe Int

horContentType :: Maybe Text

horDeleteMarker :: Maybe Bool

horETag :: Maybe Text

horExpiration :: Maybe Text

horExpires :: Maybe UTCTime

horLastModified :: Maybe UTCTime

horMetadata :: HashMap (CI Text) Text

horMissingMeta :: Maybe Int

horReplicationStatus :: Maybe ReplicationStatus

horRequestCharged :: Maybe RequestCharged

horRestore :: Maybe Text

horSSECustomerAlgorithm :: Maybe Text

horSSECustomerKeyMD5 :: Maybe Text

horSSEKMSKeyId :: Maybe Text

horServerSideEncryption :: Maybe ServerSideEncryption

horVersionId :: Maybe Text

horWebsiteRedirectLocation :: Maybe Text

Response lenses

horAcceptRanges :: Lens' HeadObjectResponse (Maybe Text)

horCacheControl :: Lens' HeadObjectResponse (Maybe Text)

Specifies caching behavior along the request/reply chain.

horContentDisposition :: Lens' HeadObjectResponse (Maybe Text)

Specifies presentational information for the object.

horContentEncoding :: Lens' HeadObjectResponse (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

horContentLanguage :: Lens' HeadObjectResponse (Maybe Text)

The language the content is in.

horContentLength :: Lens' HeadObjectResponse (Maybe Int)

Size of the body in bytes.

horContentType :: Lens' HeadObjectResponse (Maybe Text)

A standard MIME type describing the format of the object data.

horDeleteMarker :: Lens' HeadObjectResponse (Maybe Bool)

Specifies whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

horETag :: Lens' HeadObjectResponse (Maybe Text)

An ETag is an opaque identifier assigned by a web server to a specific version of a resource found at a URL

horExpiration :: Lens' HeadObjectResponse (Maybe Text)

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key value pairs providing object expiration information. The value of the rule-id is URL encoded.

horExpires :: Lens' HeadObjectResponse (Maybe UTCTime)

The date and time at which the object is no longer cacheable.

horLastModified :: Lens' HeadObjectResponse (Maybe UTCTime)

Last modified date of the object

horMetadata :: Lens' HeadObjectResponse (HashMap (CI Text) Text)

A map of metadata to store with the object in S3.

horMissingMeta :: Lens' HeadObjectResponse (Maybe Int)

This is set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

horReplicationStatus :: Lens' HeadObjectResponse (Maybe ReplicationStatus)

horRequestCharged :: Lens' HeadObjectResponse (Maybe RequestCharged)

horRestore :: Lens' HeadObjectResponse (Maybe Text)

Provides information about object restoration operation and expiration time of the restored object copy.

horSSECustomerAlgorithm :: Lens' HeadObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

horSSECustomerKeyMD5 :: Lens' HeadObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

horSSEKMSKeyId :: Lens' HeadObjectResponse (Maybe Text)

If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

horServerSideEncryption :: Lens' HeadObjectResponse (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

horVersionId :: Lens' HeadObjectResponse (Maybe Text)

Version of the object.

horWebsiteRedirectLocation :: Lens' HeadObjectResponse (Maybe Text)

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.