Network.AWS.S3.ListBucketMetricsConfigurations

Description

Lists the metrics configurations for the bucket.

Creating a Request

listBucketMetricsConfigurations

Arguments

:: BucketName
lbmcBucket

-> ListBucketMetricsConfigurations

Creates a value of ListBucketMetricsConfigurations with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbmcContinuationToken - The marker that is used to continue a metrics configuration listing that has been truncated. Use the NextContinuationToken from a previously truncated list response to continue the listing. The continuation token is an opaque value that Amazon S3 understands.

lbmcBucket - The name of the bucket containing the metrics configurations to retrieve.

data ListBucketMetricsConfigurations

See: listBucketMetricsConfigurations smart constructor.

Request Lenses

lbmcContinuationToken :: Lens' ListBucketMetricsConfigurations (Maybe Text)

The marker that is used to continue a metrics configuration listing that has been truncated. Use the NextContinuationToken from a previously truncated list response to continue the listing. The continuation token is an opaque value that Amazon S3 understands.

lbmcBucket :: Lens' ListBucketMetricsConfigurations BucketName

The name of the bucket containing the metrics configurations to retrieve.

Destructuring the Response

listBucketMetricsConfigurationsResponse

Arguments

:: Int
lbmcrsResponseStatus

-> ListBucketMetricsConfigurationsResponse

Creates a value of ListBucketMetricsConfigurationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbmcrsContinuationToken - The marker that is used as a starting point for this metrics configuration list response. This value is present if it was sent in the request.

lbmcrsMetricsConfigurationList - The list of metrics configurations for a bucket.

lbmcrsNextContinuationToken - The marker used to continue a metrics configuration listing that has been truncated. Use the NextContinuationToken from a previously truncated list response to continue the listing. The continuation token is an opaque value that Amazon S3 understands.

lbmcrsIsTruncated - Indicates whether the returned list of metrics configurations is complete. A value of true indicates that the list is not complete and the NextContinuationToken will be provided for a subsequent request.

lbmcrsResponseStatus - -- | The response status code.

data ListBucketMetricsConfigurationsResponse

See: listBucketMetricsConfigurationsResponse smart constructor.

Response Lenses

lbmcrsContinuationToken :: Lens' ListBucketMetricsConfigurationsResponse (Maybe Text)

The marker that is used as a starting point for this metrics configuration list response. This value is present if it was sent in the request.

lbmcrsMetricsConfigurationList :: Lens' ListBucketMetricsConfigurationsResponse [MetricsConfiguration]

The list of metrics configurations for a bucket.

lbmcrsNextContinuationToken :: Lens' ListBucketMetricsConfigurationsResponse (Maybe Text)

The marker used to continue a metrics configuration listing that has been truncated. Use the NextContinuationToken from a previously truncated list response to continue the listing. The continuation token is an opaque value that Amazon S3 understands.

lbmcrsIsTruncated :: Lens' ListBucketMetricsConfigurationsResponse (Maybe Bool)

Indicates whether the returned list of metrics configurations is complete. A value of true indicates that the list is not complete and the NextContinuationToken will be provided for a subsequent request.

lbmcrsResponseStatus :: Lens' ListBucketMetricsConfigurationsResponse Int

- | The response status code.