Network.AWS.S3.ListBuckets

Description

Returns a list of all buckets owned by the authenticated sender of the request.

Creating a Request

listBuckets :: ListBuckets

Creates a value of ListBuckets with the minimum fields required to make a request.

data ListBuckets

See: listBuckets smart constructor.

Destructuring the Response

listBucketsResponse

Arguments

:: Int
lbrsResponseStatus

-> ListBucketsResponse

Creates a value of ListBucketsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbrsBuckets - Undocumented member.

lbrsOwner - Undocumented member.

lbrsResponseStatus - -- | The response status code.

data ListBucketsResponse

See: listBucketsResponse smart constructor.

Response Lenses

lbrsBuckets :: Lens' ListBucketsResponse [Bucket]

Undocumented member.

lbrsOwner :: Lens' ListBucketsResponse (Maybe Owner)

Undocumented member.

lbrsResponseStatus :: Lens' ListBucketsResponse Int

- | The response status code.