Network.AWS.S3.ListMultipartUploads

Description

This operation lists in-progress multipart uploads.

Request

data ListMultipartUploads

Request constructor

listMultipartUploads

Arguments

:: Text
lmurBucket

-> ListMultipartUploads

ListMultipartUploads constructor.

The fields accessible through corresponding lenses are:

lmurBucket :: Text

lmurDelimiter :: Maybe Text

lmurEncodingType :: Maybe Text

lmurKeyMarker :: Maybe Text

lmurMaxUploads :: Maybe Int

lmurPrefix :: Maybe Text

lmurUploadIdMarker :: Maybe Text

Request lenses

lmurBucket :: Lens' ListMultipartUploads Text

lmurDelimiter :: Lens' ListMultipartUploads (Maybe Text)

Character you use to group keys.

lmurEncodingType :: Lens' ListMultipartUploads (Maybe Text)

lmurKeyMarker :: Lens' ListMultipartUploads (Maybe Text)

Together with upload-id-marker, this parameter specifies the multipart upload after which listing should begin.

lmurMaxUploads :: Lens' ListMultipartUploads (Maybe Int)

Sets the maximum number of multipart uploads, from 1 to 1,000, to return in the response body. 1,000 is the maximum number of uploads that can be returned in a response.

lmurPrefix :: Lens' ListMultipartUploads (Maybe Text)

Lists in-progress uploads only for those keys that begin with the specified prefix.

lmurUploadIdMarker :: Lens' ListMultipartUploads (Maybe Text)

Together with key-marker, specifies the multipart upload after which listing should begin. If key-marker is not specified, the upload-id-marker parameter is ignored.

Response

data ListMultipartUploadsOutput

Response constructor

listMultipartUploadsOutput :: ListMultipartUploadsOutput

ListMultipartUploadsOutput constructor.

The fields accessible through corresponding lenses are:

lmuoBucket :: Maybe Text

lmuoCommonPrefixes :: '[CommonPrefix]'

lmuoEncodingType :: Maybe Text

lmuoIsTruncated :: Maybe Bool

lmuoKeyMarker :: Maybe Text

lmuoMaxUploads :: Maybe Int

lmuoNextKeyMarker :: Maybe Text

lmuoNextUploadIdMarker :: Maybe Text

lmuoPrefix :: Maybe Text

lmuoUploadIdMarker :: Maybe Text

lmuoUploads :: '[MultipartUpload]'

Response lenses

lmuoBucket :: Lens' ListMultipartUploadsOutput (Maybe Text)

Name of the bucket to which the multipart upload was initiated.

lmuoCommonPrefixes :: Lens' ListMultipartUploadsOutput [CommonPrefix]

lmuoEncodingType :: Lens' ListMultipartUploadsOutput (Maybe Text)

Encoding type used by Amazon S3 to encode object keys in the response.

lmuoIsTruncated :: Lens' ListMultipartUploadsOutput (Maybe Bool)

Indicates whether the returned list of multipart uploads is truncated. A value of true indicates that the list was truncated. The list can be truncated if the number of multipart uploads exceeds the limit allowed or specified by max uploads.

lmuoKeyMarker :: Lens' ListMultipartUploadsOutput (Maybe Text)

The key at or after which the listing began.

lmuoMaxUploads :: Lens' ListMultipartUploadsOutput (Maybe Int)

Maximum number of multipart uploads that could have been included in the response.

lmuoNextKeyMarker :: Lens' ListMultipartUploadsOutput (Maybe Text)

When a list is truncated, this element specifies the value that should be used for the key-marker request parameter in a subsequent request.

lmuoNextUploadIdMarker :: Lens' ListMultipartUploadsOutput (Maybe Text)

When a list is truncated, this element specifies the value that should be used for the upload-id-marker request parameter in a subsequent request.

lmuoPrefix :: Lens' ListMultipartUploadsOutput (Maybe Text)

When a prefix is provided in the request, this field contains the specified prefix. The result contains only keys starting with the specified prefix.

lmuoUploadIdMarker :: Lens' ListMultipartUploadsOutput (Maybe Text)

Upload ID after which listing began.

lmuoUploads :: Lens' ListMultipartUploadsOutput [MultipartUpload]