Network.AWS.S3.ListObjectVersions

Description

Returns metadata about all of the versions of objects in a bucket.

Request

data ListObjectVersions

Request alias

type GetBucketObjectVersions = ListObjectVersions

Request constructor

listObjectVersions

Arguments

:: Text
lovBucket

-> ListObjectVersions

Smart constructor for the minimum required parameters to construct a valid ListObjectVersions request.

The fields accessible through corresponding lenses are:

Bucket :: Text

Delimiter :: Maybe Text

EncodingType :: Maybe EncodingType

KeyMarker :: Maybe Text

MaxKeys :: Maybe Integer

Prefix :: Maybe Text

VersionIdMarker :: Maybe Text

Request lenses

lovBucket :: Lens' ListObjectVersions Text

lovDelimiter :: Lens' ListObjectVersions (Maybe Text)

A delimiter is a character you use to group keys.

lovEncodingType :: Lens' ListObjectVersions (Maybe EncodingType)

Requests Amazon S3 to encode the object keys in the response and specifies the encoding method to use. An object key may contain any Unicode character; however, XML 1.0 parser cannot parse some characters, such as characters with an ASCII value from 0 to 10. For characters that are not supported in XML 1.0, you can add this parameter to request that Amazon S3 encode the keys in the response.

lovKeyMarker :: Lens' ListObjectVersions (Maybe Text)

Specifies the key to start with when listing objects in a bucket.

lovMaxKeys :: Lens' ListObjectVersions (Maybe Integer)

Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.

lovPrefix :: Lens' ListObjectVersions (Maybe Text)

Limits the response to keys that begin with the specified prefix.

lovVersionIdMarker :: Lens' ListObjectVersions (Maybe Text)

Specifies the object version you want to start listing from.

Response

data ListObjectVersionsResponse

Response constructor

listObjectVersionsResponse

Arguments

:: Bool
lovrIsTruncated

-> ListObjectVersionsResponse

Smart constructor for the minimum required parameters to construct a valid ListObjectVersionsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

IsTruncated :: Bool

KeyMarker :: Maybe Text

VersionIdMarker :: Maybe Text

NextKeyMarker :: Maybe Text

NextVersionIdMarker :: Maybe Text

Versions :: [ObjectVersion]

DeleteMarkers :: [DeleteMarkerEntry]

Name :: Maybe Text

Prefix :: Maybe Text

MaxKeys :: Maybe Integer

CommonPrefixes :: [CommonPrefix]

EncodingType :: Maybe EncodingType

Response lenses

lovrIsTruncated :: Lens' ListObjectVersionsResponse Bool

A flag that indicates whether or not Amazon S3 returned all of the results that satisfied the search criteria. If your results were truncated, you can make a follow-up paginated request using the NextKeyMarker and NextVersionIdMarker response parameters as a starting place in another request to return the rest of the results.

lovrKeyMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

Marks the last Key returned in a truncated response.

lovrVersionIdMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

lovrNextKeyMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

Use this value for the key marker request parameter in a subsequent request.

lovrNextVersionIdMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

Use this value for the next version id marker parameter in a subsequent request.

lovrVersions :: Lens' ListObjectVersionsResponse [ObjectVersion]

lovrDeleteMarkers :: Lens' ListObjectVersionsResponse [DeleteMarkerEntry]

lovrName :: Lens' ListObjectVersionsResponse (Maybe Text)

lovrPrefix :: Lens' ListObjectVersionsResponse (Maybe Text)

lovrMaxKeys :: Lens' ListObjectVersionsResponse (Maybe Integer)

lovrCommonPrefixes :: Lens' ListObjectVersionsResponse [CommonPrefix]

lovrEncodingType :: Lens' ListObjectVersionsResponse (Maybe EncodingType)

Encoding type used by Amazon S3 to encode object keys in the response.