Network.AWS.S3.ListObjects

Description

Returns some or all (up to 1000) of the objects in a bucket. You can use the request parameters as selection criteria to return a subset of the objects in a bucket.

Request

data ListObjects

Request constructor

listObjects

Arguments

:: Text
lorBucket

-> ListObjects

ListObjects constructor.

The fields accessible through corresponding lenses are:

lorBucket :: Text

lorDelimiter :: Maybe Text

lorEncodingType :: Maybe Text

lorMarker :: Maybe Text

lorMaxKeys :: Maybe Int

lorPrefix :: Maybe Text

Request lenses

lorBucket :: Lens' ListObjects Text

lorDelimiter :: Lens' ListObjects (Maybe Text)

A delimiter is a character you use to group keys.

lorEncodingType :: Lens' ListObjects (Maybe Text)

lorMarker :: Lens' ListObjects (Maybe Text)

Specifies the key to start with when listing objects in a bucket.

lorMaxKeys :: Lens' ListObjects (Maybe Int)

Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.

lorPrefix :: Lens' ListObjects (Maybe Text)

Limits the response to keys that begin with the specified prefix.

Response

data ListObjectsOutput

Response constructor

listObjectsOutput :: ListObjectsOutput

ListObjectsOutput constructor.

The fields accessible through corresponding lenses are:

looCommonPrefixes :: '[CommonPrefix]'

looContents :: '[Object]'

looEncodingType :: Maybe Text

looIsTruncated :: Maybe Bool

looMarker :: Maybe Text

looMaxKeys :: Maybe Int

looName :: Maybe Text

looNextMarker :: Maybe Text

looPrefix :: Maybe Text

Response lenses

looCommonPrefixes :: Lens' ListObjectsOutput [CommonPrefix]

looContents :: Lens' ListObjectsOutput [Object]

looEncodingType :: Lens' ListObjectsOutput (Maybe Text)

Encoding type used by Amazon S3 to encode object keys in the response.

looIsTruncated :: Lens' ListObjectsOutput (Maybe Bool)

A flag that indicates whether or not Amazon S3 returned all of the results that satisfied the search criteria.

looMarker :: Lens' ListObjectsOutput (Maybe Text)

looMaxKeys :: Lens' ListObjectsOutput (Maybe Int)

looName :: Lens' ListObjectsOutput (Maybe Text)

looNextMarker :: Lens' ListObjectsOutput (Maybe Text)

When response is truncated (the IsTruncated element value in the response is true), you can use the key name in this field as marker in the subsequent request to get next set of objects. Amazon S3 lists objects in alphabetical order Note: This element is returned only if you have delimiter request parameter specified. If response does not include the NextMaker and it is truncated, you can use the value of the last Key in the response as the marker in the subsequent request to get the next set of object keys.

looPrefix :: Lens' ListObjectsOutput (Maybe Text)