Network.AWS.S3.ListObjectsV

Description

Returns some or all (up to 1000) of the objects in a bucket. You can use the request parameters as selection criteria to return a subset of the objects in a bucket. Note: ListObjectsV2 is the revised List Objects API and we recommend you use this revised API for new application development.

Creating a Request

listObjectsV

Arguments

:: BucketName
lBucket

-> ListObjectsV

Creates a value of ListObjectsV with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lStartAfter - StartAfter is where you want Amazon S3 to start listing from. Amazon S3 starts listing after this specified key. StartAfter can be any key in the bucket

lContinuationToken - ContinuationToken indicates Amazon S3 that the list is being continued on this bucket with a token. ContinuationToken is obfuscated and is not a real key

lFetchOwner - The owner field is not present in listV2 by default, if you want to return owner field with each key in the result then set the fetch owner field to true

lPrefix - Limits the response to keys that begin with the specified prefix.

lEncodingType - Encoding type used by Amazon S3 to encode object keys in the response.

lRequestPayer - Confirms that the requester knows that she or he will be charged for the list objects request in V2 style. Bucket owners need not specify this parameter in their requests.

lMaxKeys - Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.

lDelimiter - A delimiter is a character you use to group keys.

lBucket - Name of the bucket to list.

data ListObjectsV

See: listObjectsV smart constructor.

Request Lenses

lStartAfter :: Lens' ListObjectsV (Maybe Text)

StartAfter is where you want Amazon S3 to start listing from. Amazon S3 starts listing after this specified key. StartAfter can be any key in the bucket

lContinuationToken :: Lens' ListObjectsV (Maybe Text)

ContinuationToken indicates Amazon S3 that the list is being continued on this bucket with a token. ContinuationToken is obfuscated and is not a real key

lFetchOwner :: Lens' ListObjectsV (Maybe Bool)

The owner field is not present in listV2 by default, if you want to return owner field with each key in the result then set the fetch owner field to true

lPrefix :: Lens' ListObjectsV (Maybe Text)

Limits the response to keys that begin with the specified prefix.

lEncodingType :: Lens' ListObjectsV (Maybe EncodingType)

Encoding type used by Amazon S3 to encode object keys in the response.

lRequestPayer :: Lens' ListObjectsV (Maybe RequestPayer)

Confirms that the requester knows that she or he will be charged for the list objects request in V2 style. Bucket owners need not specify this parameter in their requests.

lMaxKeys :: Lens' ListObjectsV (Maybe Int)

Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.

lDelimiter :: Lens' ListObjectsV (Maybe Delimiter)

A delimiter is a character you use to group keys.

lBucket :: Lens' ListObjectsV BucketName

Name of the bucket to list.

Destructuring the Response

listObjectsVResponse

Arguments

:: Int
lrsResponseStatus

-> ListObjectsVResponse

Creates a value of ListObjectsVResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lrsStartAfter - StartAfter is where you want Amazon S3 to start listing from. Amazon S3 starts listing after this specified key. StartAfter can be any key in the bucket

lrsKeyCount - KeyCount is the number of keys returned with this request. KeyCount will always be less than equals to MaxKeys field. Say you ask for 50 keys, your result will include less than equals 50 keys

lrsContents - Metadata about each object returned.

lrsContinuationToken - ContinuationToken indicates Amazon S3 that the list is being continued on this bucket with a token. ContinuationToken is obfuscated and is not a real key

lrsPrefix - Limits the response to keys that begin with the specified prefix.

lrsCommonPrefixes - CommonPrefixes contains all (if there are any) keys between Prefix and the next occurrence of the string specified by delimiter

lrsEncodingType - Encoding type used by Amazon S3 to encode object keys in the response.

lrsName - Name of the bucket to list.

lrsNextContinuationToken - NextContinuationToken is sent when isTruncated is true which means there are more keys in the bucket that can be listed. The next list requests to Amazon S3 can be continued with this NextContinuationToken. NextContinuationToken is obfuscated and is not a real key

lrsMaxKeys - Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.

lrsIsTruncated - A flag that indicates whether or not Amazon S3 returned all of the results that satisfied the search criteria.

lrsDelimiter - A delimiter is a character you use to group keys.

lrsResponseStatus - -- | The response status code.

data ListObjectsVResponse

See: listObjectsVResponse smart constructor.

Response Lenses

lrsStartAfter :: Lens' ListObjectsVResponse (Maybe Text)

StartAfter is where you want Amazon S3 to start listing from. Amazon S3 starts listing after this specified key. StartAfter can be any key in the bucket

lrsKeyCount :: Lens' ListObjectsVResponse (Maybe Int)

KeyCount is the number of keys returned with this request. KeyCount will always be less than equals to MaxKeys field. Say you ask for 50 keys, your result will include less than equals 50 keys

lrsContents :: Lens' ListObjectsVResponse [Object]

Metadata about each object returned.

lrsContinuationToken :: Lens' ListObjectsVResponse (Maybe Text)

ContinuationToken indicates Amazon S3 that the list is being continued on this bucket with a token. ContinuationToken is obfuscated and is not a real key

lrsPrefix :: Lens' ListObjectsVResponse (Maybe Text)

Limits the response to keys that begin with the specified prefix.

lrsCommonPrefixes :: Lens' ListObjectsVResponse [CommonPrefix]

CommonPrefixes contains all (if there are any) keys between Prefix and the next occurrence of the string specified by delimiter

lrsEncodingType :: Lens' ListObjectsVResponse (Maybe EncodingType)

Encoding type used by Amazon S3 to encode object keys in the response.

lrsName :: Lens' ListObjectsVResponse (Maybe BucketName)

Name of the bucket to list.

lrsNextContinuationToken :: Lens' ListObjectsVResponse (Maybe Text)

NextContinuationToken is sent when isTruncated is true which means there are more keys in the bucket that can be listed. The next list requests to Amazon S3 can be continued with this NextContinuationToken. NextContinuationToken is obfuscated and is not a real key

lrsMaxKeys :: Lens' ListObjectsVResponse (Maybe Int)

Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.

lrsIsTruncated :: Lens' ListObjectsVResponse (Maybe Bool)

A flag that indicates whether or not Amazon S3 returned all of the results that satisfied the search criteria.

lrsDelimiter :: Lens' ListObjectsVResponse (Maybe Delimiter)

A delimiter is a character you use to group keys.

lrsResponseStatus :: Lens' ListObjectsVResponse Int

- | The response status code.