Network.AWS.S3.ListParts

Description

Lists the parts that have been uploaded for a specific multipart upload.

Request

data ListParts

Request constructor

listParts

Arguments

:: Text
lpBucket

-> Text
lpKey

-> Text
lpUploadId

-> ListParts

ListParts constructor.

The fields accessible through corresponding lenses are:

lpBucket :: Text

lpKey :: Text

lpMaxParts :: Maybe Int

lpPartNumberMarker :: Maybe Int

lpUploadId :: Text

Request lenses

lpBucket :: Lens' ListParts Text

lpKey :: Lens' ListParts Text

lpMaxParts :: Lens' ListParts (Maybe Int)

Sets the maximum number of parts to return.

lpPartNumberMarker :: Lens' ListParts (Maybe Int)

Specifies the part after which listing should begin. Only parts with higher part numbers will be listed.

lpUploadId :: Lens' ListParts Text

Upload ID identifying the multipart upload whose parts are being listed.

Response

data ListPartsOutput

Response constructor

listPartsOutput :: ListPartsOutput

ListPartsOutput constructor.

The fields accessible through corresponding lenses are:

lpoBucket :: Maybe Text

lpoInitiator :: Maybe Initiator

lpoIsTruncated :: Maybe Bool

lpoKey :: Maybe Text

lpoMaxParts :: Maybe Int

lpoNextPartNumberMarker :: Maybe Int

lpoOwner :: Maybe Owner

lpoPartNumberMarker :: Maybe Int

lpoParts :: [Part]

lpoStorageClass :: Maybe Text

lpoUploadId :: Maybe Text

Response lenses

lpoBucket :: Lens' ListPartsOutput (Maybe Text)

Name of the bucket to which the multipart upload was initiated.

lpoInitiator :: Lens' ListPartsOutput (Maybe Initiator)

Identifies who initiated the multipart upload.

lpoIsTruncated :: Lens' ListPartsOutput (Maybe Bool)

Indicates whether the returned list of parts is truncated.

lpoKey :: Lens' ListPartsOutput (Maybe Text)

Object key for which the multipart upload was initiated.

lpoMaxParts :: Lens' ListPartsOutput (Maybe Int)

Maximum number of parts that were allowed in the response.

lpoNextPartNumberMarker :: Lens' ListPartsOutput (Maybe Int)

When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

lpoOwner :: Lens' ListPartsOutput (Maybe Owner)

lpoPartNumberMarker :: Lens' ListPartsOutput (Maybe Int)

Part number after which listing begins.

lpoParts :: Lens' ListPartsOutput [Part]

lpoStorageClass :: Lens' ListPartsOutput (Maybe Text)

The class of storage used to store the object.

lpoUploadId :: Lens' ListPartsOutput (Maybe Text)

Upload ID identifying the multipart upload whose parts are being listed.