Network.AWS.S3.PutBucketACL

Description

Sets the permissions on a bucket using access control lists (ACL).

Creating a Request

putBucketACL

Arguments

:: BucketName
pbaBucket

-> PutBucketACL

Creates a value of PutBucketACL with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pbaGrantReadACP - Allows grantee to read the bucket ACL.

pbaGrantWriteACP - Allows grantee to write the ACL for the applicable bucket.

pbaGrantRead - Allows grantee to list the objects in the bucket.

pbaGrantFullControl - Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

pbaContentMD5 - Undocumented member.

pbaAccessControlPolicy - Undocumented member.

pbaGrantWrite - Allows grantee to create, overwrite, and delete any object in the bucket.

pbaACL - The canned ACL to apply to the bucket.

pbaBucket - Undocumented member.

data PutBucketACL

See: putBucketACL smart constructor.

Request Lenses

pbaGrantReadACP :: Lens' PutBucketACL (Maybe Text)

Allows grantee to read the bucket ACL.

pbaGrantWriteACP :: Lens' PutBucketACL (Maybe Text)

Allows grantee to write the ACL for the applicable bucket.

pbaGrantRead :: Lens' PutBucketACL (Maybe Text)

Allows grantee to list the objects in the bucket.

pbaGrantFullControl :: Lens' PutBucketACL (Maybe Text)

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

pbaContentMD5 :: Lens' PutBucketACL (Maybe Text)

Undocumented member.

pbaAccessControlPolicy :: Lens' PutBucketACL (Maybe AccessControlPolicy)

Undocumented member.

pbaGrantWrite :: Lens' PutBucketACL (Maybe Text)

Allows grantee to create, overwrite, and delete any object in the bucket.

pbaACL :: Lens' PutBucketACL (Maybe BucketCannedACL)

The canned ACL to apply to the bucket.

pbaBucket :: Lens' PutBucketACL BucketName

Undocumented member.

Destructuring the Response

putBucketACLResponse :: PutBucketACLResponse

Creates a value of PutBucketACLResponse with the minimum fields required to make a request.

data PutBucketACLResponse

See: putBucketACLResponse smart constructor.