Network.AWS.S3.PutBucketAcl

Description

Sets the permissions on a bucket using access control lists (ACL).

Request

data PutBucketAcl

Request constructor

putBucketAcl

Arguments

:: Text
pbarBucket

-> PutBucketAcl

PutBucketAcl constructor.

The fields accessible through corresponding lenses are:

pbarACL :: Maybe Text

pbarAccessControlPolicy :: Maybe AccessControlPolicy

pbarBucket :: Text

pbarContentMD5 :: Maybe Text

pbarGrantFullControl :: Maybe Text

pbarGrantRead :: Maybe Text

pbarGrantReadACP :: Maybe Text

pbarGrantWrite :: Maybe Text

pbarGrantWriteACP :: Maybe Text

Request lenses

pbarACL :: Lens' PutBucketAcl (Maybe Text)

The canned ACL to apply to the bucket.

pbarAccessControlPolicy :: Lens' PutBucketAcl (Maybe AccessControlPolicy)

pbarBucket :: Lens' PutBucketAcl Text

pbarContentMD5 :: Lens' PutBucketAcl (Maybe Text)

pbarGrantFullControl :: Lens' PutBucketAcl (Maybe Text)

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

pbarGrantRead :: Lens' PutBucketAcl (Maybe Text)

Allows grantee to list the objects in the bucket.

pbarGrantReadACP :: Lens' PutBucketAcl (Maybe Text)

Allows grantee to read the bucket ACL.

pbarGrantWrite :: Lens' PutBucketAcl (Maybe Text)

Allows grantee to create, overwrite, and delete any object in the bucket.

pbarGrantWriteACP :: Lens' PutBucketAcl (Maybe Text)

Allows grantee to write the ACL for the applicable bucket.

Response

data PutBucketAclResponse

Response constructor

putBucketAclResponse :: PutBucketAclResponse