Network.AWS.S3.PutBucketAnalyticsConfiguration

Description

Sets an analytics configuration for the bucket (specified by the analytics configuration ID).

Creating a Request

putBucketAnalyticsConfiguration

Arguments

:: BucketName
pBucket

-> Text
pId

-> AnalyticsConfiguration
pAnalyticsConfiguration

-> PutBucketAnalyticsConfiguration

Creates a value of PutBucketAnalyticsConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pBucket - The name of the bucket to which an analytics configuration is stored.

pId - The identifier used to represent an analytics configuration.

pAnalyticsConfiguration - The configuration and any analyses for the analytics filter.

data PutBucketAnalyticsConfiguration

See: putBucketAnalyticsConfiguration smart constructor.

Request Lenses

pBucket :: Lens' PutBucketAnalyticsConfiguration BucketName

The name of the bucket to which an analytics configuration is stored.

pId :: Lens' PutBucketAnalyticsConfiguration Text

The identifier used to represent an analytics configuration.

pAnalyticsConfiguration :: Lens' PutBucketAnalyticsConfiguration AnalyticsConfiguration

The configuration and any analyses for the analytics filter.

Destructuring the Response

putBucketAnalyticsConfigurationResponse :: PutBucketAnalyticsConfigurationResponse

Creates a value of PutBucketAnalyticsConfigurationResponse with the minimum fields required to make a request.

data PutBucketAnalyticsConfigurationResponse

See: putBucketAnalyticsConfigurationResponse smart constructor.