Network.AWS.S3.PutBucketLifecycleConfiguration

Description

Sets lifecycle configuration for your bucket. If a lifecycle configuration exists, it replaces it.

Creating a Request

putBucketLifecycleConfiguration

Arguments

:: BucketName
pblcBucket

-> PutBucketLifecycleConfiguration

Creates a value of PutBucketLifecycleConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pblcLifecycleConfiguration - Undocumented member.

pblcBucket - Undocumented member.

data PutBucketLifecycleConfiguration

See: putBucketLifecycleConfiguration smart constructor.

Request Lenses

pblcLifecycleConfiguration :: Lens' PutBucketLifecycleConfiguration (Maybe BucketLifecycleConfiguration)

Undocumented member.

pblcBucket :: Lens' PutBucketLifecycleConfiguration BucketName

Undocumented member.

Destructuring the Response

putBucketLifecycleConfigurationResponse :: PutBucketLifecycleConfigurationResponse

Creates a value of PutBucketLifecycleConfigurationResponse with the minimum fields required to make a request.

data PutBucketLifecycleConfigurationResponse

See: putBucketLifecycleConfigurationResponse smart constructor.