Network.AWS.S3.PutBucketLogging

Description

Set the logging parameters for a bucket and to specify permissions for who can view and modify the logging parameters. To set the logging status of a bucket, you must be the bucket owner.

http://docs.aws.amazon.com/AmazonS3/latest/API/PutBucketLogging.html

Request

data PutBucketLogging

Request constructor

putBucketLogging

Arguments

:: Text
pblBucket

-> BucketLoggingStatus
pblBucketLoggingStatus

-> PutBucketLogging

PutBucketLogging constructor.

The fields accessible through corresponding lenses are:

pblBucket :: Text

pblBucketLoggingStatus :: BucketLoggingStatus

pblContentMD5 :: Maybe Text

Request lenses

pblBucket :: Lens' PutBucketLogging Text

pblBucketLoggingStatus :: Lens' PutBucketLogging BucketLoggingStatus

pblContentMD5 :: Lens' PutBucketLogging (Maybe Text)

Response

data PutBucketLoggingResponse

Response constructor

putBucketLoggingResponse :: PutBucketLoggingResponse

PutBucketLoggingResponse constructor.