Network.AWS.S3.PutBucketNotificationConfiguration

Description

Enables notifications of specified events for a bucket.

Creating a Request

putBucketNotificationConfiguration

Arguments

:: BucketName
pbncBucket

-> NotificationConfiguration
pbncNotificationConfiguration

-> PutBucketNotificationConfiguration

Creates a value of PutBucketNotificationConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pbncBucket - Undocumented member.

pbncNotificationConfiguration - Undocumented member.

data PutBucketNotificationConfiguration

See: putBucketNotificationConfiguration smart constructor.

Request Lenses

pbncBucket :: Lens' PutBucketNotificationConfiguration BucketName

Undocumented member.

pbncNotificationConfiguration :: Lens' PutBucketNotificationConfiguration NotificationConfiguration

Undocumented member.

Destructuring the Response

putBucketNotificationConfigurationResponse :: PutBucketNotificationConfigurationResponse

Creates a value of PutBucketNotificationConfigurationResponse with the minimum fields required to make a request.

data PutBucketNotificationConfigurationResponse

See: putBucketNotificationConfigurationResponse smart constructor.