Network.AWS.S3.PutBucketPolicy

Description

Replaces a policy on a bucket. If the bucket already has a policy, the one in this request completely replaces it.

Creating a Request

putBucketPolicy

Arguments

:: BucketName
pbpBucket

-> HashMap Text Value
pbpPolicy

-> PutBucketPolicy

Creates a value of PutBucketPolicy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pbpContentMD5 - Undocumented member.

pbpBucket - Undocumented member.

pbpPolicy - The bucket policy as a JSON document.

data PutBucketPolicy

See: putBucketPolicy smart constructor.

Request Lenses

pbpContentMD5 :: Lens' PutBucketPolicy (Maybe Text)

Undocumented member.

pbpBucket :: Lens' PutBucketPolicy BucketName

Undocumented member.

pbpPolicy :: Lens' PutBucketPolicy (HashMap Text Value)

The bucket policy as a JSON document.

Destructuring the Response

putBucketPolicyResponse :: PutBucketPolicyResponse

Creates a value of PutBucketPolicyResponse with the minimum fields required to make a request.

data PutBucketPolicyResponse

See: putBucketPolicyResponse smart constructor.