Network.AWS.S3.PutBucketReplication

Description

Creates a new replication configuration (or replaces an existing one, if present).

Creating a Request

putBucketReplication

Arguments

:: BucketName
pbrBucket

-> ReplicationConfiguration
pbrReplicationConfiguration

-> PutBucketReplication

Creates a value of PutBucketReplication with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pbrContentMD5 - Undocumented member.

pbrBucket - Undocumented member.

pbrReplicationConfiguration - Undocumented member.

data PutBucketReplication

See: putBucketReplication smart constructor.

Request Lenses

pbrContentMD5 :: Lens' PutBucketReplication (Maybe Text)

Undocumented member.

pbrBucket :: Lens' PutBucketReplication BucketName

Undocumented member.

pbrReplicationConfiguration :: Lens' PutBucketReplication ReplicationConfiguration

Undocumented member.

Destructuring the Response

putBucketReplicationResponse :: PutBucketReplicationResponse

Creates a value of PutBucketReplicationResponse with the minimum fields required to make a request.

data PutBucketReplicationResponse

See: putBucketReplicationResponse smart constructor.