Network.AWS.S3.PutBucketRequestPayment

Description

Sets the request payment configuration for a bucket. By default, the bucket owner pays for downloads from the bucket. This configuration parameter enables the bucket owner (only) to specify that the person requesting the download will be charged for the download.

http://docs.aws.amazon.com/AmazonS3/latest/API/PutBucketRequestPayment.html

Request

data PutBucketRequestPayment

Request constructor

putBucketRequestPayment

Arguments

:: Text
pbrpBucket

-> RequestPaymentConfiguration
pbrpRequestPaymentConfiguration

-> PutBucketRequestPayment

PutBucketRequestPayment constructor.

The fields accessible through corresponding lenses are:

pbrpBucket :: Text

pbrpContentMD5 :: Maybe Text

pbrpRequestPaymentConfiguration :: RequestPaymentConfiguration

Request lenses

pbrpBucket :: Lens' PutBucketRequestPayment Text

pbrpContentMD5 :: Lens' PutBucketRequestPayment (Maybe Text)

pbrpRequestPaymentConfiguration :: Lens' PutBucketRequestPayment RequestPaymentConfiguration

Response

data PutBucketRequestPaymentResponse

Response constructor

putBucketRequestPaymentResponse :: PutBucketRequestPaymentResponse

PutBucketRequestPaymentResponse constructor.