Network.AWS.S3.PutBucketTagging

Description

Sets the tags for a bucket.

Creating a Request

putBucketTagging

Arguments

:: BucketName
pbtBucket

-> Tagging
pbtTagging

-> PutBucketTagging

Creates a value of PutBucketTagging with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pbtContentMD5 - Undocumented member.

pbtBucket - Undocumented member.

pbtTagging - Undocumented member.

data PutBucketTagging

See: putBucketTagging smart constructor.

Request Lenses

pbtContentMD5 :: Lens' PutBucketTagging (Maybe Text)

Undocumented member.

pbtBucket :: Lens' PutBucketTagging BucketName

Undocumented member.

pbtTagging :: Lens' PutBucketTagging Tagging

Undocumented member.

Destructuring the Response

putBucketTaggingResponse :: PutBucketTaggingResponse

Creates a value of PutBucketTaggingResponse with the minimum fields required to make a request.

data PutBucketTaggingResponse

See: putBucketTaggingResponse smart constructor.