Network.AWS.S3.PutObject

Description

Adds an object to a bucket.

Request

data PutObject

Request constructor

putObject

Arguments

:: RqBody
porBody

-> Text
porBucket

-> Text
porKey

-> PutObject

PutObject constructor.

The fields accessible through corresponding lenses are:

porACL :: Maybe Text

porBody :: RqBody

porBucket :: Text

porCacheControl :: Maybe Text

porContentDisposition :: Maybe Text

porContentEncoding :: Maybe Text

porContentLanguage :: Maybe Text

porContentLength :: Maybe Int

porContentMD5 :: Maybe Text

porContentType :: Maybe Text

porExpires :: Maybe UTCTime

porGrantFullControl :: Maybe Text

porGrantRead :: Maybe Text

porGrantReadACP :: Maybe Text

porGrantWriteACP :: Maybe Text

porKey :: Text

porMetadata :: HashMap Text Text

porSSECustomerAlgorithm :: Maybe Text

porSSECustomerKey :: Maybe Text

porSSECustomerKeyMD5 :: Maybe Text

porServerSideEncryption :: Maybe Text

porStorageClass :: Maybe Text

porWebsiteRedirectLocation :: Maybe Text

Request lenses

porACL :: Lens' PutObject (Maybe Text)

The canned ACL to apply to the object.

porBody :: Lens' PutObject RqBody

porBucket :: Lens' PutObject Text

porCacheControl :: Lens' PutObject (Maybe Text)

Specifies caching behavior along the request/reply chain.

porContentDisposition :: Lens' PutObject (Maybe Text)

Specifies presentational information for the object.

porContentEncoding :: Lens' PutObject (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

porContentLanguage :: Lens' PutObject (Maybe Text)

The language the content is in.

porContentLength :: Lens' PutObject (Maybe Int)

Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically.

porContentMD5 :: Lens' PutObject (Maybe Text)

porContentType :: Lens' PutObject (Maybe Text)

A standard MIME type describing the format of the object data.

porExpires :: Lens' PutObject (Maybe UTCTime)

The date and time at which the object is no longer cacheable.

porGrantFullControl :: Lens' PutObject (Maybe Text)

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

porGrantRead :: Lens' PutObject (Maybe Text)

Allows grantee to read the object data and its metadata.

porGrantReadACP :: Lens' PutObject (Maybe Text)

Allows grantee to read the object ACL.

porGrantWriteACP :: Lens' PutObject (Maybe Text)

Allows grantee to write the ACL for the applicable object.

porKey :: Lens' PutObject Text

porMetadata :: Lens' PutObject (HashMap Text Text)

A map of metadata to store with the object in S3.

porSSECustomerAlgorithm :: Lens' PutObject (Maybe Text)

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

porSSECustomerKey :: Lens' PutObject (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

porSSECustomerKeyMD5 :: Lens' PutObject (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

porServerSideEncryption :: Lens' PutObject (Maybe Text)

The Server-side encryption algorithm used when storing this object in S3.

porStorageClass :: Lens' PutObject (Maybe Text)

The type of storage to use for the object. Defaults to STANDARD.

porWebsiteRedirectLocation :: Lens' PutObject (Maybe Text)

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

Response

data PutObjectOutput

Response constructor

putObjectOutput :: PutObjectOutput

PutObjectOutput constructor.

The fields accessible through corresponding lenses are:

pooETag :: Maybe Text

pooExpiration :: Maybe UTCTime

pooSSECustomerAlgorithm :: Maybe Text

pooSSECustomerKeyMD5 :: Maybe Text

pooServerSideEncryption :: Maybe Text

pooVersionId :: Maybe Text

Response lenses

pooETag :: Lens' PutObjectOutput (Maybe Text)

Entity tag for the uploaded object.

pooExpiration :: Lens' PutObjectOutput (Maybe UTCTime)

If the object expiration is configured, this will contain the expiration date (expiry-date) and rule ID (rule-id). The value of rule-id is URL encoded.

pooSSECustomerAlgorithm :: Lens' PutObjectOutput (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

pooSSECustomerKeyMD5 :: Lens' PutObjectOutput (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

pooServerSideEncryption :: Lens' PutObjectOutput (Maybe Text)

The Server-side encryption algorithm used when storing this object in S3.

pooVersionId :: Lens' PutObjectOutput (Maybe Text)

Version of the object.