Network.AWS.S3.PutObjectACL

Description

uses the acl subresource to set the access control list (ACL) permissions for an object that already exists in a bucket

Creating a Request

putObjectACL

Arguments

:: BucketName
poaBucket

-> ObjectKey
poaKey

-> PutObjectACL

Creates a value of PutObjectACL with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

poaVersionId - VersionId used to reference a specific version of the object.

poaGrantReadACP - Allows grantee to read the bucket ACL.

poaRequestPayer - Undocumented member.

poaGrantWriteACP - Allows grantee to write the ACL for the applicable bucket.

poaGrantRead - Allows grantee to list the objects in the bucket.

poaGrantFullControl - Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

poaContentMD5 - Undocumented member.

poaAccessControlPolicy - Undocumented member.

poaGrantWrite - Allows grantee to create, overwrite, and delete any object in the bucket.

poaACL - The canned ACL to apply to the object.

poaBucket - Undocumented member.

poaKey - Undocumented member.

data PutObjectACL

See: putObjectACL smart constructor.

Request Lenses

poaVersionId :: Lens' PutObjectACL (Maybe ObjectVersionId)

VersionId used to reference a specific version of the object.

poaGrantReadACP :: Lens' PutObjectACL (Maybe Text)

Allows grantee to read the bucket ACL.

poaRequestPayer :: Lens' PutObjectACL (Maybe RequestPayer)

Undocumented member.

poaGrantWriteACP :: Lens' PutObjectACL (Maybe Text)

Allows grantee to write the ACL for the applicable bucket.

poaGrantRead :: Lens' PutObjectACL (Maybe Text)

Allows grantee to list the objects in the bucket.

poaGrantFullControl :: Lens' PutObjectACL (Maybe Text)

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

poaContentMD5 :: Lens' PutObjectACL (Maybe Text)

Undocumented member.

poaAccessControlPolicy :: Lens' PutObjectACL (Maybe AccessControlPolicy)

Undocumented member.

poaGrantWrite :: Lens' PutObjectACL (Maybe Text)

Allows grantee to create, overwrite, and delete any object in the bucket.

poaACL :: Lens' PutObjectACL (Maybe ObjectCannedACL)

The canned ACL to apply to the object.

poaBucket :: Lens' PutObjectACL BucketName

Undocumented member.

poaKey :: Lens' PutObjectACL ObjectKey

Undocumented member.

Destructuring the Response

putObjectACLResponse

Arguments

:: Int
poarsResponseStatus

-> PutObjectACLResponse

Creates a value of PutObjectACLResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

poarsRequestCharged - Undocumented member.

poarsResponseStatus - -- | The response status code.

data PutObjectACLResponse

See: putObjectACLResponse smart constructor.

Response Lenses

poarsRequestCharged :: Lens' PutObjectACLResponse (Maybe RequestCharged)

Undocumented member.

poarsResponseStatus :: Lens' PutObjectACLResponse Int

- | The response status code.