Network.AWS.S3.PutObjectTagging

Description

Sets the supplied tag-set to an object that already exists in a bucket

Creating a Request

putObjectTagging

Arguments

:: BucketName
potBucket

-> ObjectKey
potKey

-> Tagging
potTagging

-> PutObjectTagging

Creates a value of PutObjectTagging with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

potVersionId - Undocumented member.

potContentMD5 - Undocumented member.

potBucket - Undocumented member.

potKey - Undocumented member.

potTagging - Undocumented member.

data PutObjectTagging

See: putObjectTagging smart constructor.

Request Lenses

potVersionId :: Lens' PutObjectTagging (Maybe ObjectVersionId)

Undocumented member.

potContentMD5 :: Lens' PutObjectTagging (Maybe Text)

Undocumented member.

potBucket :: Lens' PutObjectTagging BucketName

Undocumented member.

potKey :: Lens' PutObjectTagging ObjectKey

Undocumented member.

potTagging :: Lens' PutObjectTagging Tagging

Undocumented member.

Destructuring the Response

putObjectTaggingResponse

Arguments

:: Int
potrsResponseStatus

-> PutObjectTaggingResponse

Creates a value of PutObjectTaggingResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

potrsVersionId - Undocumented member.

potrsResponseStatus - -- | The response status code.

data PutObjectTaggingResponse

See: putObjectTaggingResponse smart constructor.

Response Lenses

potrsVersionId :: Lens' PutObjectTaggingResponse (Maybe ObjectVersionId)

Undocumented member.

potrsResponseStatus :: Lens' PutObjectTaggingResponse Int

- | The response status code.