Network.AWS.S3.RestoreObject

Description

Restores an archived copy of an object back into Amazon S3.

Request

data RestoreObject

Request alias

type PostObjectRestore = RestoreObject

Request constructor

restoreObject

Arguments

:: Text
roBucket

-> Text
roKey

-> RestoreObject

Smart constructor for the minimum required parameters to construct a valid RestoreObject request.

The fields accessible through corresponding lenses are:

Bucket :: Text

Key :: Text

VersionId :: Maybe Text

RestoreRequest :: Maybe RestoreRequest

Request lenses

roBucket :: Lens' RestoreObject Text

roKey :: Lens' RestoreObject Text

roVersionId :: Lens' RestoreObject (Maybe Text)

roRestoreRequest :: Lens' RestoreObject (Maybe RestoreRequest)

Response

data RestoreObjectResponse

Response constructor

restoreObjectResponse :: RestoreObjectResponse

Smart constructor for the minimum required parameters to construct a valid RestoreObjectResponse response.

This constructor is provided for convenience and testing purposes.