Network.AWS.S3.Types

Description

Amazon Simple Storage Service is storage for the Internet. Amazon S3 has a simple web services interface that you can use to store and retrieve any amount of data, at any time, from anywhere on the web. It gives any developer access to the same highly scalable, reliable, fast, inexpensive data storage infrastructure that Amazon uses to run its own global network of web sites. The service aims to maximize benefits of scale and to pass those benefits on to developers.

Service

data S3

Supported version (2006-03-01) of the Amazon Simple Storage Service service.

Errors

data S3Error

A sum type representing possible errors returned by the S3 service.

These typically include HTTPExceptions thrown by the underlying HTTP mechanisms, serialisation errors, and typed errors as specified by the service description where applicable.

Constructors

BucketAlreadyExists
The requested bucket name is not available. The bucket namespace is shared by all users of the system. Please select a different name and try again.

NoSuchBucket
The specified bucket does not exist.

NoSuchKey
The specified key does not exist.

NoSuchUpload
The specified multipart upload does not exist.

ObjectAlreadyInActiveTierError
This operation is not allowed against this storage tier.

ObjectNotInActiveTierError
The source object of the COPY operation is not in the active tier and is only stored in Amazon Glacier.

S3Client HttpException

S3Serializer String

S3Service String

_BucketAlreadyExists :: Prism' S3Error ()

The requested bucket name is not available. The bucket namespace is shared by all users of the system. Please select a different name and try again.

See: BucketAlreadyExists

_NoSuchBucket :: Prism' S3Error ()

The specified bucket does not exist.

See: NoSuchBucket

_NoSuchKey :: Prism' S3Error ()

The specified key does not exist.

See: NoSuchKey

_NoSuchUpload :: Prism' S3Error ()

The specified multipart upload does not exist.

See: NoSuchUpload

_ObjectAlreadyInActiveTierError :: Prism' S3Error ()

This operation is not allowed against this storage tier.

See: ObjectAlreadyInActiveTierError

_ObjectNotInActiveTierError :: Prism' S3Error ()

The source object of the COPY operation is not in the active tier and is only stored in Amazon Glacier.

See: ObjectNotInActiveTierError

_S3Client :: Prism' S3Error HttpException

See: S3Client

_S3Serializer :: Prism' S3Error String

See: S3Serializer

_S3Service :: Prism' S3Error String

See: S3Service

XML

xmlOptions :: Tagged a XMLOptions

BucketVersioningStatus

data BucketVersioningStatus

ExpirationStatus

data ExpirationStatus

MFADelete

data MFADelete

MFADeleteStatus

data MFADeleteStatus

BucketCannedACL

data BucketCannedACL

Constructors

BucketCannedACLAuthenticatedRead
authenticated-read

BucketCannedACLPrivate
private

BucketCannedACLPublicRead
public-read

BucketCannedACLPublicReadWrite
public-read-write

BucketLocationConstraint

data BucketLocationConstraint

Constructors

BucketLocationConstraintApNortheast1
ap-northeast-1

BucketLocationConstraintApSoutheast1
ap-southeast-1

BucketLocationConstraintApSoutheast2
ap-southeast-2

BucketLocationConstraintEu
EU

BucketLocationConstraintEuWest1
eu-west-1

BucketLocationConstraintSaEast1
sa-east-1

BucketLocationConstraintUsWest1
us-west-1

BucketLocationConstraintUsWest2
us-west-2

BucketLogsPermission

data BucketLogsPermission

Constructors

BucketLogsPermissionFullControl
FULL_CONTROL

BucketLogsPermissionRead
READ

BucketLogsPermissionWrite
WRITE

EncodingType

data EncodingType

Constructors

EncodingTypeUrl
url

Event

data Event

Constructors

EventS3ReducedRedundancyLostObject
s3:ReducedRedundancyLostObject

MetadataDirective

data MetadataDirective

Constructors

MetadataDirectiveCopy
COPY

MetadataDirectiveReplace
REPLACE

ObjectCannedACL

data ObjectCannedACL

Constructors

ObjectCannedACLAuthenticatedRead
authenticated-read

ObjectCannedACLBucketOwnerFullControl
bucket-owner-full-control

ObjectCannedACLBucketOwnerRead
bucket-owner-read

ObjectCannedACLPrivate
private

ObjectCannedACLPublicRead
public-read

ObjectCannedACLPublicReadWrite
public-read-write

ObjectStorageClass

data ObjectStorageClass

Constructors

ObjectStorageClassGlacier
GLACIER

ObjectStorageClassReducedRedundancy
REDUCED_REDUNDANCY

ObjectStorageClassStandard
STANDARD

ObjectVersionStorageClass

data ObjectVersionStorageClass

Constructors

ObjectVersionStorageClassStandard
STANDARD

Payer

data Payer

Constructors

PayerBucketOwner
BucketOwner

PayerRequester
Requester

Permission

data Permission

Constructors

PermissionFullControl
FULL_CONTROL

PermissionRead
READ

PermissionReadAcp
READ_ACP

PermissionWrite
WRITE

PermissionWriteAcp
WRITE_ACP

Protocol

data Protocol

Constructors

ProtocolHttp
http

ProtocolHttps
https

ServerSideEncryption

data ServerSideEncryption

Constructors

ServerSideEncryptionAES256
AES256

StorageClass

data StorageClass

Constructors

StorageClassReducedRedundancy
REDUCED_REDUNDANCY

StorageClassStandard
STANDARD

TransitionStorageClass

data TransitionStorageClass

Constructors

TransitionStorageClassGlacier
GLACIER

Type

data Type

Constructors

TypeAmazonCustomerByEmail
AmazonCustomerByEmail

TypeCanonicalUser
CanonicalUser

TypeGroup
Group

BucketLoggingStatus

data BucketLoggingStatus

bucketLoggingStatus :: BucketLoggingStatus

Smart constructor for the minimum required fields to construct a valid BucketLoggingStatus data type to populate a request.

The fields accessible through corresponding lenses are:

LoggingEnabled :: Maybe LoggingEnabled

blsLoggingEnabled :: Lens' BucketLoggingStatus (Maybe LoggingEnabled)

CORSConfiguration

data CORSConfiguration

cORSConfiguration :: CORSConfiguration

Smart constructor for the minimum required fields to construct a valid CORSConfiguration data type to populate a request.

The fields accessible through corresponding lenses are:

CORSRules :: [CORSRule]

corscCORSRules :: Lens' CORSConfiguration [CORSRule]

CommonPrefix

data CommonPrefix

commonPrefix :: CommonPrefix

Smart constructor for the minimum required fields to construct a valid CommonPrefix data type.

CommonPrefix is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Prefix :: Maybe Text

cprPrefix :: Lens' CommonPrefix (Maybe Text)

CompletedMultipartUpload

data CompletedMultipartUpload

completedMultipartUpload :: CompletedMultipartUpload

Smart constructor for the minimum required fields to construct a valid CompletedMultipartUpload data type to populate a request.

The fields accessible through corresponding lenses are:

Parts :: [CompletedPart]

cmu1Parts :: Lens' CompletedMultipartUpload [CompletedPart]

CreateBucketConfiguration

data CreateBucketConfiguration

createBucketConfiguration :: CreateBucketConfiguration

Smart constructor for the minimum required fields to construct a valid CreateBucketConfiguration data type to populate a request.

The fields accessible through corresponding lenses are:

LocationConstraint :: Maybe BucketLocationConstraint

cbcLocationConstraint :: Lens' CreateBucketConfiguration (Maybe BucketLocationConstraint)

Specifies the region where the bucket will be created.

ErrorDocument

data ErrorDocument

errorDocument

Arguments

:: Text
edKey

-> ErrorDocument

Smart constructor for the minimum required fields to construct a valid ErrorDocument data type to populate a request.

The fields accessible through corresponding lenses are:

Key :: Text

edKey :: Lens' ErrorDocument Text

The object key name to use when a 4XX class error occurs.

IndexDocument

data IndexDocument

indexDocument

Arguments

:: Text
idSuffix

-> IndexDocument

Smart constructor for the minimum required fields to construct a valid IndexDocument data type to populate a request.

The fields accessible through corresponding lenses are:

Suffix :: Text

idSuffix :: Lens' IndexDocument Text

A suffix that is appended to a request that is for a directory on the website endpoint (e.g. if the suffix is index.html and you make a request to samplebucketimages the data that is returned will be for the object with the key name images/index.html) The suffix must not be empty and must not include a slash character.

LifecycleConfiguration

data LifecycleConfiguration

lifecycleConfiguration

Arguments

:: [Rule]
lcRules

-> LifecycleConfiguration

Smart constructor for the minimum required fields to construct a valid LifecycleConfiguration data type to populate a request.

The fields accessible through corresponding lenses are:

Rules :: [Rule]

lcRules :: Lens' LifecycleConfiguration [Rule]

NoncurrentVersionExpiration

data NoncurrentVersionExpiration

Specifies when noncurrent object versions expire. Upon expiration, Amazon S3 permanently deletes the noncurrent object versions. You set this lifecycle configuration action on a bucket that has versioning enabled (or suspended) to request that Amazon S3 delete noncurrent object versions at a specific period in the object's lifetime.

noncurrentVersionExpiration :: NoncurrentVersionExpiration

Smart constructor for the minimum required fields to construct a valid NoncurrentVersionExpiration data type to populate a request.

The fields accessible through corresponding lenses are:

NoncurrentDays :: Maybe Integer

nveNoncurrentDays :: Lens' NoncurrentVersionExpiration (Maybe Integer)

Specifies the number of days an object is noncurrent before Amazon S3 can perform the associated action. For information about the noncurrent days calculations, see How Amazon S3 Calculates When an Object Became Noncurrent in the Amazon Simple Storage Service Developer Guide.

NotificationConfiguration

data NotificationConfiguration

notificationConfiguration

Arguments

:: TopicConfiguration
ncTopicConfiguration

-> NotificationConfiguration

Smart constructor for the minimum required fields to construct a valid NotificationConfiguration data type to populate a request.

The fields accessible through corresponding lenses are:

TopicConfiguration :: TopicConfiguration

ncTopicConfiguration :: Lens' NotificationConfiguration TopicConfiguration

RequestPaymentConfiguration

data RequestPaymentConfiguration

requestPaymentConfiguration

Arguments

:: Payer
rpcPayer

-> RequestPaymentConfiguration

Smart constructor for the minimum required fields to construct a valid RequestPaymentConfiguration data type to populate a request.

The fields accessible through corresponding lenses are:

Payer :: Payer

rpcPayer :: Lens' RequestPaymentConfiguration Payer

Specifies who pays for the download and request fees.

RestoreRequest

data RestoreRequest

restoreRequest

Arguments

:: Integer
rr1Days

-> RestoreRequest

Smart constructor for the minimum required fields to construct a valid RestoreRequest data type to populate a request.

The fields accessible through corresponding lenses are:

Days :: Integer

rr1Days :: Lens' RestoreRequest Integer

Lifetime of the active copy in days.

Tagging

data Tagging

tagging

Arguments

:: [Tag]
t1TagSet

-> Tagging

Smart constructor for the minimum required fields to construct a valid Tagging data type to populate a request.

The fields accessible through corresponding lenses are:

TagSet :: [Tag]

t1TagSet :: Lens' Tagging [Tag]

AccessControlPolicy

data AccessControlPolicy

accessControlPolicy :: AccessControlPolicy

Smart constructor for the minimum required fields to construct a valid AccessControlPolicy data type to populate a request.

The fields accessible through corresponding lenses are:

Grants :: [Grant]

Owner :: Maybe Owner

acpGrants :: Lens' AccessControlPolicy [Grant]

A list of grants.

acpOwner :: Lens' AccessControlPolicy (Maybe Owner)

Bucket

data Bucket

bucket :: Bucket

Smart constructor for the minimum required fields to construct a valid Bucket data type.

Bucket is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Name :: Maybe Text

CreationDate :: Maybe RFC822

bName :: Lens' Bucket (Maybe Text)

The name of the bucket.

bCreationDate :: Lens' Bucket (Maybe RFC822)

Date the bucket was created.

CORSRule

data CORSRule

cORSRule :: CORSRule

Smart constructor for the minimum required fields to construct a valid CORSRule data type to populate a request.

The fields accessible through corresponding lenses are:

AllowedHeaders :: [Text]

AllowedMethods :: [Text]

AllowedOrigins :: [Text]

ExposeHeaders :: [Text]

MaxAgeSeconds :: Maybe Integer

corsrAllowedHeaders :: Lens' CORSRule [Text]

Specifies which headers are allowed in a pre-flight OPTIONS request.

corsrAllowedMethods :: Lens' CORSRule [Text]

Identifies HTTP methods that the domain/origin specified in the rule is allowed to execute.

corsrAllowedOrigins :: Lens' CORSRule [Text]

One or more origins you want customers to be able to access the bucket from.

corsrExposeHeaders :: Lens' CORSRule [Text]

One or more headers in the response that you want customers to be able to access from their applications (for example, from a JavaScript XMLHttpRequest object).

corsrMaxAgeSeconds :: Lens' CORSRule (Maybe Integer)

The time in seconds that your browser is to cache the preflight response for the specified resource.

CompletedPart

data CompletedPart

completedPart :: CompletedPart

Smart constructor for the minimum required fields to construct a valid CompletedPart data type to populate a request.

The fields accessible through corresponding lenses are:

ETag :: Maybe Text

PartNumber :: Maybe Integer

cpETag :: Lens' CompletedPart (Maybe Text)

Entity tag returned when the part was uploaded.

cpPartNumber :: Lens' CompletedPart (Maybe Integer)

Part number that identifies the part.

Condition

data Condition

A container for describing a condition that must be met for the specified redirect to apply. For example, 1. If request is for pages in the /docs folder, redirect to the /documents folder. 2. If request results in HTTP error 4xx, redirect request to another host where you might process the error.

condition :: Condition

Smart constructor for the minimum required fields to construct a valid Condition data type to populate a request.

The fields accessible through corresponding lenses are:

HttpErrorCodeReturnedEquals :: Maybe Text

KeyPrefixEquals :: Maybe Text

cHttpErrorCodeReturnedEquals :: Lens' Condition (Maybe Text)

The HTTP error code when the redirect is applied. In the event of an error, if the error code equals this value, then the specified redirect is applied. Required when parent element Condition is specified and sibling KeyPrefixEquals is not specified. If both are specified, then both must be true for the redirect to be applied.

cKeyPrefixEquals :: Lens' Condition (Maybe Text)

The object key name prefix when the redirect is applied. For example, to redirect requests for ExamplePage.html, the key prefix will be ExamplePage.html. To redirect request for all pages with the prefix docs/, the key prefix will be docs, which identifies all objects in the docs folder. Required when the parent element Condition is specified and sibling HttpErrorCodeReturnedEquals is not specified. If both conditions are specified, both must be true for the redirect to be applied.

CopyObjectResult

data CopyObjectResult

copyObjectResult :: CopyObjectResult

Smart constructor for the minimum required fields to construct a valid CopyObjectResult data type.

CopyObjectResult is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ETag :: Maybe Text

LastModified :: Maybe RFC822

corrETag :: Lens' CopyObjectResult (Maybe Text)

corrLastModified :: Lens' CopyObjectResult (Maybe RFC822)

CopyPartResult

data CopyPartResult

copyPartResult :: CopyPartResult

Smart constructor for the minimum required fields to construct a valid CopyPartResult data type.

CopyPartResult is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ETag :: Maybe Text

LastModified :: Maybe RFC822

cprrETag :: Lens' CopyPartResult (Maybe Text)

Entity tag of the object.

cprrLastModified :: Lens' CopyPartResult (Maybe RFC822)

Date and time at which the object was uploaded.

Delete

data Delete

delete

Arguments

:: [ObjectIdentifier]
dObjects

-> Delete

Smart constructor for the minimum required fields to construct a valid Delete data type to populate a request.

The fields accessible through corresponding lenses are:

Objects :: [ObjectIdentifier]

Quiet :: Maybe Bool

dObjects :: Lens' Delete [ObjectIdentifier]

dQuiet :: Lens' Delete (Maybe Bool)

Element to enable quiet mode for the request. When you add this element, you must set its value to true.

DeleteMarkerEntry

data DeleteMarkerEntry

deleteMarkerEntry :: DeleteMarkerEntry

Smart constructor for the minimum required fields to construct a valid DeleteMarkerEntry data type.

DeleteMarkerEntry is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Owner :: Maybe Owner

Key :: Maybe Text

VersionId :: Maybe Text

IsLatest :: Maybe Bool

LastModified :: Maybe RFC822

dmeOwner :: Lens' DeleteMarkerEntry (Maybe Owner)

dmeKey :: Lens' DeleteMarkerEntry (Maybe Text)

The object key.

dmeVersionId :: Lens' DeleteMarkerEntry (Maybe Text)

Version ID of an object.

dmeIsLatest :: Lens' DeleteMarkerEntry (Maybe Bool)

Specifies whether the object is (true) or is not (false) the latest version of an object.

dmeLastModified :: Lens' DeleteMarkerEntry (Maybe RFC822)

Date and time the object was last modified.

DeletedObject

data DeletedObject

deletedObject :: DeletedObject

Smart constructor for the minimum required fields to construct a valid DeletedObject data type.

DeletedObject is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Key :: Maybe Text

VersionId :: Maybe Text

DeleteMarker :: Maybe Bool

DeleteMarkerVersionId :: Maybe Text

do1rKey :: Lens' DeletedObject (Maybe Text)

do1rVersionId :: Lens' DeletedObject (Maybe Text)

do1rDeleteMarker :: Lens' DeletedObject (Maybe Bool)

do1rDeleteMarkerVersionId :: Lens' DeletedObject (Maybe Text)

Error

data Error

error' :: Error

Smart constructor for the minimum required fields to construct a valid Error data type.

Error is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Key :: Maybe Text

VersionId :: Maybe Text

Code :: Maybe Text

Message :: Maybe Text

eKey :: Lens' Error (Maybe Text)

eVersionId :: Lens' Error (Maybe Text)

eCode :: Lens' Error (Maybe Text)

eMessage :: Lens' Error (Maybe Text)

Grant

data Grant

grant :: Grant

Smart constructor for the minimum required fields to construct a valid Grant data type to populate a request.

The fields accessible through corresponding lenses are:

Grantee :: Maybe Grantee

Permission :: Maybe Permission

gGrantee :: Lens' Grant (Maybe Grantee)

gPermission :: Lens' Grant (Maybe Permission)

Specifies the permission given to the grantee.

Grantee

data Grantee

grantee

Arguments

:: Type
g1Type

-> Grantee

Smart constructor for the minimum required fields to construct a valid Grantee data type to populate a request.

The fields accessible through corresponding lenses are:

DisplayName :: Maybe Text

EmailAddress :: Maybe Text

ID :: Maybe Text

Type :: Type

URI :: Maybe Text

g1DisplayName :: Lens' Grantee (Maybe Text)

Screen name of the grantee.

g1EmailAddress :: Lens' Grantee (Maybe Text)

Email address of the grantee.

g1ID :: Lens' Grantee (Maybe Text)

The canonical user ID of the grantee.

g1Type :: Lens' Grantee Type

Type of grantee.

g1URI :: Lens' Grantee (Maybe Text)

URI of the grantee group.

Initiator

data Initiator

Identifies who initiated the multipart upload.

initiator :: Initiator

Smart constructor for the minimum required fields to construct a valid Initiator data type to populate a request.

The fields accessible through corresponding lenses are:

ID :: Maybe Text

DisplayName :: Maybe Text

iID :: Lens' Initiator (Maybe Text)

If the principal is an AWS account, it provides the Canonical User ID. If the principal is an IAM User, it provides a user ARN value.

iDisplayName :: Lens' Initiator (Maybe Text)

Name of the Principal.

LifecycleExpiration

data LifecycleExpiration

lifecycleExpiration :: LifecycleExpiration

Smart constructor for the minimum required fields to construct a valid LifecycleExpiration data type to populate a request.

The fields accessible through corresponding lenses are:

Date :: Maybe RFC822

Days :: Maybe Integer

leDate :: Lens' LifecycleExpiration (Maybe RFC822)

Indicates at what date the object is to be moved or deleted. Should be in GMT ISO 8601 Format.

leDays :: Lens' LifecycleExpiration (Maybe Integer)

Indicates the lifetime, in days, of the objects that are subject to the rule. The value must be a non-zero positive integer.

LoggingEnabled

data LoggingEnabled

loggingEnabled :: LoggingEnabled

Smart constructor for the minimum required fields to construct a valid LoggingEnabled data type to populate a request.

The fields accessible through corresponding lenses are:

TargetBucket :: Maybe Text

TargetGrants :: [TargetGrant]

TargetPrefix :: Maybe Text

le1TargetBucket :: Lens' LoggingEnabled (Maybe Text)

Specifies the bucket where you want Amazon S3 to store server access logs. You can have your logs delivered to any bucket that you own, including the same bucket that is being logged. You can also configure multiple buckets to deliver their logs to the same target bucket. In this case you should choose a different TargetPrefix for each source bucket so that the delivered log files can be distinguished by key.

le1TargetGrants :: Lens' LoggingEnabled [TargetGrant]

le1TargetPrefix :: Lens' LoggingEnabled (Maybe Text)

This element lets you specify a prefix for the keys that the log files will be stored under.

MultipartUpload

data MultipartUpload

multipartUpload :: MultipartUpload

Smart constructor for the minimum required fields to construct a valid MultipartUpload data type.

MultipartUpload is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

UploadId :: Maybe Text

Key :: Maybe Text

Initiated :: Maybe RFC822

StorageClass :: Maybe StorageClass

Owner :: Maybe Owner

Initiator :: Maybe Initiator

muUploadId :: Lens' MultipartUpload (Maybe Text)

Upload ID that identifies the multipart upload.

muKey :: Lens' MultipartUpload (Maybe Text)

Key of the object for which the multipart upload was initiated.

muInitiated :: Lens' MultipartUpload (Maybe RFC822)

Date and time at which the multipart upload was initiated.

muStorageClass :: Lens' MultipartUpload (Maybe StorageClass)

The class of storage used to store the object.

muOwner :: Lens' MultipartUpload (Maybe Owner)

muInitiator :: Lens' MultipartUpload (Maybe Initiator)

Identifies who initiated the multipart upload.

NoncurrentVersionTransition

data NoncurrentVersionTransition

Container for the transition rule that describes when noncurrent objects transition to the GLACIER storage class. If your bucket is versioning-enabled (or versioning is suspended), you can set this action to request that Amazon S3 transition noncurrent object versions to the GLACIER storage class at a specific period in the object's lifetime.

noncurrentVersionTransition :: NoncurrentVersionTransition

Smart constructor for the minimum required fields to construct a valid NoncurrentVersionTransition data type to populate a request.

The fields accessible through corresponding lenses are:

NoncurrentDays :: Maybe Integer

StorageClass :: Maybe TransitionStorageClass

nvtNoncurrentDays :: Lens' NoncurrentVersionTransition (Maybe Integer)

Specifies the number of days an object is noncurrent before Amazon S3 can perform the associated action. For information about the noncurrent days calculations, see How Amazon S3 Calculates When an Object Became Noncurrent in the Amazon Simple Storage Service Developer Guide.

nvtStorageClass :: Lens' NoncurrentVersionTransition (Maybe TransitionStorageClass)

The class of storage used to store the object.

Object

data Object

object

Arguments

:: Text
orKey

-> RFC822
orLastModified

-> Text
orETag

-> Integer
orSize

-> ObjectStorageClass
orStorageClass

-> Owner
orOwner

-> Object

Smart constructor for the minimum required fields to construct a valid Object data type.

Object is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Key :: Text

LastModified :: RFC822

ETag :: Text

Size :: Integer

StorageClass :: ObjectStorageClass

Owner :: Owner

orKey :: Lens' Object Text

orLastModified :: Lens' Object RFC822

orETag :: Lens' Object Text

orSize :: Lens' Object Integer

orStorageClass :: Lens' Object ObjectStorageClass

The class of storage used to store the object.

orOwner :: Lens' Object Owner

ObjectIdentifier

data ObjectIdentifier

objectIdentifier

Arguments

:: Text
oiKey

-> ObjectIdentifier

Smart constructor for the minimum required fields to construct a valid ObjectIdentifier data type to populate a request.

The fields accessible through corresponding lenses are:

Key :: Text

VersionId :: Maybe Text

oiKey :: Lens' ObjectIdentifier Text

Key name of the object to delete.

oiVersionId :: Lens' ObjectIdentifier (Maybe Text)

VersionId for the specific version of the object to delete.

ObjectVersion

data ObjectVersion

objectVersion :: ObjectVersion

Smart constructor for the minimum required fields to construct a valid ObjectVersion data type.

ObjectVersion is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

ETag :: Maybe Text

Size :: Maybe Integer

StorageClass :: Maybe ObjectVersionStorageClass

Key :: Maybe Text

VersionId :: Maybe Text

IsLatest :: Maybe Bool

LastModified :: Maybe RFC822

Owner :: Maybe Owner

ovETag :: Lens' ObjectVersion (Maybe Text)

ovSize :: Lens' ObjectVersion (Maybe Integer)

Size in bytes of the object.

ovStorageClass :: Lens' ObjectVersion (Maybe ObjectVersionStorageClass)

The class of storage used to store the object.

ovKey :: Lens' ObjectVersion (Maybe Text)

The object key.

ovVersionId :: Lens' ObjectVersion (Maybe Text)

Version ID of an object.

ovIsLatest :: Lens' ObjectVersion (Maybe Bool)

Specifies whether the object is (true) or is not (false) the latest version of an object.

ovLastModified :: Lens' ObjectVersion (Maybe RFC822)

Date and time the object was last modified.

ovOwner :: Lens' ObjectVersion (Maybe Owner)

Owner

data Owner

owner :: Owner

Smart constructor for the minimum required fields to construct a valid Owner data type to populate a request.

The fields accessible through corresponding lenses are:

DisplayName :: Maybe Text

ID :: Maybe Text

oDisplayName :: Lens' Owner (Maybe Text)

oID :: Lens' Owner (Maybe Text)

Part

data Part

part :: Part

Smart constructor for the minimum required fields to construct a valid Part data type.

Part is exclusively used in responses and this constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

PartNumber :: Maybe Integer

LastModified :: Maybe RFC822

ETag :: Maybe Text

Size :: Maybe Integer

pPartNumber :: Lens' Part (Maybe Integer)

Part number identifying the part.

pLastModified :: Lens' Part (Maybe RFC822)

Date and time at which the part was uploaded.

pETag :: Lens' Part (Maybe Text)

Entity tag returned when the part was uploaded.

pSize :: Lens' Part (Maybe Integer)

Size of the uploaded part data.

Redirect

data Redirect

Container for redirect information. You can redirect requests to another host, to another page, or with another protocol. In the event of an error, you can can specify a different error code to return.

redirect :: Redirect

Smart constructor for the minimum required fields to construct a valid Redirect data type to populate a request.

The fields accessible through corresponding lenses are:

HostName :: Maybe Text

HttpRedirectCode :: Maybe Text

Protocol :: Maybe Protocol

ReplaceKeyPrefixWith :: Maybe Text

ReplaceKeyWith :: Maybe Text

r1HostName :: Lens' Redirect (Maybe Text)

The host name to use in the redirect request.

r1HttpRedirectCode :: Lens' Redirect (Maybe Text)

The HTTP redirect code to use on the response. Not required if one of the siblings is present.

r1Protocol :: Lens' Redirect (Maybe Protocol)

Protocol to use (http, https) when redirecting requests. The default is the protocol that is used in the original request.

r1ReplaceKeyPrefixWith :: Lens' Redirect (Maybe Text)

The object key prefix to use in the redirect request. For example, to redirect requests for all pages with prefix docs (objects in the docs folder) to documents/, you can set a condition block with KeyPrefixEquals set to docs and in the Redirect set ReplaceKeyPrefixWith to documents. Not required if one of the siblings is present. Can be present only if ReplaceKeyWith is not provided.

r1ReplaceKeyWith :: Lens' Redirect (Maybe Text)

The specific object key to use in the redirect request. For example, redirect request to error.html. Not required if one of the sibling is present. Can be present only if ReplaceKeyPrefixWith is not provided.

RedirectAllRequestsTo

data RedirectAllRequestsTo

redirectAllRequestsTo

Arguments

:: Text
rartHostName

-> RedirectAllRequestsTo

Smart constructor for the minimum required fields to construct a valid RedirectAllRequestsTo data type to populate a request.

The fields accessible through corresponding lenses are:

HostName :: Text

Protocol :: Maybe Protocol

rartHostName :: Lens' RedirectAllRequestsTo Text

Name of the host where requests will be redirected.

rartProtocol :: Lens' RedirectAllRequestsTo (Maybe Protocol)

Protocol to use (http, https) when redirecting requests. The default is the protocol that is used in the original request.

RoutingRule

data RoutingRule

routingRule

Arguments

:: Redirect
rrRedirect

-> RoutingRule

Smart constructor for the minimum required fields to construct a valid RoutingRule data type to populate a request.

The fields accessible through corresponding lenses are:

Condition :: Maybe Condition

Redirect :: Redirect

rrCondition :: Lens' RoutingRule (Maybe Condition)

A container for describing a condition that must be met for the specified redirect to apply. For example, 1. If request is for pages in the /docs folder, redirect to the /documents folder. 2. If request results in HTTP error 4xx, redirect request to another host where you might process the error.

rrRedirect :: Lens' RoutingRule Redirect

Container for redirect information. You can redirect requests to another host, to another page, or with another protocol. In the event of an error, you can can specify a different error code to return.

Rule

data Rule

rule

Arguments

:: Text
rPrefix

-> Switch ExpirationStatus
rStatus

-> Rule

Smart constructor for the minimum required fields to construct a valid Rule data type to populate a request.

The fields accessible through corresponding lenses are:

Expiration :: Maybe LifecycleExpiration

ID :: Maybe Text

Prefix :: Text

Status :: Switch ExpirationStatus

Transition :: Maybe Transition

NoncurrentVersionTransition :: Maybe NoncurrentVersionTransition

NoncurrentVersionExpiration :: Maybe NoncurrentVersionExpiration

rExpiration :: Lens' Rule (Maybe LifecycleExpiration)

rID :: Lens' Rule (Maybe Text)

Unique identifier for the rule. The value cannot be longer than 255 characters.

rPrefix :: Lens' Rule Text

Prefix identifying one or more objects to which the rule applies.

rStatus :: Lens' Rule (Switch ExpirationStatus)

If Enabled, the rule is currently being applied. If Disabled, the rule is not currently being applied.

rTransition :: Lens' Rule (Maybe Transition)

rNoncurrentVersionTransition :: Lens' Rule (Maybe NoncurrentVersionTransition)

Container for the transition rule that describes when noncurrent objects transition to the GLACIER storage class. If your bucket is versioning-enabled (or versioning is suspended), you can set this action to request that Amazon S3 transition noncurrent object versions to the GLACIER storage class at a specific period in the object's lifetime.

rNoncurrentVersionExpiration :: Lens' Rule (Maybe NoncurrentVersionExpiration)

Specifies when noncurrent object versions expire. Upon expiration, Amazon S3 permanently deletes the noncurrent object versions. You set this lifecycle configuration action on a bucket that has versioning enabled (or suspended) to request that Amazon S3 delete noncurrent object versions at a specific period in the object's lifetime.

Tag

data Tag

tag

Arguments

:: Text
t2Key

-> Text
t2Value

-> Tag

Smart constructor for the minimum required fields to construct a valid Tag data type to populate a request.

The fields accessible through corresponding lenses are:

Key :: Text

Value :: Text

t2Key :: Lens' Tag Text

Name of the tag.

t2Value :: Lens' Tag Text

Value of the tag.

TargetGrant

data TargetGrant

targetGrant :: TargetGrant

Smart constructor for the minimum required fields to construct a valid TargetGrant data type to populate a request.

The fields accessible through corresponding lenses are:

Grantee :: Maybe Grantee

Permission :: Maybe BucketLogsPermission

tgGrantee :: Lens' TargetGrant (Maybe Grantee)

tgPermission :: Lens' TargetGrant (Maybe BucketLogsPermission)

Logging permissions assigned to the Grantee for the bucket.

TopicConfiguration

data TopicConfiguration

topicConfiguration :: TopicConfiguration

Smart constructor for the minimum required fields to construct a valid TopicConfiguration data type to populate a request.

The fields accessible through corresponding lenses are:

Event :: Maybe Event

Topic :: Maybe Text

tcEvent :: Lens' TopicConfiguration (Maybe Event)

Bucket event for which to send notifications.

tcTopic :: Lens' TopicConfiguration (Maybe Text)

Amazon SNS topic to which Amazon S3 will publish a message to report the specified events for the bucket.

Transition

data Transition

transition :: Transition

Smart constructor for the minimum required fields to construct a valid Transition data type to populate a request.

The fields accessible through corresponding lenses are:

Date :: Maybe RFC822

Days :: Maybe Integer

StorageClass :: Maybe TransitionStorageClass

tDate :: Lens' Transition (Maybe RFC822)

Indicates at what date the object is to be moved or deleted. Should be in GMT ISO 8601 Format.

tDays :: Lens' Transition (Maybe Integer)

Indicates the lifetime, in days, of the objects that are subject to the rule. The value must be a non-zero positive integer.

tStorageClass :: Lens' Transition (Maybe TransitionStorageClass)

The class of storage used to store the object.

VersioningConfiguration

data VersioningConfiguration

versioningConfiguration :: VersioningConfiguration

Smart constructor for the minimum required fields to construct a valid VersioningConfiguration data type to populate a request.

The fields accessible through corresponding lenses are:

Status :: Maybe (Switch BucketVersioningStatus)

MfaDelete :: Maybe (Switch MFADelete)

vcStatus :: Lens' VersioningConfiguration (Maybe (Switch BucketVersioningStatus))

The versioning state of the bucket.

vcMfaDelete :: Lens' VersioningConfiguration (Maybe (Switch MFADelete))

Specifies whether MFA delete is enabled in the bucket versioning configuration. This element is only returned if the bucket has been configured with MFA delete. If the bucket has never been so configured, this element is not returned.

WebsiteConfiguration

data WebsiteConfiguration

websiteConfiguration :: WebsiteConfiguration

Smart constructor for the minimum required fields to construct a valid WebsiteConfiguration data type to populate a request.

The fields accessible through corresponding lenses are:

ErrorDocument :: Maybe ErrorDocument

IndexDocument :: Maybe IndexDocument

RedirectAllRequestsTo :: Maybe RedirectAllRequestsTo

RoutingRules :: [RoutingRule]

wcErrorDocument :: Lens' WebsiteConfiguration (Maybe ErrorDocument)

wcIndexDocument :: Lens' WebsiteConfiguration (Maybe IndexDocument)

wcRedirectAllRequestsTo :: Lens' WebsiteConfiguration (Maybe RedirectAllRequestsTo)

wcRoutingRules :: Lens' WebsiteConfiguration [RoutingRule]

Common

newtype BucketName

Constructors

BucketName Text

newtype ObjectKey

Constructors

ObjectKey Text

newtype ObjectVersionId

Constructors

ObjectVersionId Text

newtype ETag

Constructors

ETag Text