Network.AWS.S3.UploadPart

Description

Uploads a part in a multipart upload.

Note: After you initiate multipart upload and upload one or more parts, you must either complete or abort multipart upload in order to stop getting charged for storage of the uploaded parts. Only after you either complete or abort multipart upload, Amazon S3 frees up the parts storage and stops charging you for the parts storage.

http://docs.aws.amazon.com/AmazonS3/latest/API/UploadPart.html

Request

data UploadPart

Request constructor

uploadPart

Arguments

:: RqBody
upBody

-> Text
upBucket

-> Text
upKey

-> Int
upPartNumber

-> Text
upUploadId

-> UploadPart

UploadPart constructor.

The fields accessible through corresponding lenses are:

upBody :: RqBody

upBucket :: Text

upContentLength :: Maybe Int

upContentMD5 :: Maybe Text

upKey :: Text

upPartNumber :: Int

upRequestPayer :: Maybe RequestPayer

upSSECustomerAlgorithm :: Maybe Text

upSSECustomerKey :: Maybe Text

upSSECustomerKeyMD5 :: Maybe Text

upUploadId :: Text

Request lenses

upBody :: Lens' UploadPart RqBody

upBucket :: Lens' UploadPart Text

upContentLength :: Lens' UploadPart (Maybe Int)

Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically.

upContentMD5 :: Lens' UploadPart (Maybe Text)

upKey :: Lens' UploadPart Text

upPartNumber :: Lens' UploadPart Int

Part number of part being uploaded.

upRequestPayer :: Lens' UploadPart (Maybe RequestPayer)

upSSECustomerAlgorithm :: Lens' UploadPart (Maybe Text)

Specifies the algorithm to use to when encrypting the object (e.g., AES256, aws:kms).

upSSECustomerKey :: Lens' UploadPart (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

upSSECustomerKeyMD5 :: Lens' UploadPart (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

upUploadId :: Lens' UploadPart Text

Upload ID identifying the multipart upload whose part is being uploaded.

Response

data UploadPartResponse

Response constructor

uploadPartResponse :: UploadPartResponse

UploadPartResponse constructor.

The fields accessible through corresponding lenses are:

uprETag :: Maybe Text

uprRequestCharged :: Maybe RequestCharged

uprSSECustomerAlgorithm :: Maybe Text

uprSSECustomerKeyMD5 :: Maybe Text

uprSSEKMSKeyId :: Maybe Text

uprServerSideEncryption :: Maybe ServerSideEncryption

Response lenses

uprETag :: Lens' UploadPartResponse (Maybe Text)

Entity tag for the uploaded object.

uprRequestCharged :: Lens' UploadPartResponse (Maybe RequestCharged)

uprSSECustomerAlgorithm :: Lens' UploadPartResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

uprSSECustomerKeyMD5 :: Lens' UploadPartResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

uprSSEKMSKeyId :: Lens' UploadPartResponse (Maybe Text)

If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

uprServerSideEncryption :: Lens' UploadPartResponse (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).