Network.AWS.S3.UploadPartCopy

Description

Uploads a part by copying data from an existing object as data source.

Request

data UploadPartCopy

Request constructor

uploadPartCopy

Arguments

:: Text
upcrBucket

-> Text
upcrCopySource

-> Text
upcrKey

-> Int
upcrPartNumber

-> Text
upcrUploadId

-> UploadPartCopy

UploadPartCopy constructor.

The fields accessible through corresponding lenses are:

upcrBucket :: Text

upcrCopySource :: Text

upcrCopySourceIfMatch :: Maybe Text

upcrCopySourceIfModifiedSince :: Maybe UTCTime

upcrCopySourceIfNoneMatch :: Maybe Text

upcrCopySourceIfUnmodifiedSince :: Maybe UTCTime

upcrCopySourceRange :: Maybe Text

upcrCopySourceSSECustomerAlgorithm :: Maybe Text

upcrCopySourceSSECustomerKey :: Maybe Text

upcrCopySourceSSECustomerKeyMD5 :: Maybe Text

upcrKey :: Text

upcrPartNumber :: Int

upcrSSECustomerAlgorithm :: Maybe Text

upcrSSECustomerKey :: Maybe Text

upcrSSECustomerKeyMD5 :: Maybe Text

upcrUploadId :: Text

Request lenses

upcrBucket :: Lens' UploadPartCopy Text

upcrCopySource :: Lens' UploadPartCopy Text

The name of the source bucket and key name of the source object, separated by a slash (/). Must be URL-encoded.

upcrCopySourceIfMatch :: Lens' UploadPartCopy (Maybe Text)

Copies the object if its entity tag (ETag) matches the specified tag.

upcrCopySourceIfModifiedSince :: Lens' UploadPartCopy (Maybe UTCTime)

Copies the object if it has been modified since the specified time.

upcrCopySourceIfNoneMatch :: Lens' UploadPartCopy (Maybe Text)

Copies the object if its entity tag (ETag) is different than the specified ETag.

upcrCopySourceIfUnmodifiedSince :: Lens' UploadPartCopy (Maybe UTCTime)

Copies the object if it hasn't been modified since the specified time.

upcrCopySourceRange :: Lens' UploadPartCopy (Maybe Text)

The range of bytes to copy from the source object. The range value must use the form bytes=first-last, where the first and last are the zero-based byte offsets to copy. For example, bytes=0-9 indicates that you want to copy the first ten bytes of the source. You can copy a range only if the source object is greater than 5 GB.

upcrCopySourceSSECustomerAlgorithm :: Lens' UploadPartCopy (Maybe Text)

Specifies the algorithm to use when decrypting the source object (e.g., AES256).

upcrCopySourceSSECustomerKey :: Lens' UploadPartCopy (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

upcrCopySourceSSECustomerKeyMD5 :: Lens' UploadPartCopy (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

upcrKey :: Lens' UploadPartCopy Text

upcrPartNumber :: Lens' UploadPartCopy Int

Part number of part being copied.

upcrSSECustomerAlgorithm :: Lens' UploadPartCopy (Maybe Text)

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

upcrSSECustomerKey :: Lens' UploadPartCopy (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

upcrSSECustomerKeyMD5 :: Lens' UploadPartCopy (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

upcrUploadId :: Lens' UploadPartCopy Text

Upload ID identifying the multipart upload whose part is being copied.

Response

data UploadPartCopyOutput

Response constructor

uploadPartCopyOutput :: UploadPartCopyOutput

UploadPartCopyOutput constructor.

The fields accessible through corresponding lenses are:

upcoCopyPartResult :: Maybe CopyPartResult

upcoCopySourceVersionId :: Maybe Text

upcoSSECustomerAlgorithm :: Maybe Text

upcoSSECustomerKeyMD5 :: Maybe Text

upcoServerSideEncryption :: Maybe Text

Response lenses

upcoCopyPartResult :: Lens' UploadPartCopyOutput (Maybe CopyPartResult)

upcoCopySourceVersionId :: Lens' UploadPartCopyOutput (Maybe Text)

The version of the source object that was copied, if you have enabled versioning on the source bucket.

upcoSSECustomerAlgorithm :: Lens' UploadPartCopyOutput (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

upcoSSECustomerKeyMD5 :: Lens' UploadPartCopyOutput (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

upcoServerSideEncryption :: Lens' UploadPartCopyOutput (Maybe Text)

The Server-side encryption algorithm used when storing this object in S3.