Network.AWS.S3.V2006_03_01.CompleteMultipartUpload

Description

Completes a multipart upload by assembling previously uploaded parts.

Request

data CompleteMultipartUpload

Request constructor

mkCompleteMultipartUpload

Arguments

:: BucketName
cmuBucket

-> ObjectKey
cmuKey

-> Text
cmuUploadId

-> CompleteMultipartUpload

Smart constructor for the minimum required parameters to construct a valid CompleteMultipartUpload request.

Request lenses

cmuBucket :: Lens' CompleteMultipartUpload BucketName

cmuKey :: Lens' CompleteMultipartUpload ObjectKey

cmuMultipartUpload :: Lens' CompleteMultipartUpload (Maybe CompletedMultipartUpload)

cmuUploadId :: Lens' CompleteMultipartUpload Text

Response

data CompleteMultipartUploadResponse

Response lenses

cmurLocation :: Lens' CompleteMultipartUploadResponse (Maybe Text)

cmurBucket :: Lens' CompleteMultipartUploadResponse (Maybe BucketName)

cmurKey :: Lens' CompleteMultipartUploadResponse (Maybe ObjectKey)

cmurExpiration :: Lens' CompleteMultipartUploadResponse (Maybe RFC822)

If the object expiration is configured, this will contain the expiration date (expiry-date) and rule ID (rule-id). The value of rule-id is URL encoded.

cmurETag :: Lens' CompleteMultipartUploadResponse (Maybe ETag)

Entity tag of the object.

cmurServerSideEncryption :: Lens' CompleteMultipartUploadResponse (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3.

cmurVersionId :: Lens' CompleteMultipartUploadResponse (Maybe ObjectVersionId)

Version of the object.