Network.AWS.S3.V2006_03_01.CreateMultipartUpload

Description

Initiates a multipart upload and returns an upload ID. Note: After you initiate multipart upload and upload one or more parts, you must either complete or abort multipart upload in order to stop getting charged for storage of the uploaded parts. Only after you either complete or abort multipart upload, Amazon S3 frees up the parts storage and stops charging you for the parts storage.

Request

data CreateMultipartUpload

Request alias

type InitiateMultipartUpload = CreateMultipartUpload

Request constructor

mkCreateMultipartUpload

Arguments

:: ObjectKey
cmu2Key

-> BucketName
cmu2Bucket

-> CreateMultipartUpload

Smart constructor for the minimum required parameters to construct a valid CreateMultipartUpload request.

Request lenses

cmu2ACL :: Lens' CreateMultipartUpload (Maybe ObjectCannedACL)

The canned ACL to apply to the object.

cmu2Bucket :: Lens' CreateMultipartUpload BucketName

cmu2CacheControl :: Lens' CreateMultipartUpload (Maybe Text)

Specifies caching behavior along the request/reply chain.

cmu2ContentDisposition :: Lens' CreateMultipartUpload (Maybe Text)

Specifies presentational information for the object.

cmu2ContentEncoding :: Lens' CreateMultipartUpload (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

cmu2ContentLanguage :: Lens' CreateMultipartUpload (Maybe Text)

The language the content is in.

cmu2ContentType :: Lens' CreateMultipartUpload (Maybe Text)

A standard MIME type describing the format of the object data.

cmu2Expires :: Lens' CreateMultipartUpload (Maybe RFC822)

The date and time at which the object is no longer cacheable.

cmu2GrantFullControl :: Lens' CreateMultipartUpload (Maybe Text)

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

cmu2GrantRead :: Lens' CreateMultipartUpload (Maybe Text)

Allows grantee to read the object data and its metadata.

cmu2GrantReadACP :: Lens' CreateMultipartUpload (Maybe Text)

Allows grantee to read the object ACL.

cmu2GrantWriteACP :: Lens' CreateMultipartUpload (Maybe Text)

Allows grantee to write the ACL for the applicable object.

cmu2Key :: Lens' CreateMultipartUpload ObjectKey

cmu2Metadata :: Lens' CreateMultipartUpload (Map Text Text)

A map of metadata to store with the object in S3.

cmu2ServerSideEncryption :: Lens' CreateMultipartUpload (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3.

cmu2StorageClass :: Lens' CreateMultipartUpload (Maybe StorageClass)

The type of storage to use for the object. Defaults to STANDARD.

cmu2WebsiteRedirectLocation :: Lens' CreateMultipartUpload (Maybe Text)

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

cmu2SSECustomerAlgorithm :: Lens' CreateMultipartUpload (Maybe Text)

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

cmu2SSECustomerKey :: Lens' CreateMultipartUpload (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

cmu2SSECustomerKeyMD5 :: Lens' CreateMultipartUpload (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

Response

data CreateMultipartUploadResponse

Response lenses

cmursrsBucket :: Lens' CreateMultipartUploadResponse (Maybe BucketName)

Name of the bucket to which the multipart upload was initiated.

cmursrsKey :: Lens' CreateMultipartUploadResponse (Maybe ObjectKey)

Object key for which the multipart upload was initiated.

cmursrsUploadId :: Lens' CreateMultipartUploadResponse (Maybe Text)

ID for the initiated multipart upload.

cmursrsServerSideEncryption :: Lens' CreateMultipartUploadResponse (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3.

cmursrsSSECustomerAlgorithm :: Lens' CreateMultipartUploadResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

cmursrsSSECustomerKeyMD5 :: Lens' CreateMultipartUploadResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.