Network.AWS.S3.V2006_03_01.DeleteBucketLifecycle

Description

Deletes the lifecycle configuration from the bucket.

Request

data DeleteBucketLifecycle

Request constructor

mkDeleteBucketLifecycle

Arguments

:: BucketName
dblBucket

-> DeleteBucketLifecycle

Smart constructor for the minimum required parameters to construct a valid DeleteBucketLifecycle request.

The fields accessible through corresponding lenses are:

Bucket :: BucketName

Request lenses

dblBucket :: Lens' DeleteBucketLifecycle BucketName

Response

data DeleteBucketLifecycleResponse

Response constructor

mkDeleteBucketLifecycleResponse :: DeleteBucketLifecycleResponse

Smart constructor for the minimum required parameters to construct a valid DeleteBucketLifecycleResponse response.

This constructor is provided for convenience and testing purposes.