Network.AWS.S3.V2006_03_01.DeleteObjects

Description

This operation enables you to delete multiple objects from a bucket using a single HTTP request. You may specify up to 1000 keys.

Documentation

type DeleteMultipleObjects = DeleteObjects

deleteObjects

Arguments

:: Delete
_dosDelete

-> BucketName
_dosBucket

-> DeleteObjects

Minimum specification for a DeleteObjects request.

data DeleteObjects

Constructors

DeleteObjects

Fields

_dosDelete :: Delete

_dosBucket :: BucketName

_dosMFA :: Maybe Text

The concatenation of the authentication device's serial number, a space, and the value that is displayed on your authentication device.

dosMFA :: Lens' DeleteObjects (Maybe Text)

dosDelete :: Lens' DeleteObjects Delete

dosBucket :: Lens' DeleteObjects BucketName

data DeleteObjectsResponse

Constructors

DeleteObjectsResponse

Fields

_dopDeleted :: [DeletedObject]

_dopErrors :: [Error]

dopErrors :: Lens' DeleteObjectsResponse [Error]

dopDeleted :: Lens' DeleteObjectsResponse [DeletedObject]