Network.AWS.S3.V2006_03_01.GetBucketLogging

Description

Returns the logging status of a bucket and the permissions users have to view and modify that status. To use GET, you must be the bucket owner.

Documentation

data GetBucketLogging

Constructors

GetBucketLogging

Fields

_gbltBucket :: BucketName

gbltBucket :: Iso' GetBucketLogging BucketName

data GetBucketLoggingResponse

Constructors

GetBucketLoggingResponse

Fields

_gblqLoggingEnabled :: Maybe LoggingEnabled

gblqLoggingEnabled :: Iso' GetBucketLoggingResponse (Maybe LoggingEnabled)