Network.AWS.S3.V2006_03_01.GetBucketWebsite

Description

Returns the website configuration for a bucket.

Request

data GetBucketWebsite

Request constructor

mkGetBucketWebsite

Arguments

:: BucketName
gbwBucket

-> GetBucketWebsite

Smart constructor for the minimum required parameters to construct a valid GetBucketWebsite request.

The fields accessible through corresponding lenses are:

Bucket :: BucketName

Request lenses

gbwBucket :: Lens' GetBucketWebsite BucketName

Response

data GetBucketWebsiteResponse

Response constructor

mkGetBucketWebsiteResponse :: GetBucketWebsiteResponse

Smart constructor for the minimum required parameters to construct a valid GetBucketWebsiteResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

RedirectAllRequestsTo :: Maybe RedirectAllRequestsTo

IndexDocument :: Maybe IndexDocument

ErrorDocument :: Maybe ErrorDocument

RoutingRules :: [RoutingRule]

Response lenses

gbwrRedirectAllRequestsTo :: Lens' GetBucketWebsiteResponse (Maybe RedirectAllRequestsTo)

gbwrIndexDocument :: Lens' GetBucketWebsiteResponse (Maybe IndexDocument)

gbwrErrorDocument :: Lens' GetBucketWebsiteResponse (Maybe ErrorDocument)

gbwrRoutingRules :: Lens' GetBucketWebsiteResponse [RoutingRule]