Network.AWS.S3.V2006_03_01.HeadBucket

Description

This operation is useful to determine if a bucket exists and you have permission to access it.

Request

data HeadBucket

Request constructor

mkHeadBucket

Arguments

:: BucketName
hbBucket

-> HeadBucket

Smart constructor for the minimum required parameters to construct a valid HeadBucket request.

The fields accessible through corresponding lenses are:

Bucket :: BucketName

Request lenses

hbBucket :: Lens' HeadBucket BucketName

Response

data HeadBucketResponse

Response constructor

mkHeadBucketResponse :: HeadBucketResponse

Smart constructor for the minimum required parameters to construct a valid HeadBucketResponse response.

This constructor is provided for convenience and testing purposes.