Network.AWS.S3.V2006_03_01.ListMultipartUploads

Description

This operation lists in-progress multipart uploads.

Request

data ListMultipartUploads

Request constructor

mkListMultipartUploads

Arguments

:: BucketName
lmuBucket

-> ListMultipartUploads

Smart constructor for the minimum required parameters to construct a valid ListMultipartUploads request.

Request lenses

lmuBucket :: Lens' ListMultipartUploads BucketName

lmuDelimiter :: Lens' ListMultipartUploads (Maybe Char)

Character you use to group keys.

lmuEncodingType :: Lens' ListMultipartUploads (Maybe EncodingType)

Requests Amazon S3 to encode the object keys in the response and specifies the encoding method to use. An object key may contain any Unicode character; however, XML 1.0 parser cannot parse some characters, such as characters with an ASCII value from 0 to 10. For characters that are not supported in XML 1.0, you can add this parameter to request that Amazon S3 encode the keys in the response.

lmuKeyMarker :: Lens' ListMultipartUploads (Maybe Text)

Together with upload-id-marker, this parameter specifies the multipart upload after which listing should begin.

lmuMaxUploads :: Lens' ListMultipartUploads (Maybe Integer)

Sets the maximum number of multipart uploads, from 1 to 1,000, to return in the response body. 1,000 is the maximum number of uploads that can be returned in a response.

lmuPrefix :: Lens' ListMultipartUploads (Maybe Text)

Lists in-progress uploads only for those keys that begin with the specified prefix.

lmuUploadIdMarker :: Lens' ListMultipartUploads (Maybe Text)

Together with key-marker, specifies the multipart upload after which listing should begin. If key-marker is not specified, the upload-id-marker parameter is ignored.

Response

data ListMultipartUploadsResponse

Response lenses

lmurBucket :: Lens' ListMultipartUploadsResponse (Maybe BucketName)

Name of the bucket to which the multipart upload was initiated.

lmurKeyMarker :: Lens' ListMultipartUploadsResponse (Maybe Text)

The key at or after which the listing began.

lmurUploadIdMarker :: Lens' ListMultipartUploadsResponse (Maybe Text)

Upload ID after which listing began.

lmurNextKeyMarker :: Lens' ListMultipartUploadsResponse (Maybe Text)

When a list is truncated, this element specifies the value that should be used for the key-marker request parameter in a subsequent request.

lmurPrefix :: Lens' ListMultipartUploadsResponse (Maybe Text)

When a prefix is provided in the request, this field contains the specified prefix. The result contains only keys starting with the specified prefix.

lmurNextUploadIdMarker :: Lens' ListMultipartUploadsResponse (Maybe Text)

When a list is truncated, this element specifies the value that should be used for the upload-id-marker request parameter in a subsequent request.

lmurMaxUploads :: Lens' ListMultipartUploadsResponse (Maybe Integer)

Maximum number of multipart uploads that could have been included in the response.

lmurIsTruncated :: Lens' ListMultipartUploadsResponse Bool

Indicates whether the returned list of multipart uploads is truncated. A value of true indicates that the list was truncated. The list can be truncated if the number of multipart uploads exceeds the limit allowed or specified by max uploads.

lmurUploads :: Lens' ListMultipartUploadsResponse [MultipartUpload]

lmurCommonPrefixes :: Lens' ListMultipartUploadsResponse [CommonPrefix]

lmurEncodingType :: Lens' ListMultipartUploadsResponse (Maybe EncodingType)

Encoding type used by Amazon S3 to encode object keys in the response.