Network.AWS.S3.V2006_03_01.ListObjectVersions

Description

Returns metadata about all of the versions of objects in a bucket.

Request

data ListObjectVersions

Request alias

type GetBucketObjectVersions = ListObjectVersions

Request constructor

mkListObjectVersions

Arguments

:: BucketName
lovBucket

-> ListObjectVersions

Smart constructor for the minimum required parameters to construct a valid ListObjectVersions request.

Request lenses

lovBucket :: Lens' ListObjectVersions BucketName

lovDelimiter :: Lens' ListObjectVersions (Maybe Char)

A delimiter is a character you use to group keys.

lovEncodingType :: Lens' ListObjectVersions (Maybe EncodingType)

Requests Amazon S3 to encode the object keys in the response and specifies the encoding method to use. An object key may contain any Unicode character; however, XML 1.0 parser cannot parse some characters, such as characters with an ASCII value from 0 to 10. For characters that are not supported in XML 1.0, you can add this parameter to request that Amazon S3 encode the keys in the response.

lovKeyMarker :: Lens' ListObjectVersions (Maybe Text)

Specifies the key to start with when listing objects in a bucket.

lovMaxKeys :: Lens' ListObjectVersions (Maybe Integer)

Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.

lovPrefix :: Lens' ListObjectVersions (Maybe Text)

Limits the response to keys that begin with the specified prefix.

lovVersionIdMarker :: Lens' ListObjectVersions (Maybe Text)

Specifies the object version you want to start listing from.

Response

data ListObjectVersionsResponse

Response lenses

lovrsIsTruncated :: Lens' ListObjectVersionsResponse Bool

A flag that indicates whether or not Amazon S3 returned all of the results that satisfied the search criteria. If your results were truncated, you can make a follow-up paginated request using the NextKeyMarker and NextVersionIdMarker response parameters as a starting place in another request to return the rest of the results.

lovrsKeyMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

Marks the last Key returned in a truncated response.

lovrsVersionIdMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

lovrsNextKeyMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

Use this value for the key marker request parameter in a subsequent request.

lovrsNextVersionIdMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

Use this value for the next version id marker parameter in a subsequent request.

lovrsVersions :: Lens' ListObjectVersionsResponse [ObjectVersion]

lovrsDeleteMarkers :: Lens' ListObjectVersionsResponse [DeleteMarkerEntry]

lovrsName :: Lens' ListObjectVersionsResponse (Maybe BucketName)

lovrsPrefix :: Lens' ListObjectVersionsResponse (Maybe Text)

lovrsMaxKeys :: Lens' ListObjectVersionsResponse (Maybe Integer)

lovrsCommonPrefixes :: Lens' ListObjectVersionsResponse [CommonPrefix]

lovrsEncodingType :: Lens' ListObjectVersionsResponse (Maybe EncodingType)

Encoding type used by Amazon S3 to encode object keys in the response.