Network.AWS.S3.V2006_03_01.ListObjects

Description

Returns some or all (up to 1000) of the objects in a bucket. You can use the request parameters as selection criteria to return a subset of the objects in a bucket.

Documentation

type GetBucket = ListObjects

listObjects

Arguments

:: BucketName
_lorBucket

-> ListObjects

Minimum specification for a ListObjects request.

data ListObjects

Constructors

ListObjects

Fields

_lorBucket :: BucketName

_lorDelimiter :: Maybe Char

A delimiter is a character you use to group keys.

_lorEncodingType :: Maybe EncodingType

Requests Amazon S3 to encode the object keys in the response and specifies the encoding method to use. An object key may contain any Unicode character; however, XML 1.0 parser cannot parse some characters, such as characters with an ASCII value from 0 to 10. For characters that are not supported in XML 1.0, you can add this parameter to request that Amazon S3 encode the keys in the response.

_lorMarker :: Maybe Text

Specifies the key to start with when listing objects in a bucket.

_lorMaxKeys :: Maybe Integer

Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.

_lorPrefix :: Maybe Text

Limits the response to keys that begin with the specified prefix.

lorPrefix :: Lens' ListObjects (Maybe Text)

lorMaxKeys :: Lens' ListObjects (Maybe Integer)

lorMarker :: Lens' ListObjects (Maybe Text)

lorEncodingType :: Lens' ListObjects (Maybe EncodingType)

lorDelimiter :: Lens' ListObjects (Maybe Char)

lorBucket :: Lens' ListObjects BucketName

data ListObjectsResponse

Constructors

ListObjectsResponse

Fields

_looIsTruncated :: Bool

A flag that indicates whether or not Amazon S3 returned all of the results that satisfied the search criteria.

_looName :: BucketName

_looCommonPrefixes :: [CommonPrefix]

_looEncodingType :: Maybe EncodingType

Encoding type used by Amazon S3 to encode object keys in the response.

_looMarker :: Maybe Text

_looMaxKeys :: Maybe Integer

_looNextMarker :: Maybe Text

When response is truncated (the IsTruncated element value in the response is true), you can use the key name in this field as marker in the subsequent request to get next set of objects. Amazon S3 lists objects in alphabetical order Note: This element is returned only if you have delimiter request parameter specified. If response does not include the NextMaker and it is truncated, you can use the value of the last Key in the response as the marker in the subsequent request to get the next set of object keys.

_looContents :: [Object]

_looPrefix :: Maybe Text

looPrefix :: Lens' ListObjectsResponse (Maybe Text)

looNextMarker :: Lens' ListObjectsResponse (Maybe Text)

looName :: Lens' ListObjectsResponse BucketName

looMaxKeys :: Lens' ListObjectsResponse (Maybe Integer)

looMarker :: Lens' ListObjectsResponse (Maybe Text)

looIsTruncated :: Lens' ListObjectsResponse Bool

looEncodingType :: Lens' ListObjectsResponse (Maybe EncodingType)

looContents :: Lens' ListObjectsResponse [Object]

looCommonPrefixes :: Lens' ListObjectsResponse [CommonPrefix]