Network.AWS.S3.V2006_03_01.ListParts

Description

Lists the parts that have been uploaded for a specific multipart upload.

Request

data ListParts

Request constructor

mkListParts

Arguments

:: BucketName
lpBucket

-> ObjectKey
lpKey

-> Text
lpUploadId

-> ListParts

Smart constructor for the minimum required parameters to construct a valid ListParts request.

The fields accessible through corresponding lenses are:

Bucket :: BucketName

Key :: ObjectKey

MaxParts :: Maybe Integer

PartNumberMarker :: Maybe Integer

UploadId :: Text

Request lenses

lpBucket :: Lens' ListParts BucketName

lpKey :: Lens' ListParts ObjectKey

lpMaxParts :: Lens' ListParts (Maybe Integer)

Sets the maximum number of parts to return.

lpPartNumberMarker :: Lens' ListParts (Maybe Integer)

Specifies the part after which listing should begin. Only parts with higher part numbers will be listed.

lpUploadId :: Lens' ListParts Text

Upload ID identifying the multipart upload whose parts are being listed.

Response

data ListPartsResponse

Response constructor

mkListPartsResponse

Arguments

:: Bool
lprIsTruncated

-> ListPartsResponse

Smart constructor for the minimum required parameters to construct a valid ListPartsResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Bucket :: Maybe BucketName

Key :: Maybe ObjectKey

UploadId :: Maybe Text

PartNumberMarker :: Maybe Integer

NextPartNumberMarker :: Maybe Integer

MaxParts :: Maybe Integer

IsTruncated :: Bool

Parts :: [Part]

Initiator :: Maybe Initiator

Owner :: Maybe Owner

StorageClass :: Maybe StorageClass

Response lenses

lprBucket :: Lens' ListPartsResponse (Maybe BucketName)

Name of the bucket to which the multipart upload was initiated.

lprKey :: Lens' ListPartsResponse (Maybe ObjectKey)

Object key for which the multipart upload was initiated.

lprUploadId :: Lens' ListPartsResponse (Maybe Text)

Upload ID identifying the multipart upload whose parts are being listed.

lprPartNumberMarker :: Lens' ListPartsResponse (Maybe Integer)

Part number after which listing begins.

lprNextPartNumberMarker :: Lens' ListPartsResponse (Maybe Integer)

When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

lprMaxParts :: Lens' ListPartsResponse (Maybe Integer)

Maximum number of parts that were allowed in the response.

lprIsTruncated :: Lens' ListPartsResponse Bool

Indicates whether the returned list of parts is truncated.

lprParts :: Lens' ListPartsResponse [Part]

lprInitiator :: Lens' ListPartsResponse (Maybe Initiator)

Identifies who initiated the multipart upload.

lprOwner :: Lens' ListPartsResponse (Maybe Owner)

lprStorageClass :: Lens' ListPartsResponse (Maybe StorageClass)

The class of storage used to store the object.