Network.AWS.S3.V2006_03_01.Monadic

Description

Amazon Simple Storage Service is storage for the Internet. Amazon S3 has a simple web services interface that you can use to store and retrieve any amount of data, at any time, from anywhere on the web. It gives any developer access to the same highly scalable, reliable, fast, inexpensive data storage infrastructure that Amazon uses to run its own global network of web sites. The service aims to maximize benefits of scale and to pass those benefits on to developers.

The State operator variants from Control.Lens.Setter such as .= can be used to modify any additional request parameters before sending.

AbortMultipartUpload

Aborts a multipart upload. To verify that all parts have been removed, so you don't get charged for the part storage, you should call the List Parts operation and ensure the parts list is empty.

See: AbortMultipartUpload

abortMultipartUpload

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
amuBucket

-> ObjectKey
amuKey

-> Text
amuUploadId

-> State AbortMultipartUpload a

-> m AbortMultipartUploadResponse

abortMultipartUploadCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
amuBucket

-> ObjectKey
amuKey

-> Text
amuUploadId

-> State AbortMultipartUpload a

-> m (Either ServiceEr AbortMultipartUploadResponse)

CompleteMultipartUpload

Completes a multipart upload by assembling previously uploaded parts.

See: CompleteMultipartUpload

completeMultipartUpload

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
cmuBucket

-> ObjectKey
cmuKey

-> Text
cmuUploadId

-> State CompleteMultipartUpload a

-> m CompleteMultipartUploadResponse

completeMultipartUploadCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
cmuBucket

-> ObjectKey
cmuKey

-> Text
cmuUploadId

-> State CompleteMultipartUpload a

-> m (Either ServiceEr CompleteMultipartUploadResponse)

CopyObject

Creates a copy of an object that is already stored in Amazon S3.

See: CopyObject

copyObject

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> ObjectKey
coKey

-> BucketName
coBucket

-> Text
coCopySource

-> State CopyObject a

-> m CopyObjectResponse

copyObjectCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> ObjectKey
coKey

-> BucketName
coBucket

-> Text
coCopySource

-> State CopyObject a

-> m (Either ServiceEr CopyObjectResponse)

CreateBucket

Creates a new bucket.

See: CreateBucket

createBucket

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
cbBucket

-> State CreateBucket a

-> m CreateBucketResponse

createBucketCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
cbBucket

-> State CreateBucket a

-> m (Either ServiceEr CreateBucketResponse)

CreateMultipartUpload

Initiates a multipart upload and returns an upload ID. Note: After you initiate multipart upload and upload one or more parts, you must either complete or abort multipart upload in order to stop getting charged for storage of the uploaded parts. Only after you either complete or abort multipart upload, Amazon S3 frees up the parts storage and stops charging you for the parts storage.

See: CreateMultipartUpload

createMultipartUpload

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> ObjectKey
cmu2Key

-> BucketName
cmu2Bucket

-> State CreateMultipartUpload a

-> m CreateMultipartUploadResponse

createMultipartUploadCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> ObjectKey
cmu2Key

-> BucketName
cmu2Bucket

-> State CreateMultipartUpload a

-> m (Either ServiceEr CreateMultipartUploadResponse)

DeleteBucket

Deletes the bucket. All objects (including all object versions and Delete Markers) in the bucket must be deleted before the bucket itself can be deleted.

See: DeleteBucket

deleteBucket

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
dbBucket

-> State DeleteBucket a

-> m DeleteBucketResponse

deleteBucketCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
dbBucket

-> State DeleteBucket a

-> m (Either ServiceEr DeleteBucketResponse)

DeleteBucketCors

Deletes the cors configuration information set for the bucket.

See: DeleteBucketCors

deleteBucketCors

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
dbcBucket

-> State DeleteBucketCors a

-> m DeleteBucketCorsResponse

deleteBucketCorsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
dbcBucket

-> State DeleteBucketCors a

-> m (Either ServiceEr DeleteBucketCorsResponse)

DeleteBucketLifecycle

Deletes the lifecycle configuration from the bucket.

See: DeleteBucketLifecycle

deleteBucketLifecycle

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
dblBucket

-> State DeleteBucketLifecycle a

-> m DeleteBucketLifecycleResponse

deleteBucketLifecycleCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
dblBucket

-> State DeleteBucketLifecycle a

-> m (Either ServiceEr DeleteBucketLifecycleResponse)

DeleteBucketPolicy

Deletes the policy from the bucket.

See: DeleteBucketPolicy

deleteBucketPolicy

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
dbpBucket

-> State DeleteBucketPolicy a

-> m DeleteBucketPolicyResponse

deleteBucketPolicyCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
dbpBucket

-> State DeleteBucketPolicy a

-> m (Either ServiceEr DeleteBucketPolicyResponse)

DeleteBucketTagging

Deletes the tags from the bucket.

See: DeleteBucketTagging

deleteBucketTagging

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
dbtBucket

-> State DeleteBucketTagging a

-> m DeleteBucketTaggingResponse

deleteBucketTaggingCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
dbtBucket

-> State DeleteBucketTagging a

-> m (Either ServiceEr DeleteBucketTaggingResponse)

DeleteBucketWebsite

This operation removes the website configuration from the bucket.

See: DeleteBucketWebsite

deleteBucketWebsite

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
dbwBucket

-> State DeleteBucketWebsite a

-> m DeleteBucketWebsiteResponse

deleteBucketWebsiteCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
dbwBucket

-> State DeleteBucketWebsite a

-> m (Either ServiceEr DeleteBucketWebsiteResponse)

DeleteObject

Removes the null version (if there is one) of an object and inserts a delete marker, which becomes the latest version of the object. If there isn't a null version, Amazon S3 does not remove any objects.

See: DeleteObject

deleteObject

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
doBucket

-> ObjectKey
doKey

-> State DeleteObject a

-> m DeleteObjectResponse

deleteObjectCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
doBucket

-> ObjectKey
doKey

-> State DeleteObject a

-> m (Either ServiceEr DeleteObjectResponse)

DeleteObjects

This operation enables you to delete multiple objects from a bucket using a single HTTP request. You may specify up to 1000 keys.

See: DeleteObjects

deleteObjects

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
do1Bucket

-> Delete
do1Delete

-> State DeleteObjects a

-> m DeleteObjectsResponse

deleteObjectsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
do1Bucket

-> Delete
do1Delete

-> State DeleteObjects a

-> m (Either ServiceEr DeleteObjectsResponse)

GetBucketAcl

Gets the access control policy for the bucket.

See: GetBucketAcl

getBucketAcl

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
gbaBucket

-> State GetBucketAcl a

-> m GetBucketAclResponse

getBucketAclCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
gbaBucket

-> State GetBucketAcl a

-> m (Either ServiceEr GetBucketAclResponse)

GetBucketCors

Returns the cors configuration for the bucket.

See: GetBucketCors

getBucketCors

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
gbcBucket

-> State GetBucketCors a

-> m GetBucketCorsResponse

getBucketCorsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
gbcBucket

-> State GetBucketCors a

-> m (Either ServiceEr GetBucketCorsResponse)

GetBucketLifecycle

Returns the lifecycle configuration information set on the bucket.

See: GetBucketLifecycle

getBucketLifecycle

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
gblBucket

-> State GetBucketLifecycle a

-> m GetBucketLifecycleResponse

getBucketLifecycleCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
gblBucket

-> State GetBucketLifecycle a

-> m (Either ServiceEr GetBucketLifecycleResponse)

GetBucketLocation

Returns the region the bucket resides in.

See: GetBucketLocation

getBucketLocation

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
gbl1Bucket

-> State GetBucketLocation a

-> m GetBucketLocationResponse

getBucketLocationCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
gbl1Bucket

-> State GetBucketLocation a

-> m (Either ServiceEr GetBucketLocationResponse)

GetBucketLogging

Returns the logging status of a bucket and the permissions users have to view and modify that status. To use GET, you must be the bucket owner.

See: GetBucketLogging

getBucketLogging

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
gbl2Bucket

-> State GetBucketLogging a

-> m GetBucketLoggingResponse

getBucketLoggingCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
gbl2Bucket

-> State GetBucketLogging a

-> m (Either ServiceEr GetBucketLoggingResponse)

GetBucketNotification

Return the notification configuration of a bucket.

See: GetBucketNotification

getBucketNotification

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
gbnBucket

-> State GetBucketNotification a

-> m GetBucketNotificationResponse

getBucketNotificationCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
gbnBucket

-> State GetBucketNotification a

-> m (Either ServiceEr GetBucketNotificationResponse)

GetBucketPolicy

Returns the policy of a specified bucket.

See: GetBucketPolicy

getBucketPolicy

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
gbpBucket

-> State GetBucketPolicy a

-> m GetBucketPolicyResponse

getBucketPolicyCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
gbpBucket

-> State GetBucketPolicy a

-> m (Either ServiceEr GetBucketPolicyResponse)

GetBucketRequestPayment

Returns the request payment configuration of a bucket.

See: GetBucketRequestPayment

getBucketRequestPayment

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
gbrpBucket

-> State GetBucketRequestPayment a

-> m GetBucketRequestPaymentResponse

getBucketRequestPaymentCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
gbrpBucket

-> State GetBucketRequestPayment a

-> m (Either ServiceEr GetBucketRequestPaymentResponse)

GetBucketTagging

Returns the tag set associated with the bucket.

See: GetBucketTagging

getBucketTagging

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
gbtBucket

-> State GetBucketTagging a

-> m GetBucketTaggingResponse

getBucketTaggingCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
gbtBucket

-> State GetBucketTagging a

-> m (Either ServiceEr GetBucketTaggingResponse)

GetBucketVersioning

Returns the versioning state of a bucket.

See: GetBucketVersioning

getBucketVersioning

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
gbvBucket

-> State GetBucketVersioning a

-> m GetBucketVersioningResponse

getBucketVersioningCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
gbvBucket

-> State GetBucketVersioning a

-> m (Either ServiceEr GetBucketVersioningResponse)

GetBucketWebsite

Returns the website configuration for a bucket.

See: GetBucketWebsite

getBucketWebsite

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
gbwBucket

-> State GetBucketWebsite a

-> m GetBucketWebsiteResponse

getBucketWebsiteCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
gbwBucket

-> State GetBucketWebsite a

-> m (Either ServiceEr GetBucketWebsiteResponse)

GetObject

Retrieves objects from Amazon S3.

See: GetObject

getObject

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
goBucket

-> ObjectKey
goKey

-> State GetObject a

-> m GetObjectResponse

getObjectCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
goBucket

-> ObjectKey
goKey

-> State GetObject a

-> m (Either ServiceEr GetObjectResponse)

GetObjectAcl

Returns the access control list (ACL) of an object.

See: GetObjectAcl

getObjectAcl

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
goaBucket

-> ObjectKey
goaKey

-> State GetObjectAcl a

-> m GetObjectAclResponse

getObjectAclCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
goaBucket

-> ObjectKey
goaKey

-> State GetObjectAcl a

-> m (Either ServiceEr GetObjectAclResponse)

GetObjectTorrent

Return torrent files from a bucket.

See: GetObjectTorrent

getObjectTorrent

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
gotBucket

-> ObjectKey
gotKey

-> State GetObjectTorrent a

-> m GetObjectTorrentResponse

getObjectTorrentCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
gotBucket

-> ObjectKey
gotKey

-> State GetObjectTorrent a

-> m (Either ServiceEr GetObjectTorrentResponse)

HeadBucket

This operation is useful to determine if a bucket exists and you have permission to access it.

See: HeadBucket

headBucket

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
hbBucket

-> State HeadBucket a

-> m HeadBucketResponse

headBucketCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
hbBucket

-> State HeadBucket a

-> m (Either ServiceEr HeadBucketResponse)

HeadObject

The HEAD operation retrieves metadata from an object without returning the object itself. This operation is useful if you're only interested in an object's metadata. To use HEAD, you must have READ access to the object.

See: HeadObject

headObject

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
hoBucket

-> ObjectKey
hoKey

-> State HeadObject a

-> m HeadObjectResponse

headObjectCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
hoBucket

-> ObjectKey
hoKey

-> State HeadObject a

-> m (Either ServiceEr HeadObjectResponse)

ListBuckets

Returns a list of all buckets owned by the authenticated sender of the request.

See: ListBuckets

listBuckets :: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m) => State ListBuckets a -> m ListBucketsResponse

listBucketsCatch :: (MonadCatch m, MonadResource m, MonadReader Env m) => State ListBuckets a -> m (Either ServiceEr ListBucketsResponse)

ListMultipartUploads

This operation lists in-progress multipart uploads.

See: ListMultipartUploads

listMultipartUploads

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env (ResumableSource m))

=> BucketName
lmuBucket

-> State ListMultipartUploads a

-> ResumableSource m ListMultipartUploadsResponse

listMultipartUploadsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env (ResumableSource m))

=> BucketName
lmuBucket

-> State ListMultipartUploads a

-> ResumableSource m (Either ServiceEr ListMultipartUploadsResponse)

ListObjectVersions

Returns metadata about all of the versions of objects in a bucket.

See: ListObjectVersions

listObjectVersions

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env (ResumableSource m))

=> BucketName
lovBucket

-> State ListObjectVersions a

-> ResumableSource m ListObjectVersionsResponse

listObjectVersionsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env (ResumableSource m))

=> BucketName
lovBucket

-> State ListObjectVersions a

-> ResumableSource m (Either ServiceEr ListObjectVersionsResponse)

ListObjects

Returns some or all (up to 1000) of the objects in a bucket. You can use the request parameters as selection criteria to return a subset of the objects in a bucket.

See: ListObjects

listObjects

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env (ResumableSource m))

=> BucketName
loBucket

-> State ListObjects a

-> ResumableSource m ListObjectsResponse

listObjectsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env (ResumableSource m))

=> BucketName
loBucket

-> State ListObjects a

-> ResumableSource m (Either ServiceEr ListObjectsResponse)

ListParts

Lists the parts that have been uploaded for a specific multipart upload.

See: ListParts

listParts

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env (ResumableSource m))

=> BucketName
lpBucket

-> ObjectKey
lpKey

-> Text
lpUploadId

-> State ListParts a

-> ResumableSource m ListPartsResponse

listPartsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env (ResumableSource m))

=> BucketName
lpBucket

-> ObjectKey
lpKey

-> Text
lpUploadId

-> State ListParts a

-> ResumableSource m (Either ServiceEr ListPartsResponse)

PutBucketAcl

Sets the permissions on a bucket using access control lists (ACL).

See: PutBucketAcl

putBucketAcl

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
pbaBucket

-> State PutBucketAcl a

-> m PutBucketAclResponse

putBucketAclCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
pbaBucket

-> State PutBucketAcl a

-> m (Either ServiceEr PutBucketAclResponse)

PutBucketCors

Sets the cors configuration for a bucket.

See: PutBucketCors

putBucketCors

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
pbcBucket

-> State PutBucketCors a

-> m PutBucketCorsResponse

putBucketCorsCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
pbcBucket

-> State PutBucketCors a

-> m (Either ServiceEr PutBucketCorsResponse)

PutBucketLifecycle

Sets lifecycle configuration for your bucket. If a lifecycle configuration exists, it replaces it.

See: PutBucketLifecycle

putBucketLifecycle

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
pblBucket

-> State PutBucketLifecycle a

-> m PutBucketLifecycleResponse

putBucketLifecycleCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
pblBucket

-> State PutBucketLifecycle a

-> m (Either ServiceEr PutBucketLifecycleResponse)

PutBucketLogging

Set the logging parameters for a bucket and to specify permissions for who can view and modify the logging parameters. To set the logging status of a bucket, you must be the bucket owner.

See: PutBucketLogging

putBucketLogging

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
pbl1Bucket

-> BucketLoggingStatus
pbl1BucketLoggingStatus

-> State PutBucketLogging a

-> m PutBucketLoggingResponse

putBucketLoggingCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
pbl1Bucket

-> BucketLoggingStatus
pbl1BucketLoggingStatus

-> State PutBucketLogging a

-> m (Either ServiceEr PutBucketLoggingResponse)

PutBucketNotification

Enables notifications of specified events for a bucket.

See: PutBucketNotification

putBucketNotification

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
pbnBucket

-> NotificationConfiguration
pbnNotificationConfiguration

-> State PutBucketNotification a

-> m PutBucketNotificationResponse

putBucketNotificationCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
pbnBucket

-> NotificationConfiguration
pbnNotificationConfiguration

-> State PutBucketNotification a

-> m (Either ServiceEr PutBucketNotificationResponse)

PutBucketPolicy

Replaces a policy on a bucket. If the bucket already has a policy, the one in this request completely replaces it.

See: PutBucketPolicy

putBucketPolicy

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
pbpBucket

-> Text
pbpPolicy

-> State PutBucketPolicy a

-> m PutBucketPolicyResponse

putBucketPolicyCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
pbpBucket

-> Text
pbpPolicy

-> State PutBucketPolicy a

-> m (Either ServiceEr PutBucketPolicyResponse)

PutBucketRequestPayment

Sets the request payment configuration for a bucket. By default, the bucket owner pays for downloads from the bucket. This configuration parameter enables the bucket owner (only) to specify that the person requesting the download will be charged for the download.

See: PutBucketRequestPayment

putBucketRequestPayment

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
pbrpBucket

-> RequestPaymentConfiguration
pbrpRequestPaymentConfiguration

-> State PutBucketRequestPayment a

-> m PutBucketRequestPaymentResponse

putBucketRequestPaymentCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
pbrpBucket

-> RequestPaymentConfiguration
pbrpRequestPaymentConfiguration

-> State PutBucketRequestPayment a

-> m (Either ServiceEr PutBucketRequestPaymentResponse)

PutBucketTagging

Sets the tags for a bucket.

See: PutBucketTagging

putBucketTagging

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
pbtBucket

-> Tagging
pbtTagging

-> State PutBucketTagging a

-> m PutBucketTaggingResponse

putBucketTaggingCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
pbtBucket

-> Tagging
pbtTagging

-> State PutBucketTagging a

-> m (Either ServiceEr PutBucketTaggingResponse)

PutBucketVersioning

Sets the versioning state of an existing bucket. To set the versioning state, you must be the bucket owner.

See: PutBucketVersioning

putBucketVersioning

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
pbvBucket

-> VersioningConfiguration
pbvVersioningConfiguration

-> State PutBucketVersioning a

-> m PutBucketVersioningResponse

putBucketVersioningCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
pbvBucket

-> VersioningConfiguration
pbvVersioningConfiguration

-> State PutBucketVersioning a

-> m (Either ServiceEr PutBucketVersioningResponse)

PutBucketWebsite

Set the website configuration for a bucket.

See: PutBucketWebsite

putBucketWebsite

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
pbwBucket

-> WebsiteConfiguration
pbwWebsiteConfiguration

-> State PutBucketWebsite a

-> m PutBucketWebsiteResponse

putBucketWebsiteCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
pbwBucket

-> WebsiteConfiguration
pbwWebsiteConfiguration

-> State PutBucketWebsite a

-> m (Either ServiceEr PutBucketWebsiteResponse)

PutObject

Adds an object to a bucket.

See: PutObject

putObject

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> ObjectKey
poKey

-> RqBody
poBody

-> BucketName
poBucket

-> State PutObject a

-> m PutObjectResponse

putObjectCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> ObjectKey
poKey

-> RqBody
poBody

-> BucketName
poBucket

-> State PutObject a

-> m (Either ServiceEr PutObjectResponse)

PutObjectAcl

uses the acl subresource to set the access control list (ACL) permissions for an object that already exists in a bucket.

See: PutObjectAcl

putObjectAcl

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> ObjectKey
poaKey

-> BucketName
poaBucket

-> State PutObjectAcl a

-> m PutObjectAclResponse

putObjectAclCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> ObjectKey
poaKey

-> BucketName
poaBucket

-> State PutObjectAcl a

-> m (Either ServiceEr PutObjectAclResponse)

RestoreObject

Restores an archived copy of an object back into Amazon S3.

See: RestoreObject

restoreObject

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
roBucket

-> ObjectKey
roKey

-> State RestoreObject a

-> m RestoreObjectResponse

restoreObjectCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
roBucket

-> ObjectKey
roKey

-> State RestoreObject a

-> m (Either ServiceEr RestoreObjectResponse)

UploadPart

Uploads a part in a multipart upload. Note: After you initiate multipart upload and upload one or more parts, you must either complete or abort multipart upload in order to stop getting charged for storage of the uploaded parts. Only after you either complete or abort multipart upload, Amazon S3 frees up the parts storage and stops charging you for the parts storage.

See: UploadPart

uploadPart

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> RqBody
upBody

-> BucketName
upBucket

-> ObjectKey
upKey

-> Integer
upPartNumber

-> Text
upUploadId

-> State UploadPart a

-> m UploadPartResponse

uploadPartCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> RqBody
upBody

-> BucketName
upBucket

-> ObjectKey
upKey

-> Integer
upPartNumber

-> Text
upUploadId

-> State UploadPart a

-> m (Either ServiceEr UploadPartResponse)

UploadPartCopy

Uploads a part by copying data from an existing object as data source.

See: UploadPartCopy

uploadPartCopy

Arguments

:: (MonadCatch m, MonadResource m, MonadError Error m, MonadReader Env m)

=> BucketName
upcBucket

-> Text
upcUploadId

-> Text
upcCopySource

-> ObjectKey
upcKey

-> Integer
upcPartNumber

-> State UploadPartCopy a

-> m UploadPartCopyResponse

uploadPartCopyCatch

Arguments

:: (MonadCatch m, MonadResource m, MonadReader Env m)

=> BucketName
upcBucket

-> Text
upcUploadId

-> Text
upcCopySource

-> ObjectKey
upcKey

-> Integer
upcPartNumber

-> State UploadPartCopy a

-> m (Either ServiceEr UploadPartCopyResponse)

Re-exported

module Network.AWS.S3.V2006_03_01

(.=) :: MonadState s m => ASetter s s a b -> b -> m ()

(?=) :: MonadState s m => ASetter s s a (Maybe b) -> b -> m ()

(<>=) :: (MonadState s m, Monoid a) => ASetter' s a -> a -> m ()

(%=) :: (Profunctor p, MonadState s m) => Setting p s s a b -> p a b -> m ()