Network.AWS.S3.V2006_03_01.PutBucketAcl

Description

Sets the permissions on a bucket using access control lists (ACL).

Documentation

putBucketAcl

Arguments

:: BucketName
_pbarBucket

-> PutBucketAcl

Minimum specification for a PutBucketAcl request.

data PutBucketAcl

Constructors

PutBucketAcl

Fields

_pbarBucket :: BucketName

_pbarAccessControlPolicy :: Maybe AccessControlPolicy

_pbarACL :: Maybe BucketCannedACL

The canned ACL to apply to the bucket.

_pbarContentMD5 :: Maybe Text

_pbarGrantFullControl :: Maybe Text

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

_pbarGrantRead :: Maybe Text

Allows grantee to list the objects in the bucket.

_pbarGrantReadACP :: Maybe Text

Allows grantee to read the bucket ACL.

_pbarGrantWrite :: Maybe Text

Allows grantee to create, overwrite, and delete any object in the bucket.

_pbarGrantWriteACP :: Maybe Text

Allows grantee to write the ACL for the applicable bucket.

pbarGrantWriteACP :: Lens' PutBucketAcl (Maybe Text)

pbarGrantWrite :: Lens' PutBucketAcl (Maybe Text)

pbarGrantReadACP :: Lens' PutBucketAcl (Maybe Text)

pbarGrantRead :: Lens' PutBucketAcl (Maybe Text)

pbarGrantFullControl :: Lens' PutBucketAcl (Maybe Text)

pbarContentMD5 :: Lens' PutBucketAcl (Maybe Text)

pbarBucket :: Lens' PutBucketAcl BucketName

pbarAccessControlPolicy :: Lens' PutBucketAcl (Maybe AccessControlPolicy)

pbarACL :: Lens' PutBucketAcl (Maybe BucketCannedACL)

data PutBucketAclResponse

Constructors

PutBucketAclResponse