Network.AWS.S3.V2006_03_01.PutBucketLogging

Description

Set the logging parameters for a bucket and to specify permissions for who can view and modify the logging parameters. To set the logging status of a bucket, you must be the bucket owner.

Request

data PutBucketLogging

Request constructor

mkPutBucketLogging

Arguments

:: BucketName
pbl1Bucket

-> BucketLoggingStatus
pbl1BucketLoggingStatus

-> PutBucketLogging

Smart constructor for the minimum required parameters to construct a valid PutBucketLogging request.

Request lenses

pbl1Bucket :: Lens' PutBucketLogging BucketName

pbl1BucketLoggingStatus :: Lens' PutBucketLogging BucketLoggingStatus

pbl1ContentMD5 :: Lens' PutBucketLogging (Maybe Text)

Response

data PutBucketLoggingResponse

Response constructor

mkPutBucketLoggingResponse :: PutBucketLoggingResponse

Smart constructor for the minimum required parameters to construct a valid PutBucketLoggingResponse response.

This constructor is provided for convenience and testing purposes.