Network.AWS.S3.V2006_03_01.PutBucketVersioning

Description

Sets the versioning state of an existing bucket. To set the versioning state, you must be the bucket owner.

Request

data PutBucketVersioning

Request constructor

mkPutBucketVersioning

Arguments

:: BucketName
pbvBucket

-> VersioningConfiguration
pbvVersioningConfiguration

-> PutBucketVersioning

Smart constructor for the minimum required parameters to construct a valid PutBucketVersioning request.

Request lenses

pbvBucket :: Lens' PutBucketVersioning BucketName

pbvContentMD5 :: Lens' PutBucketVersioning (Maybe Text)

pbvMFA :: Lens' PutBucketVersioning (Maybe Text)

The concatenation of the authentication device's serial number, a space, and the value that is displayed on your authentication device.

pbvVersioningConfiguration :: Lens' PutBucketVersioning VersioningConfiguration

Response

data PutBucketVersioningResponse

Response constructor

mkPutBucketVersioningResponse :: PutBucketVersioningResponse

Smart constructor for the minimum required parameters to construct a valid PutBucketVersioningResponse response.

This constructor is provided for convenience and testing purposes.