Network.AWS.S3.V2006_03_01.PutObjectAcl

Description

uses the acl subresource to set the access control list (ACL) permissions for an object that already exists in a bucket.

Request

data PutObjectAcl

Request constructor

mkPutObjectAcl

Arguments

:: ObjectKey
poaKey

-> BucketName
poaBucket

-> PutObjectAcl

Smart constructor for the minimum required parameters to construct a valid PutObjectAcl request.

The fields accessible through corresponding lenses are:

ACL :: Maybe ObjectCannedACL

AccessControlPolicy :: Maybe AccessControlPolicy

Bucket :: BucketName

ContentMD5 :: Maybe Text

GrantFullControl :: Maybe Text

GrantRead :: Maybe Text

GrantReadACP :: Maybe Text

GrantWrite :: Maybe Text

GrantWriteACP :: Maybe Text

Key :: ObjectKey

Request lenses

poaACL :: Lens' PutObjectAcl (Maybe ObjectCannedACL)

The canned ACL to apply to the object.

poaAccessControlPolicy :: Lens' PutObjectAcl (Maybe AccessControlPolicy)

poaBucket :: Lens' PutObjectAcl BucketName

poaContentMD5 :: Lens' PutObjectAcl (Maybe Text)

poaGrantFullControl :: Lens' PutObjectAcl (Maybe Text)

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

poaGrantRead :: Lens' PutObjectAcl (Maybe Text)

Allows grantee to list the objects in the bucket.

poaGrantReadACP :: Lens' PutObjectAcl (Maybe Text)

Allows grantee to read the bucket ACL.

poaGrantWrite :: Lens' PutObjectAcl (Maybe Text)

Allows grantee to create, overwrite, and delete any object in the bucket.

poaGrantWriteACP :: Lens' PutObjectAcl (Maybe Text)

Allows grantee to write the ACL for the applicable bucket.

poaKey :: Lens' PutObjectAcl ObjectKey

Response

data PutObjectAclResponse

Response constructor

mkPutObjectAclResponse :: PutObjectAclResponse

Smart constructor for the minimum required parameters to construct a valid PutObjectAclResponse response.

This constructor is provided for convenience and testing purposes.