Network.AWS.S3.V2006_03_01.Types

Description

Amazon Simple Storage Service is storage for the Internet. Amazon S3 has a simple web services interface that you can use to store and retrieve any amount of data, at any time, from anywhere on the web. It gives any developer access to the same highly scalable, reliable, fast, inexpensive data storage infrastructure that Amazon uses to run its own global network of web sites. The service aims to maximize benefits of scale and to pass those benefits on to developers.

Documentation

data S3

Supported version (2006-03-01) of the Amazon Simple Storage Service service.

xmlOptions :: Tagged a XMLOptions

data BucketVersioningStatus

The versioning state of the bucket.

data ExpirationStatus

If Enabled, the rule is currently being applied. If Disabled, the rule is not currently being applied.

data MFADelete

Specifies whether MFA delete is enabled in the bucket versioning configuration. This element is only returned if the bucket has been configured with MFA delete. If the bucket has never been so configured, this element is not returned.

data MFADeleteStatus

Specifies whether MFA delete is enabled in the bucket versioning configuration. This element is only returned if the bucket has been configured with MFA delete. If the bucket has never been so configured, this element is not returned.

data BucketCannedACL

The canned ACL to apply to the bucket.

Constructors

BucketCannedACLAuthenticatedRead
authenticated-read

BucketCannedACLPrivate
private

BucketCannedACLPublicRead
public-read

BucketCannedACLPublicReadWrite
public-read-write

newtype BucketLocationConstraint

Specifies the region where the bucket will be created.

Constructors

BucketLocationConstraint Region

data BucketLogsPermission

Logging permissions assigned to the Grantee for the bucket.

Constructors

BucketLogsPermissionFullControl
FULL_CONTROL

BucketLogsPermissionRead
READ

BucketLogsPermissionWrite
WRITE

data EncodingType

Requests Amazon S3 to encode the object keys in the response and specifies the encoding method to use. An object key may contain any Unicode character; however, XML 1.0 parser cannot parse some characters, such as characters with an ASCII value from 0 to 10. For characters that are not supported in XML 1.0, you can add this parameter to request that Amazon S3 encode the keys in the response.

Constructors

EncodingTypeUrl
url

data Event

Bucket event for which to send notifications.

Constructors

EventS3ReducedRedundancyLostObject
s3:ReducedRedundancyLostObject

data MetadataDirective

Specifies whether the metadata is copied from the source object or replaced with metadata provided in the request.

Constructors

MetadataDirectiveCopy
COPY

MetadataDirectiveReplace
REPLACE

data ObjectCannedACL

The canned ACL to apply to the object.

Constructors

ObjectCannedACLAuthenticatedRead
authenticated-read

ObjectCannedACLBucketOwnerFullControl
bucket-owner-full-control

ObjectCannedACLBucketOwnerRead
bucket-owner-read

ObjectCannedACLPrivate
private

ObjectCannedACLPublicRead
public-read

ObjectCannedACLPublicReadWrite
public-read-write

data ObjectStorageClass

The class of storage used to store the object.

Constructors

ObjectStorageClassGlacier
GLACIER

ObjectStorageClassReducedRedundancy
REDUCED_REDUNDANCY

ObjectStorageClassStandard
STANDARD

data ObjectVersionStorageClass

The class of storage used to store the object.

Constructors

ObjectVersionStorageClassStandard
STANDARD

data Payer

Specifies who pays for the download and request fees.

Constructors

PayerBucketOwner
BucketOwner

PayerRequester
Requester

data Permission

Specifies the permission given to the grantee.

Constructors

PermissionFullControl
FULL_CONTROL

PermissionRead
READ

PermissionReadAcp
READ_ACP

PermissionWrite
WRITE

PermissionWriteAcp
WRITE_ACP

data Protocol

Protocol to use (http, https) when redirecting requests. The default is the protocol that is used in the original request.

Constructors

ProtocolHttp
http

ProtocolHttps
https

data ServerSideEncryption

The Server-side encryption algorithm used when storing this object in S3.

Constructors

ServerSideEncryptionAES256
AES256

data StorageClass

The type of storage to use for the object. Defaults to STANDARD.

Constructors

StorageClassReducedRedundancy
REDUCED_REDUNDANCY

StorageClassStandard
STANDARD

data TransitionStorageClass

The class of storage used to store the object.

Constructors

TransitionStorageClassGlacier
GLACIER

data Type

Type of grantee.

Constructors

TypeAmazonCustomerByEmail
AmazonCustomerByEmail

TypeCanonicalUser
CanonicalUser

TypeGroup
Group

newtype BucketLoggingStatus

Constructors

BucketLoggingStatus

Fields

_blsLoggingEnabled :: Maybe LoggingEnabled

newtype CORSConfiguration

Constructors

CORSConfiguration

Fields

_corscCORSRules :: [CORSRule]

newtype CommonPrefix

Constructors

CommonPrefix

Fields

_cqPrefix :: Maybe Text

newtype CompletedMultipartUpload

Constructors

CompletedMultipartUpload

Fields

_cmuParts :: [CompletedPart]

newtype CreateBucketConfiguration

Constructors

CreateBucketConfiguration

Fields

_cbcLocationConstraint :: Maybe BucketLocationConstraint

Specifies the region where the bucket will be created.

newtype ErrorDocument

Constructors

ErrorDocument

Fields

_edKey :: ObjectKey

The object key name to use when a 4XX class error occurs.

newtype IndexDocument

Constructors

IndexDocument

Fields

_ihSuffix :: Text

A suffix that is appended to a request that is for a directory on the website endpoint (e.g. if the suffix is index.html and you make a request to samplebucketimages the data that is returned will be for the object with the key name images/index.html) The suffix must not be empty and must not include a slash character.

newtype LifecycleConfiguration

Constructors

LifecycleConfiguration

Fields

_lcRules :: [Rule]

newtype NoncurrentVersionExpiration

Specifies when noncurrent object versions expire. Upon expiration, Amazon S3 permanently deletes the noncurrent object versions. You set this lifecycle configuration action on a bucket that has versioning enabled (or suspended) to request that Amazon S3 delete noncurrent object versions at a specific period in the object's lifetime.

Constructors

NoncurrentVersionExpiration

Fields

_nveNoncurrentDays :: Maybe Integer

Specifies the number of days an object is noncurrent before Amazon S3 can perform the associated action. For information about the noncurrent days calculations, see How Amazon S3 Calculates When an Object Became Noncurrent in the Amazon Simple Storage Service Developer Guide.

newtype NotificationConfiguration

Constructors

NotificationConfiguration

Fields

_ncTopicConfiguration :: TopicConfiguration

newtype RequestPaymentConfiguration

Constructors

RequestPaymentConfiguration

Fields

_rpcPayer :: Payer

Specifies who pays for the download and request fees.

newtype RestoreRequest

Constructors

RestoreRequest

Fields

_rzDays :: Integer

Lifetime of the active copy in days.

newtype Tagging

Constructors

Tagging

Fields

_tiTagSet :: [Tag]

data AccessControlPolicy

Constructors

AccessControlPolicy

Fields

_acpGrants :: [Grant]

A list of grants.

_acpOwner :: Maybe Owner

data Bucket

Constructors

Bucket

Fields

_hName :: Maybe BucketName

The name of the bucket.

_hCreationDate :: Maybe RFC822

Date the bucket was created.

data CORSRule

Constructors

CORSRule

Fields

_corssAllowedMethods :: [Text]

Identifies HTTP methods that the domain/origin specified in the rule is allowed to execute.

_corssMaxAgeSeconds :: Maybe Integer

The time in seconds that your browser is to cache the preflight response for the specified resource.

_corssAllowedHeaders :: [Text]

Specifies which headers are allowed in a pre-flight OPTIONS request.

_corssAllowedOrigins :: [Text]

One or more origins you want customers to be able to access the bucket from.

_corssExposeHeaders :: [Text]

One or more headers in the response that you want customers to be able to access from their applications (for example, from a JavaScript XMLHttpRequest object).

data CompletedPart

Constructors

CompletedPart

Fields

_cyETag :: Maybe ETag

Entity tag returned when the part was uploaded.

_cyPartNumber :: Maybe Integer

Part number that identifies the part.

data Condition

A container for describing a condition that must be met for the specified redirect to apply. For example, 1. If request is for pages in the /docs folder, redirect to the /documents folder. 2. If request results in HTTP error 4xx, redirect request to another host where you might process the error.

Constructors

Condition

Fields

_cpKeyPrefixEquals :: Maybe Text

The object key name prefix when the redirect is applied. For example, to redirect requests for ExamplePage.html, the key prefix will be ExamplePage.html. To redirect request for all pages with the prefix docs, the key prefix will be docs, which identifies all objects in the docs/ folder. Required when the parent element Condition is specified and sibling HttpErrorCodeReturnedEquals is not specified. If both conditions are specified, both must be true for the redirect to be applied.

_cpHttpErrorCodeReturnedEquals :: Maybe Text

The HTTP error code when the redirect is applied. In the event of an error, if the error code equals this value, then the specified redirect is applied. Required when parent element Condition is specified and sibling KeyPrefixEquals is not specified. If both are specified, then both must be true for the redirect to be applied.

data CopyObjectResult

Constructors

CopyObjectResult

Fields

_cosETag :: Maybe ETag

_cosLastModified :: Maybe RFC822

data CopyPartResult

Constructors

CopyPartResult

Fields

_cprETag :: Maybe ETag

Entity tag of the object.

_cprLastModified :: Maybe RFC822

Date and time at which the object was uploaded.

data Delete

Constructors

Delete

Fields

_diQuiet :: Maybe Bool

Element to enable quiet mode for the request. When you add this element, you must set its value to true.

_diObjects :: [ObjectIdentifier]

data DeleteMarkerEntry

Constructors

DeleteMarkerEntry

Fields

_dmeVersionId :: Maybe ObjectVersionId

Version ID of an object.

_dmeIsLatest :: Maybe Bool

Specifies whether the object is (true) or is not (false) the latest version of an object.

_dmeOwner :: Maybe Owner

_dmeKey :: Maybe ObjectKey

The object key.

_dmeLastModified :: Maybe RFC822

Date and time the object was last modified.

data DeletedObject

Constructors

DeletedObject

Fields

_ddyVersionId :: Maybe ObjectVersionId

_ddyDeleteMarker :: Maybe Bool

_ddyDeleteMarkerVersionId :: Maybe Text

_ddyKey :: Maybe ObjectKey

data Error

Constructors

Error

Fields

_erVersionId :: Maybe ObjectVersionId

_erKey :: Maybe ObjectKey

_erCode :: Maybe Text

_erMessage :: Maybe Text

data Grant

Constructors

Grant

Fields

_mPermission :: Maybe Permission

Specifies the permission given to the grantee.

_mGrantee :: Maybe Grantee

data Grantee

Constructors

Grantee

Fields

_gURI :: Maybe Text

URI of the grantee group.

_gEmailAddress :: Maybe Text

Email address of the grantee.

_gDisplayName :: Maybe Text

Screen name of the grantee.

_gID :: Maybe Text

The canonical user ID of the grantee.

_gType :: Type

Type of grantee.

data Initiator

Identifies who initiated the multipart upload.

Constructors

Initiator

Fields

_irDisplayName :: Maybe Text

Name of the Principal.

_irID :: Maybe Text

If the principal is an AWS account, it provides the Canonical User ID. If the principal is an IAM User, it provides a user ARN value.

data LifecycleExpiration

Constructors

LifecycleExpiration

Fields

_lfDays :: Maybe Integer

Indicates the lifetime, in days, of the objects that are subject to the rule. The value must be a non-zero positive integer.

_lfDate :: Maybe RFC822

Indicates at what date the object is to be moved or deleted. Should be in GMT ISO 8601 Format.

data LoggingEnabled

Constructors

LoggingEnabled

Fields

_leTargetBucket :: Maybe Text

Specifies the bucket where you want Amazon S3 to store server access logs. You can have your logs delivered to any bucket that you own, including the same bucket that is being logged. You can also configure multiple buckets to deliver their logs to the same target bucket. In this case you should choose a different TargetPrefix for each source bucket so that the delivered log files can be distinguished by key.

_leTargetGrants :: [TargetGrant]

_leTargetPrefix :: Maybe Text

This element lets you specify a prefix for the keys that the log files will be stored under.

data MultipartUpload

Constructors

MultipartUpload

Fields

_mwInitiated :: Maybe RFC822

Date and time at which the multipart upload was initiated.

_mwInitiator :: Maybe Initiator

Identifies who initiated the multipart upload.

_mwOwner :: Maybe Owner

_mwKey :: Maybe ObjectKey

Key of the object for which the multipart upload was initiated.

_mwStorageClass :: Maybe StorageClass

The class of storage used to store the object.

_mwUploadId :: Maybe Text

Upload ID that identifies the multipart upload.

data NoncurrentVersionTransition

Container for the transition rule that describes when noncurrent objects transition to the GLACIER storage class. If your bucket is versioning-enabled (or versioning is suspended), you can set this action to request that Amazon S3 transition noncurrent object versions to the GLACIER storage class at a specific period in the object's lifetime.

Constructors

NoncurrentVersionTransition

Fields

_nvtStorageClass :: Maybe TransitionStorageClass

The class of storage used to store the object.

_nvtNoncurrentDays :: Maybe Integer

Specifies the number of days an object is noncurrent before Amazon S3 can perform the associated action. For information about the noncurrent days calculations, see How Amazon S3 Calculates When an Object Became Noncurrent in the Amazon Simple Storage Service Developer Guide.

data Object

Constructors

Object

Fields

_oouETag :: ETag

_oouSize :: Integer

_oouOwner :: Owner

_oouKey :: ObjectKey

_oouStorageClass :: ObjectStorageClass

The class of storage used to store the object.

_oouLastModified :: RFC822

data ObjectIdentifier

Constructors

ObjectIdentifier

Fields

_oiVersionId :: Maybe ObjectVersionId

VersionId for the specific version of the object to delete.

_oiKey :: ObjectKey

Key name of the object to delete.

data ObjectVersion

Constructors

ObjectVersion

Fields

_ovETag :: Maybe ETag

_ovVersionId :: Maybe ObjectVersionId

Version ID of an object.

_ovSize :: Maybe Integer

Size in bytes of the object.

_ovIsLatest :: Maybe Bool

Specifies whether the object is (true) or is not (false) the latest version of an object.

_ovOwner :: Maybe Owner

_ovKey :: Maybe ObjectKey

The object key.

_ovStorageClass :: Maybe ObjectVersionStorageClass

The class of storage used to store the object.

_ovLastModified :: Maybe RFC822

Date and time the object was last modified.

data Owner

Constructors

Owner

Fields

_oDisplayName :: Maybe Text

_oID :: Maybe Text

data Part

Constructors

Part

Fields

_pvETag :: Maybe ETag

Entity tag returned when the part was uploaded.

_pvSize :: Maybe Integer

Size of the uploaded part data.

_pvPartNumber :: Maybe Integer

Part number identifying the part.

_pvLastModified :: Maybe RFC822

Date and time at which the part was uploaded.

data Redirect

Container for redirect information. You can redirect requests to another host, to another page, or with another protocol. In the event of an error, you can can specify a different error code to return.

Constructors

Redirect

Fields

_rtHostName :: Maybe Text

The host name to use in the redirect request.

_rtProtocol :: Maybe Protocol

Protocol to use (http, https) when redirecting requests. The default is the protocol that is used in the original request.

_rtHttpRedirectCode :: Maybe Text

The HTTP redirect code to use on the response. Not required if one of the siblings is present.

_rtReplaceKeyWith :: Maybe Text

The specific object key to use in the redirect request. For example, redirect request to error.html. Not required if one of the sibling is present. Can be present only if ReplaceKeyPrefixWith is not provided.

_rtReplaceKeyPrefixWith :: Maybe Text

The object key prefix to use in the redirect request. For example, to redirect requests for all pages with prefix docs/ (objects in the docs folder) to documents, you can set a condition block with KeyPrefixEquals set to docs/ and in the Redirect set ReplaceKeyPrefixWith to /documents. Not required if one of the siblings is present. Can be present only if ReplaceKeyWith is not provided.

data RedirectAllRequestsTo

Constructors

RedirectAllRequestsTo

Fields

_rartHostName :: Text

Name of the host where requests will be redirected.

_rartProtocol :: Maybe Protocol

Protocol to use (http, https) when redirecting requests. The default is the protocol that is used in the original request.

data RoutingRule

Constructors

RoutingRule

Fields

_rsRedirect :: Redirect

Container for redirect information. You can redirect requests to another host, to another page, or with another protocol. In the event of an error, you can can specify a different error code to return.

_rsCondition :: Maybe Condition

A container for describing a condition that must be met for the specified redirect to apply. For example, 1. If request is for pages in the docs folder, redirect to the documents folder. 2. If request results in HTTP error 4xx, redirect request to another host where you might process the error.

data Rule

Constructors

Rule

Fields

_rhStatus :: Switch ExpirationStatus

If Enabled, the rule is currently being applied. If Disabled, the rule is not currently being applied.

_rhNoncurrentVersionExpiration :: Maybe NoncurrentVersionExpiration

Specifies when noncurrent object versions expire. Upon expiration, Amazon S3 permanently deletes the noncurrent object versions. You set this lifecycle configuration action on a bucket that has versioning enabled (or suspended) to request that Amazon S3 delete noncurrent object versions at a specific period in the object's lifetime.

_rhTransition :: Maybe Transition

_rhPrefix :: Text

Prefix identifying one or more objects to which the rule applies.

_rhExpiration :: Maybe LifecycleExpiration

_rhNoncurrentVersionTransition :: Maybe NoncurrentVersionTransition

Container for the transition rule that describes when noncurrent objects transition to the GLACIER storage class. If your bucket is versioning-enabled (or versioning is suspended), you can set this action to request that Amazon S3 transition noncurrent object versions to the GLACIER storage class at a specific period in the object's lifetime.

_rhID :: Maybe Text

Unique identifier for the rule. The value cannot be longer than 255 characters.

data Tag

Constructors

Tag

Fields

_tjValue :: Text

Value of the tag.

_tjKey :: ObjectKey

Name of the tag.

data TargetGrant

Constructors

TargetGrant

Fields

_thPermission :: Maybe BucketLogsPermission

Logging permissions assigned to the Grantee for the bucket.

_thGrantee :: Maybe Grantee

data TopicConfiguration

Constructors

TopicConfiguration

Fields

_tcEvent :: Maybe Event

Bucket event for which to send notifications.

_tcTopic :: Maybe Text

Amazon SNS topic to which Amazon S3 will publish a message to report the specified events for the bucket.

data Transition

Constructors

Transition

Fields

_tnDays :: Maybe Integer

Indicates the lifetime, in days, of the objects that are subject to the rule. The value must be a non-zero positive integer.

_tnDate :: Maybe RFC822

Indicates at what date the object is to be moved or deleted. Should be in GMT ISO 8601 Format.

_tnStorageClass :: Maybe TransitionStorageClass

The class of storage used to store the object.

data VersioningConfiguration

Constructors

VersioningConfiguration

Fields

_vcStatus :: Maybe (Switch BucketVersioningStatus)

The versioning state of the bucket.

_vcMfaDelete :: Maybe (Switch MFADelete)

Specifies whether MFA delete is enabled in the bucket versioning configuration. This element is only returned if the bucket has been configured with MFA delete. If the bucket has never been so configured, this element is not returned.

data WebsiteConfiguration

Constructors

WebsiteConfiguration

Fields

_wcRedirectAllRequestsTo :: Maybe RedirectAllRequestsTo

_wcErrorDocument :: Maybe ErrorDocument

_wcIndexDocument :: Maybe IndexDocument

_wcRoutingRules :: [RoutingRule]

blsLoggingEnabled :: Iso' BucketLoggingStatus (Maybe LoggingEnabled)

corscCORSRules :: Iso' CORSConfiguration [CORSRule]

cqPrefix :: Iso' CommonPrefix (Maybe Text)

cmuParts :: Iso' CompletedMultipartUpload [CompletedPart]

cbcLocationConstraint :: Iso' CreateBucketConfiguration (Maybe BucketLocationConstraint)

edKey :: Iso' ErrorDocument ObjectKey

ihSuffix :: Iso' IndexDocument Text

lcRules :: Iso' LifecycleConfiguration [Rule]

nveNoncurrentDays :: Iso' NoncurrentVersionExpiration (Maybe Integer)

ncTopicConfiguration :: Iso' NotificationConfiguration TopicConfiguration

rpcPayer :: Iso' RequestPaymentConfiguration Payer

rzDays :: Iso' RestoreRequest Integer

tiTagSet :: Iso' Tagging [Tag]

acpOwner :: Lens' AccessControlPolicy (Maybe Owner)

acpGrants :: Lens' AccessControlPolicy [Grant]

hName :: Lens' Bucket (Maybe BucketName)

hCreationDate :: Lens' Bucket (Maybe RFC822)

corssMaxAgeSeconds :: Lens' CORSRule (Maybe Integer)

corssExposeHeaders :: Lens' CORSRule [Text]

corssAllowedOrigins :: Lens' CORSRule [Text]

corssAllowedMethods :: Lens' CORSRule [Text]

corssAllowedHeaders :: Lens' CORSRule [Text]

cyPartNumber :: Lens' CompletedPart (Maybe Integer)

cyETag :: Lens' CompletedPart (Maybe ETag)

cpKeyPrefixEquals :: Lens' Condition (Maybe Text)

cpHttpErrorCodeReturnedEquals :: Lens' Condition (Maybe Text)

cosLastModified :: Lens' CopyObjectResult (Maybe RFC822)

cosETag :: Lens' CopyObjectResult (Maybe ETag)

cprLastModified :: Lens' CopyPartResult (Maybe RFC822)

cprETag :: Lens' CopyPartResult (Maybe ETag)

diQuiet :: Lens' Delete (Maybe Bool)

diObjects :: Lens' Delete [ObjectIdentifier]

dmeVersionId :: Lens' DeleteMarkerEntry (Maybe ObjectVersionId)

dmeOwner :: Lens' DeleteMarkerEntry (Maybe Owner)

dmeLastModified :: Lens' DeleteMarkerEntry (Maybe RFC822)

dmeKey :: Lens' DeleteMarkerEntry (Maybe ObjectKey)

dmeIsLatest :: Lens' DeleteMarkerEntry (Maybe Bool)

ddyVersionId :: Lens' DeletedObject (Maybe ObjectVersionId)

ddyKey :: Lens' DeletedObject (Maybe ObjectKey)

ddyDeleteMarkerVersionId :: Lens' DeletedObject (Maybe Text)

ddyDeleteMarker :: Lens' DeletedObject (Maybe Bool)

erVersionId :: Lens' Error (Maybe ObjectVersionId)

erMessage :: Lens' Error (Maybe Text)

erKey :: Lens' Error (Maybe ObjectKey)

erCode :: Lens' Error (Maybe Text)

mPermission :: Lens' Grant (Maybe Permission)

mGrantee :: Lens' Grant (Maybe Grantee)

gURI :: Lens' Grantee (Maybe Text)

gType :: Lens' Grantee Type

gID :: Lens' Grantee (Maybe Text)

gEmailAddress :: Lens' Grantee (Maybe Text)

gDisplayName :: Lens' Grantee (Maybe Text)

irID :: Lens' Initiator (Maybe Text)

irDisplayName :: Lens' Initiator (Maybe Text)

lfDays :: Lens' LifecycleExpiration (Maybe Integer)

lfDate :: Lens' LifecycleExpiration (Maybe RFC822)

leTargetPrefix :: Lens' LoggingEnabled (Maybe Text)

leTargetGrants :: Lens' LoggingEnabled [TargetGrant]

leTargetBucket :: Lens' LoggingEnabled (Maybe Text)

mwUploadId :: Lens' MultipartUpload (Maybe Text)

mwStorageClass :: Lens' MultipartUpload (Maybe StorageClass)

mwOwner :: Lens' MultipartUpload (Maybe Owner)

mwKey :: Lens' MultipartUpload (Maybe ObjectKey)

mwInitiator :: Lens' MultipartUpload (Maybe Initiator)

mwInitiated :: Lens' MultipartUpload (Maybe RFC822)

nvtStorageClass :: Lens' NoncurrentVersionTransition (Maybe TransitionStorageClass)

nvtNoncurrentDays :: Lens' NoncurrentVersionTransition (Maybe Integer)

oouStorageClass :: Lens' Object ObjectStorageClass

oouSize :: Lens' Object Integer

oouOwner :: Lens' Object Owner

oouLastModified :: Lens' Object RFC822

oouKey :: Lens' Object ObjectKey

oouETag :: Lens' Object ETag

oiVersionId :: Lens' ObjectIdentifier (Maybe ObjectVersionId)

oiKey :: Lens' ObjectIdentifier ObjectKey

ovVersionId :: Lens' ObjectVersion (Maybe ObjectVersionId)

ovStorageClass :: Lens' ObjectVersion (Maybe ObjectVersionStorageClass)

ovSize :: Lens' ObjectVersion (Maybe Integer)

ovOwner :: Lens' ObjectVersion (Maybe Owner)

ovLastModified :: Lens' ObjectVersion (Maybe RFC822)

ovKey :: Lens' ObjectVersion (Maybe ObjectKey)

ovIsLatest :: Lens' ObjectVersion (Maybe Bool)

ovETag :: Lens' ObjectVersion (Maybe ETag)

oID :: Lens' Owner (Maybe Text)

oDisplayName :: Lens' Owner (Maybe Text)

pvSize :: Lens' Part (Maybe Integer)

pvPartNumber :: Lens' Part (Maybe Integer)

pvLastModified :: Lens' Part (Maybe RFC822)

pvETag :: Lens' Part (Maybe ETag)

rtReplaceKeyWith :: Lens' Redirect (Maybe Text)

rtReplaceKeyPrefixWith :: Lens' Redirect (Maybe Text)

rtProtocol :: Lens' Redirect (Maybe Protocol)

rtHttpRedirectCode :: Lens' Redirect (Maybe Text)

rtHostName :: Lens' Redirect (Maybe Text)

rartProtocol :: Lens' RedirectAllRequestsTo (Maybe Protocol)

rartHostName :: Lens' RedirectAllRequestsTo Text

rsRedirect :: Lens' RoutingRule Redirect

rsCondition :: Lens' RoutingRule (Maybe Condition)

rhTransition :: Lens' Rule (Maybe Transition)

rhStatus :: Lens' Rule (Switch ExpirationStatus)

rhPrefix :: Lens' Rule Text

rhNoncurrentVersionTransition :: Lens' Rule (Maybe NoncurrentVersionTransition)

rhNoncurrentVersionExpiration :: Lens' Rule (Maybe NoncurrentVersionExpiration)

rhID :: Lens' Rule (Maybe Text)

rhExpiration :: Lens' Rule (Maybe LifecycleExpiration)

tjValue :: Lens' Tag Text

tjKey :: Lens' Tag ObjectKey

thPermission :: Lens' TargetGrant (Maybe BucketLogsPermission)

thGrantee :: Lens' TargetGrant (Maybe Grantee)

tcTopic :: Lens' TopicConfiguration (Maybe Text)

tcEvent :: Lens' TopicConfiguration (Maybe Event)

tnStorageClass :: Lens' Transition (Maybe TransitionStorageClass)

tnDays :: Lens' Transition (Maybe Integer)

tnDate :: Lens' Transition (Maybe RFC822)

vcStatus :: Lens' VersioningConfiguration (Maybe (Switch BucketVersioningStatus))

vcMfaDelete :: Lens' VersioningConfiguration (Maybe (Switch MFADelete))

wcRoutingRules :: Lens' WebsiteConfiguration [RoutingRule]

wcRedirectAllRequestsTo :: Lens' WebsiteConfiguration (Maybe RedirectAllRequestsTo)

wcIndexDocument :: Lens' WebsiteConfiguration (Maybe IndexDocument)

wcErrorDocument :: Lens' WebsiteConfiguration (Maybe ErrorDocument)