Network.AWS.S3.V2006_03_01.UploadPartCopy

Description

Uploads a part by copying data from an existing object as data source.

Documentation

uploadPartCopy

Arguments

:: Text
_upcrCopySource

-> BucketName
_upcrBucket

-> Text
_upcrUploadId

-> ObjectKey
_upcrKey

-> Integer
_upcrPartNumber

-> UploadPartCopy

Minimum specification for a UploadPartCopy request.

data UploadPartCopy

Constructors

UploadPartCopy

Fields

_upcrCopySource :: Text

The name of the source bucket and key name of the source object, separated by a slash (/). Must be URL-encoded.

_upcrBucket :: BucketName

_upcrUploadId :: Text

Upload ID identifying the multipart upload whose part is being copied.

_upcrKey :: ObjectKey

_upcrPartNumber :: Integer

Part number of part being copied.

_upcrCopySourceIfMatch :: Maybe Text

Copies the object if its entity tag (ETag) matches the specified tag.

_upcrCopySourceIfModifiedSince :: Maybe RFC822

Copies the object if it has been modified since the specified time.

_upcrCopySourceIfNoneMatch :: Maybe Text

Copies the object if its entity tag (ETag) is different than the specified ETag.

_upcrCopySourceIfUnmodifiedSince :: Maybe RFC822

Copies the object if it hasn't been modified since the specified time.

_upcrCopySourceRange :: Maybe Text

The range of bytes to copy from the source object. The range value must use the form bytes=first-last, where the first and last are the zero-based byte offsets to copy. For example, bytes=0-9 indicates that you want to copy the first ten bytes of the source. You can copy a range only if the source object is greater than 5 GB.

_upcrCopySourceSSECustomerAlgorithm :: Maybe Text

Specifies the algorithm to use when decrypting the source object (e.g., AES256).

_upcrCopySourceSSECustomerKey :: Maybe Text

Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

_upcrCopySourceSSECustomerKeyMD5 :: Maybe Text

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

_upcrSSECustomerAlgorithm :: Maybe Text

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

_upcrSSECustomerKey :: Maybe Text

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

_upcrSSECustomerKeyMD5 :: Maybe Text

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

upcrUploadId :: Lens' UploadPartCopy Text

upcrSSECustomerKeyMD5 :: Lens' UploadPartCopy (Maybe Text)

upcrSSECustomerKey :: Lens' UploadPartCopy (Maybe Text)

upcrSSECustomerAlgorithm :: Lens' UploadPartCopy (Maybe Text)

upcrPartNumber :: Lens' UploadPartCopy Integer

upcrKey :: Lens' UploadPartCopy ObjectKey

upcrCopySourceSSECustomerKeyMD5 :: Lens' UploadPartCopy (Maybe Text)

upcrCopySourceSSECustomerKey :: Lens' UploadPartCopy (Maybe Text)

upcrCopySourceSSECustomerAlgorithm :: Lens' UploadPartCopy (Maybe Text)

upcrCopySourceRange :: Lens' UploadPartCopy (Maybe Text)

upcrCopySourceIfUnmodifiedSince :: Lens' UploadPartCopy (Maybe RFC822)

upcrCopySourceIfNoneMatch :: Lens' UploadPartCopy (Maybe Text)

upcrCopySourceIfModifiedSince :: Lens' UploadPartCopy (Maybe RFC822)

upcrCopySourceIfMatch :: Lens' UploadPartCopy (Maybe Text)

upcrCopySource :: Lens' UploadPartCopy Text

upcrBucket :: Lens' UploadPartCopy BucketName

data UploadPartCopyResponse

Constructors

UploadPartCopyResponse

Fields

_upcoCopyPartResult :: Maybe CopyPartResult

_upcoCopySourceVersionId :: Maybe Text

The version of the source object that was copied, if you have enabled versioning on the source bucket.

_upcoSSECustomerAlgorithm :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

_upcoSSECustomerKeyMD5 :: Maybe Text

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

_upcoServerSideEncryption :: Maybe ServerSideEncryption

The Server-side encryption algorithm used when storing this object in S3.

upcoServerSideEncryption :: Lens' UploadPartCopyResponse (Maybe ServerSideEncryption)

upcoSSECustomerKeyMD5 :: Lens' UploadPartCopyResponse (Maybe Text)

upcoSSECustomerAlgorithm :: Lens' UploadPartCopyResponse (Maybe Text)

upcoCopySourceVersionId :: Lens' UploadPartCopyResponse (Maybe Text)

upcoCopyPartResult :: Lens' UploadPartCopyResponse (Maybe CopyPartResult)