Network.AWS.S3.AbortMultipartUpload

Description

Aborts a multipart upload.

To verify that all parts have been removed, so you don't get charged for the part storage, you should call the List Parts operation and ensure the parts list is empty.

Creating a Request

abortMultipartUpload

Arguments

:: BucketName
amuBucket

-> ObjectKey
amuKey

-> Text
amuUploadId

-> AbortMultipartUpload

Creates a value of AbortMultipartUpload with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

amuRequestPayer - Undocumented member.

amuBucket - Undocumented member.

amuKey - Undocumented member.

amuUploadId - Undocumented member.

data AbortMultipartUpload

See: abortMultipartUpload smart constructor.

Request Lenses

amuRequestPayer :: Lens' AbortMultipartUpload (Maybe RequestPayer)

Undocumented member.

amuBucket :: Lens' AbortMultipartUpload BucketName

Undocumented member.

amuKey :: Lens' AbortMultipartUpload ObjectKey

Undocumented member.

amuUploadId :: Lens' AbortMultipartUpload Text

Undocumented member.

Destructuring the Response

abortMultipartUploadResponse

Arguments

:: Int
amursResponseStatus

-> AbortMultipartUploadResponse

Creates a value of AbortMultipartUploadResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

amursRequestCharged - Undocumented member.

amursResponseStatus - -- | The response status code.

data AbortMultipartUploadResponse

See: abortMultipartUploadResponse smart constructor.

Response Lenses

amursRequestCharged :: Lens' AbortMultipartUploadResponse (Maybe RequestCharged)

Undocumented member.

amursResponseStatus :: Lens' AbortMultipartUploadResponse Int

- | The response status code.