Network.AWS.S3.CompleteMultipartUpload

Description

Completes a multipart upload by assembling previously uploaded parts.

Creating a Request

completeMultipartUpload

Arguments

:: BucketName
cBucket

-> ObjectKey
cKey

-> Text
cUploadId

-> CompleteMultipartUpload

Creates a value of CompleteMultipartUpload with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cRequestPayer - Undocumented member.

cMultipartUpload - Undocumented member.

cBucket - Undocumented member.

cKey - Undocumented member.

cUploadId - Undocumented member.

data CompleteMultipartUpload

See: completeMultipartUpload smart constructor.

Request Lenses

cRequestPayer :: Lens' CompleteMultipartUpload (Maybe RequestPayer)

Undocumented member.

cMultipartUpload :: Lens' CompleteMultipartUpload (Maybe CompletedMultipartUpload)

Undocumented member.

cBucket :: Lens' CompleteMultipartUpload BucketName

Undocumented member.

cKey :: Lens' CompleteMultipartUpload ObjectKey

Undocumented member.

cUploadId :: Lens' CompleteMultipartUpload Text

Undocumented member.

Destructuring the Response

completeMultipartUploadResponse

Arguments

:: Int
crsResponseStatus

-> CompleteMultipartUploadResponse

Creates a value of CompleteMultipartUploadResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

crsRequestCharged - Undocumented member.

crsETag - Entity tag of the object.

crsVersionId - Version of the object.

crsLocation - Undocumented member.

crsExpiration - If the object expiration is configured, this will contain the expiration date (expiry-date) and rule ID (rule-id). The value of rule-id is URL encoded.

crsBucket - Undocumented member.

crsKey - Undocumented member.

crsSSEKMSKeyId - If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

crsServerSideEncryption - The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

crsResponseStatus - -- | The response status code.

data CompleteMultipartUploadResponse

See: completeMultipartUploadResponse smart constructor.

Response Lenses

crsRequestCharged :: Lens' CompleteMultipartUploadResponse (Maybe RequestCharged)

Undocumented member.

crsETag :: Lens' CompleteMultipartUploadResponse (Maybe ETag)

Entity tag of the object.

crsVersionId :: Lens' CompleteMultipartUploadResponse (Maybe ObjectVersionId)

Version of the object.

crsLocation :: Lens' CompleteMultipartUploadResponse (Maybe Text)

Undocumented member.

crsExpiration :: Lens' CompleteMultipartUploadResponse (Maybe Text)

If the object expiration is configured, this will contain the expiration date (expiry-date) and rule ID (rule-id). The value of rule-id is URL encoded.

crsBucket :: Lens' CompleteMultipartUploadResponse (Maybe BucketName)

Undocumented member.

crsKey :: Lens' CompleteMultipartUploadResponse (Maybe ObjectKey)

Undocumented member.

crsSSEKMSKeyId :: Lens' CompleteMultipartUploadResponse (Maybe Text)

If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

crsServerSideEncryption :: Lens' CompleteMultipartUploadResponse (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

crsResponseStatus :: Lens' CompleteMultipartUploadResponse Int

- | The response status code.