Network.AWS.S3.CompleteMultipartUpload

Description

Completes a multipart upload by assembling previously uploaded parts.

http://docs.aws.amazon.com/AmazonS3/latest/API/CompleteMultipartUpload.html

Request

data CompleteMultipartUpload

Request constructor

completeMultipartUpload

Arguments

:: Text
cmu1Bucket

-> Text
cmu1Key

-> Text
cmu1UploadId

-> CompleteMultipartUpload

CompleteMultipartUpload constructor.

The fields accessible through corresponding lenses are:

cmu1Bucket :: Text

cmu1Key :: Text

cmu1MultipartUpload :: Maybe CompletedMultipartUpload

cmu1UploadId :: Text

Request lenses

cmu1Bucket :: Lens' CompleteMultipartUpload Text

cmu1Key :: Lens' CompleteMultipartUpload Text

cmu1MultipartUpload :: Lens' CompleteMultipartUpload (Maybe CompletedMultipartUpload)

cmu1UploadId :: Lens' CompleteMultipartUpload Text

Response

data CompleteMultipartUploadResponse

Response constructor

completeMultipartUploadResponse :: CompleteMultipartUploadResponse

CompleteMultipartUploadResponse constructor.

The fields accessible through corresponding lenses are:

cmur1Bucket :: Maybe Text

cmur1ETag :: Maybe Text

cmur1Expiration :: Maybe UTCTime

cmur1Key :: Maybe Text

cmur1Location :: Maybe Text

cmur1SSEKMSKeyId :: Maybe Text

cmur1ServerSideEncryption :: Maybe Text

cmur1VersionId :: Maybe Text

Response lenses

cmur1Bucket :: Lens' CompleteMultipartUploadResponse (Maybe Text)

cmur1ETag :: Lens' CompleteMultipartUploadResponse (Maybe Text)

Entity tag of the object.

cmur1Expiration :: Lens' CompleteMultipartUploadResponse (Maybe UTCTime)

If the object expiration is configured, this will contain the expiration date (expiry-date) and rule ID (rule-id). The value of rule-id is URL encoded.

cmur1Key :: Lens' CompleteMultipartUploadResponse (Maybe Text)

cmur1Location :: Lens' CompleteMultipartUploadResponse (Maybe Text)

cmur1SSEKMSKeyId :: Lens' CompleteMultipartUploadResponse (Maybe Text)

If present, specifies the AWS KMS key used to encrypt the object.

cmur1ServerSideEncryption :: Lens' CompleteMultipartUploadResponse (Maybe Text)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

cmur1VersionId :: Lens' CompleteMultipartUploadResponse (Maybe Text)

Version of the object.