Network.AWS.S3.CopyObject

Description

Creates a copy of an object that is already stored in Amazon S3.

Creating a Request

copyObject

Arguments

:: BucketName
coBucket

-> Text
coCopySource

-> ObjectKey
coKey

-> CopyObject

Creates a value of CopyObject with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

coCopySourceIfModifiedSince - Copies the object if it has been modified since the specified time.

coCopySourceIfUnmodifiedSince - Copies the object if it hasn't been modified since the specified time.

coCopySourceSSECustomerKeyMD5 - Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

coTaggingDirective - Specifies whether the object tag-set are copied from the source object or replaced with tag-set provided in the request.

coMetadataDirective - Specifies whether the metadata is copied from the source object or replaced with metadata provided in the request.

coExpires - The date and time at which the object is no longer cacheable.

coGrantReadACP - Allows grantee to read the object ACL.

coCopySourceIfNoneMatch - Copies the object if its entity tag (ETag) is different than the specified ETag.

coSSECustomerAlgorithm - Specifies the algorithm to use to when encrypting the object (e.g., AES256).

coSSECustomerKey - Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

coRequestPayer - Undocumented member.

coGrantWriteACP - Allows grantee to write the ACL for the applicable object.

coCopySourceIfMatch - Copies the object if its entity tag (ETag) matches the specified tag.

coWebsiteRedirectLocation - If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

coGrantRead - Allows grantee to read the object data and its metadata.

coStorageClass - The type of storage to use for the object. Defaults to STANDARD.

coSSECustomerKeyMD5 - Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

coSSEKMSKeyId - Specifies the AWS KMS key ID to use for object encryption. All GET and PUT requests for an object protected by AWS KMS will fail if not made via SSL or using SigV4. Documentation on configuring any of the officially supported AWS SDKs and CLI can be found at http://docs.aws.amazon.com/AmazonS3/latest/dev/UsingAWSSDK.html#specify-signature-version

coGrantFullControl - Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

coContentEncoding - Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

coTagging - The tag-set for the object destination object this value must be used in conjunction with the TaggingDirective. The tag-set must be encoded as URL Query parameters

coMetadata - A map of metadata to store with the object in S3.

coCacheControl - Specifies caching behavior along the request/reply chain.

coContentLanguage - The language the content is in.

coCopySourceSSECustomerKey - Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

coCopySourceSSECustomerAlgorithm - Specifies the algorithm to use when decrypting the source object (e.g., AES256).

coACL - The canned ACL to apply to the object.

coContentDisposition - Specifies presentational information for the object.

coServerSideEncryption - The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

coContentType - A standard MIME type describing the format of the object data.

coBucket - Undocumented member.

coCopySource - The name of the source bucket and key name of the source object, separated by a slash (/). Must be URL-encoded.

coKey - Undocumented member.

data CopyObject

See: copyObject smart constructor.

Request Lenses

coCopySourceIfModifiedSince :: Lens' CopyObject (Maybe UTCTime)

Copies the object if it has been modified since the specified time.

coCopySourceIfUnmodifiedSince :: Lens' CopyObject (Maybe UTCTime)

Copies the object if it hasn't been modified since the specified time.

coCopySourceSSECustomerKeyMD5 :: Lens' CopyObject (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

coTaggingDirective :: Lens' CopyObject (Maybe TaggingDirective)

Specifies whether the object tag-set are copied from the source object or replaced with tag-set provided in the request.

coMetadataDirective :: Lens' CopyObject (Maybe MetadataDirective)

Specifies whether the metadata is copied from the source object or replaced with metadata provided in the request.

coExpires :: Lens' CopyObject (Maybe UTCTime)

The date and time at which the object is no longer cacheable.

coGrantReadACP :: Lens' CopyObject (Maybe Text)

Allows grantee to read the object ACL.

coCopySourceIfNoneMatch :: Lens' CopyObject (Maybe Text)

Copies the object if its entity tag (ETag) is different than the specified ETag.

coSSECustomerAlgorithm :: Lens' CopyObject (Maybe Text)

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

coSSECustomerKey :: Lens' CopyObject (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

coRequestPayer :: Lens' CopyObject (Maybe RequestPayer)

Undocumented member.

coGrantWriteACP :: Lens' CopyObject (Maybe Text)

Allows grantee to write the ACL for the applicable object.

coCopySourceIfMatch :: Lens' CopyObject (Maybe Text)

Copies the object if its entity tag (ETag) matches the specified tag.

coWebsiteRedirectLocation :: Lens' CopyObject (Maybe Text)

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

coGrantRead :: Lens' CopyObject (Maybe Text)

Allows grantee to read the object data and its metadata.

coStorageClass :: Lens' CopyObject (Maybe StorageClass)

The type of storage to use for the object. Defaults to STANDARD.

coSSECustomerKeyMD5 :: Lens' CopyObject (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

coSSEKMSKeyId :: Lens' CopyObject (Maybe Text)

Specifies the AWS KMS key ID to use for object encryption. All GET and PUT requests for an object protected by AWS KMS will fail if not made via SSL or using SigV4. Documentation on configuring any of the officially supported AWS SDKs and CLI can be found at http://docs.aws.amazon.com/AmazonS3/latest/dev/UsingAWSSDK.html#specify-signature-version

coGrantFullControl :: Lens' CopyObject (Maybe Text)

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

coContentEncoding :: Lens' CopyObject (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

coTagging :: Lens' CopyObject (Maybe Text)

The tag-set for the object destination object this value must be used in conjunction with the TaggingDirective. The tag-set must be encoded as URL Query parameters

coMetadata :: Lens' CopyObject (HashMap Text Text)

A map of metadata to store with the object in S3.

coCacheControl :: Lens' CopyObject (Maybe Text)

Specifies caching behavior along the request/reply chain.

coContentLanguage :: Lens' CopyObject (Maybe Text)

The language the content is in.

coCopySourceSSECustomerKey :: Lens' CopyObject (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

coCopySourceSSECustomerAlgorithm :: Lens' CopyObject (Maybe Text)

Specifies the algorithm to use when decrypting the source object (e.g., AES256).

coACL :: Lens' CopyObject (Maybe ObjectCannedACL)

The canned ACL to apply to the object.

coContentDisposition :: Lens' CopyObject (Maybe Text)

Specifies presentational information for the object.

coServerSideEncryption :: Lens' CopyObject (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

coContentType :: Lens' CopyObject (Maybe Text)

A standard MIME type describing the format of the object data.

coBucket :: Lens' CopyObject BucketName

Undocumented member.

coCopySource :: Lens' CopyObject Text

The name of the source bucket and key name of the source object, separated by a slash (/). Must be URL-encoded.

coKey :: Lens' CopyObject ObjectKey

Undocumented member.

Destructuring the Response

copyObjectResponse

Arguments

:: Int
corsResponseStatus

-> CopyObjectResponse

Creates a value of CopyObjectResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

corsRequestCharged - Undocumented member.

corsVersionId - Version ID of the newly created copy.

corsExpiration - If the object expiration is configured, the response includes this header.

corsSSECustomerAlgorithm - If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

corsCopySourceVersionId - Undocumented member.

corsSSECustomerKeyMD5 - If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

corsSSEKMSKeyId - If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

corsServerSideEncryption - The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

corsCopyObjectResult - Undocumented member.

corsResponseStatus - -- | The response status code.

data CopyObjectResponse

See: copyObjectResponse smart constructor.

Response Lenses

corsRequestCharged :: Lens' CopyObjectResponse (Maybe RequestCharged)

Undocumented member.

corsVersionId :: Lens' CopyObjectResponse (Maybe ObjectVersionId)

Version ID of the newly created copy.

corsExpiration :: Lens' CopyObjectResponse (Maybe Text)

If the object expiration is configured, the response includes this header.

corsSSECustomerAlgorithm :: Lens' CopyObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

corsCopySourceVersionId :: Lens' CopyObjectResponse (Maybe Text)

Undocumented member.

corsSSECustomerKeyMD5 :: Lens' CopyObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

corsSSEKMSKeyId :: Lens' CopyObjectResponse (Maybe Text)

If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

corsServerSideEncryption :: Lens' CopyObjectResponse (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

corsCopyObjectResult :: Lens' CopyObjectResponse (Maybe CopyObjectResult)

Undocumented member.

corsResponseStatus :: Lens' CopyObjectResponse Int

- | The response status code.