Network.AWS.S3.CopyObject

Description

Creates a copy of an object that is already stored in Amazon S3.

http://docs.aws.amazon.com/AmazonS3/latest/API/CopyObject.html

Request

data CopyObject

Request constructor

copyObject

Arguments

:: Text
coBucket

-> Text
coCopySource

-> Text
coKey

-> CopyObject

CopyObject constructor.

The fields accessible through corresponding lenses are:

coACL :: Maybe ObjectCannedACL

coBucket :: Text

coCacheControl :: Maybe Text

coContentDisposition :: Maybe Text

coContentEncoding :: Maybe Text

coContentLanguage :: Maybe Text

coContentType :: Maybe Text

coCopySource :: Text

coCopySourceIfMatch :: Maybe Text

coCopySourceIfModifiedSince :: Maybe UTCTime

coCopySourceIfNoneMatch :: Maybe Text

coCopySourceIfUnmodifiedSince :: Maybe UTCTime

coCopySourceSSECustomerAlgorithm :: Maybe Text

coCopySourceSSECustomerKey :: Maybe Text

coCopySourceSSECustomerKeyMD5 :: Maybe Text

coExpires :: Maybe UTCTime

coGrantFullControl :: Maybe Text

coGrantRead :: Maybe Text

coGrantReadACP :: Maybe Text

coGrantWriteACP :: Maybe Text

coKey :: Text

coMetadata :: HashMap (CI Text) Text

coMetadataDirective :: Maybe MetadataDirective

coRequestPayer :: Maybe RequestPayer

coSSECustomerAlgorithm :: Maybe Text

coSSECustomerKey :: Maybe Text

coSSECustomerKeyMD5 :: Maybe Text

coSSEKMSKeyId :: Maybe Text

coServerSideEncryption :: Maybe ServerSideEncryption

coStorageClass :: Maybe StorageClass

coWebsiteRedirectLocation :: Maybe Text

Request lenses

coACL :: Lens' CopyObject (Maybe ObjectCannedACL)

The canned ACL to apply to the object.

coBucket :: Lens' CopyObject Text

coCacheControl :: Lens' CopyObject (Maybe Text)

Specifies caching behavior along the request/reply chain.

coContentDisposition :: Lens' CopyObject (Maybe Text)

Specifies presentational information for the object.

coContentEncoding :: Lens' CopyObject (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

coContentLanguage :: Lens' CopyObject (Maybe Text)

The language the content is in.

coContentType :: Lens' CopyObject (Maybe Text)

A standard MIME type describing the format of the object data.

coCopySource :: Lens' CopyObject Text

The name of the source bucket and key name of the source object, separated by a slash (/). Must be URL-encoded.

coCopySourceIfMatch :: Lens' CopyObject (Maybe Text)

Copies the object if its entity tag (ETag) matches the specified tag.

coCopySourceIfModifiedSince :: Lens' CopyObject (Maybe UTCTime)

Copies the object if it has been modified since the specified time.

coCopySourceIfNoneMatch :: Lens' CopyObject (Maybe Text)

Copies the object if its entity tag (ETag) is different than the specified ETag.

coCopySourceIfUnmodifiedSince :: Lens' CopyObject (Maybe UTCTime)

Copies the object if it hasn't been modified since the specified time.

coCopySourceSSECustomerAlgorithm :: Lens' CopyObject (Maybe Text)

Specifies the algorithm to use when decrypting the source object (e.g., AES256).

coCopySourceSSECustomerKey :: Lens' CopyObject (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

coCopySourceSSECustomerKeyMD5 :: Lens' CopyObject (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

coExpires :: Lens' CopyObject (Maybe UTCTime)

The date and time at which the object is no longer cacheable.

coGrantFullControl :: Lens' CopyObject (Maybe Text)

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

coGrantRead :: Lens' CopyObject (Maybe Text)

Allows grantee to read the object data and its metadata.

coGrantReadACP :: Lens' CopyObject (Maybe Text)

Allows grantee to read the object ACL.

coGrantWriteACP :: Lens' CopyObject (Maybe Text)

Allows grantee to write the ACL for the applicable object.

coKey :: Lens' CopyObject Text

coMetadata :: Lens' CopyObject (HashMap (CI Text) Text)

A map of metadata to store with the object in S3.

coMetadataDirective :: Lens' CopyObject (Maybe MetadataDirective)

Specifies whether the metadata is copied from the source object or replaced with metadata provided in the request.

coRequestPayer :: Lens' CopyObject (Maybe RequestPayer)

coSSECustomerAlgorithm :: Lens' CopyObject (Maybe Text)

Specifies the algorithm to use to when encrypting the object (e.g., AES256, aws:kms).

coSSECustomerKey :: Lens' CopyObject (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

coSSECustomerKeyMD5 :: Lens' CopyObject (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

coSSEKMSKeyId :: Lens' CopyObject (Maybe Text)

Specifies the AWS KMS key ID to use for object encryption. All GET and PUT requests for an object protected by AWS KMS will fail if not made via SSL or using SigV4. Documentation on configuring any of the officially supported AWS SDKs and CLI can be found at http://docs.aws.amazon.com/AmazonS3/latest/dev/UsingAWSSDK.html#specify-signature-version

coServerSideEncryption :: Lens' CopyObject (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

coStorageClass :: Lens' CopyObject (Maybe StorageClass)

The type of storage to use for the object. Defaults to STANDARD.

coWebsiteRedirectLocation :: Lens' CopyObject (Maybe Text)

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

Response

data CopyObjectResponse

Response constructor

copyObjectResponse :: CopyObjectResponse

CopyObjectResponse constructor.

The fields accessible through corresponding lenses are:

corCopyObjectResult :: Maybe CopyObjectResult

corCopySourceVersionId :: Maybe Text

corExpiration :: Maybe Text

corRequestCharged :: Maybe RequestCharged

corSSECustomerAlgorithm :: Maybe Text

corSSECustomerKeyMD5 :: Maybe Text

corSSEKMSKeyId :: Maybe Text

corServerSideEncryption :: Maybe ServerSideEncryption

Response lenses

corCopyObjectResult :: Lens' CopyObjectResponse (Maybe CopyObjectResult)

corCopySourceVersionId :: Lens' CopyObjectResponse (Maybe Text)

corExpiration :: Lens' CopyObjectResponse (Maybe Text)

If the object expiration is configured, the response includes this header.

corRequestCharged :: Lens' CopyObjectResponse (Maybe RequestCharged)

corSSECustomerAlgorithm :: Lens' CopyObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

corSSECustomerKeyMD5 :: Lens' CopyObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

corSSEKMSKeyId :: Lens' CopyObjectResponse (Maybe Text)

If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

corServerSideEncryption :: Lens' CopyObjectResponse (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).