Network.AWS.S3.CreateBucket

Description

Creates a new bucket.

Creating a Request

createBucket

Arguments

:: BucketName
cbBucket

-> CreateBucket

Creates a value of CreateBucket with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cbGrantReadACP - Allows grantee to read the bucket ACL.

cbGrantWriteACP - Allows grantee to write the ACL for the applicable bucket.

cbGrantRead - Allows grantee to list the objects in the bucket.

cbGrantFullControl - Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

cbCreateBucketConfiguration - Undocumented member.

cbGrantWrite - Allows grantee to create, overwrite, and delete any object in the bucket.

cbACL - The canned ACL to apply to the bucket.

cbBucket - Undocumented member.

data CreateBucket

See: createBucket smart constructor.

Request Lenses

cbGrantReadACP :: Lens' CreateBucket (Maybe Text)

Allows grantee to read the bucket ACL.

cbGrantWriteACP :: Lens' CreateBucket (Maybe Text)

Allows grantee to write the ACL for the applicable bucket.

cbGrantRead :: Lens' CreateBucket (Maybe Text)

Allows grantee to list the objects in the bucket.

cbGrantFullControl :: Lens' CreateBucket (Maybe Text)

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

cbCreateBucketConfiguration :: Lens' CreateBucket (Maybe CreateBucketConfiguration)

Undocumented member.

cbGrantWrite :: Lens' CreateBucket (Maybe Text)

Allows grantee to create, overwrite, and delete any object in the bucket.

cbACL :: Lens' CreateBucket (Maybe BucketCannedACL)

The canned ACL to apply to the bucket.

cbBucket :: Lens' CreateBucket BucketName

Undocumented member.

Destructuring the Response

createBucketResponse

Arguments

:: Int
cbrsResponseStatus

-> CreateBucketResponse

Creates a value of CreateBucketResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cbrsLocation - Undocumented member.

cbrsResponseStatus - -- | The response status code.

data CreateBucketResponse

See: createBucketResponse smart constructor.

Response Lenses

cbrsLocation :: Lens' CreateBucketResponse (Maybe Text)

Undocumented member.

cbrsResponseStatus :: Lens' CreateBucketResponse Int

- | The response status code.