Network.AWS.S3.CreateMultipartUpload

Description

Initiates a multipart upload and returns an upload ID.

Note: After you initiate multipart upload and upload one or more parts, you must either complete or abort multipart upload in order to stop getting charged for storage of the uploaded parts. Only after you either complete or abort multipart upload, Amazon S3 frees up the parts storage and stops charging you for the parts storage.

Creating a Request

createMultipartUpload

Arguments

:: BucketName
cmuBucket

-> ObjectKey
cmuKey

-> CreateMultipartUpload

Creates a value of CreateMultipartUpload with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmuExpires - The date and time at which the object is no longer cacheable.

cmuGrantReadACP - Allows grantee to read the object ACL.

cmuSSECustomerAlgorithm - Specifies the algorithm to use to when encrypting the object (e.g., AES256).

cmuSSECustomerKey - Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

cmuRequestPayer - Undocumented member.

cmuGrantWriteACP - Allows grantee to write the ACL for the applicable object.

cmuWebsiteRedirectLocation - If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

cmuGrantRead - Allows grantee to read the object data and its metadata.

cmuStorageClass - The type of storage to use for the object. Defaults to STANDARD.

cmuSSECustomerKeyMD5 - Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

cmuSSEKMSKeyId - Specifies the AWS KMS key ID to use for object encryption. All GET and PUT requests for an object protected by AWS KMS will fail if not made via SSL or using SigV4. Documentation on configuring any of the officially supported AWS SDKs and CLI can be found at http://docs.aws.amazon.com/AmazonS3/latest/dev/UsingAWSSDK.html#specify-signature-version

cmuGrantFullControl - Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

cmuContentEncoding - Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

cmuTagging - The tag-set for the object. The tag-set must be encoded as URL Query parameters

cmuMetadata - A map of metadata to store with the object in S3.

cmuCacheControl - Specifies caching behavior along the request/reply chain.

cmuContentLanguage - The language the content is in.

cmuACL - The canned ACL to apply to the object.

cmuContentDisposition - Specifies presentational information for the object.

cmuServerSideEncryption - The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

cmuContentType - A standard MIME type describing the format of the object data.

cmuBucket - Undocumented member.

cmuKey - Undocumented member.

data CreateMultipartUpload

See: createMultipartUpload smart constructor.

Request Lenses

cmuExpires :: Lens' CreateMultipartUpload (Maybe UTCTime)

The date and time at which the object is no longer cacheable.

cmuGrantReadACP :: Lens' CreateMultipartUpload (Maybe Text)

Allows grantee to read the object ACL.

cmuSSECustomerAlgorithm :: Lens' CreateMultipartUpload (Maybe Text)

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

cmuSSECustomerKey :: Lens' CreateMultipartUpload (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

cmuRequestPayer :: Lens' CreateMultipartUpload (Maybe RequestPayer)

Undocumented member.

cmuGrantWriteACP :: Lens' CreateMultipartUpload (Maybe Text)

Allows grantee to write the ACL for the applicable object.

cmuWebsiteRedirectLocation :: Lens' CreateMultipartUpload (Maybe Text)

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

cmuGrantRead :: Lens' CreateMultipartUpload (Maybe Text)

Allows grantee to read the object data and its metadata.

cmuStorageClass :: Lens' CreateMultipartUpload (Maybe StorageClass)

The type of storage to use for the object. Defaults to STANDARD.

cmuSSECustomerKeyMD5 :: Lens' CreateMultipartUpload (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

cmuSSEKMSKeyId :: Lens' CreateMultipartUpload (Maybe Text)

Specifies the AWS KMS key ID to use for object encryption. All GET and PUT requests for an object protected by AWS KMS will fail if not made via SSL or using SigV4. Documentation on configuring any of the officially supported AWS SDKs and CLI can be found at http://docs.aws.amazon.com/AmazonS3/latest/dev/UsingAWSSDK.html#specify-signature-version

cmuGrantFullControl :: Lens' CreateMultipartUpload (Maybe Text)

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

cmuContentEncoding :: Lens' CreateMultipartUpload (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

cmuTagging :: Lens' CreateMultipartUpload (Maybe Text)

The tag-set for the object. The tag-set must be encoded as URL Query parameters

cmuMetadata :: Lens' CreateMultipartUpload (HashMap Text Text)

A map of metadata to store with the object in S3.

cmuCacheControl :: Lens' CreateMultipartUpload (Maybe Text)

Specifies caching behavior along the request/reply chain.

cmuContentLanguage :: Lens' CreateMultipartUpload (Maybe Text)

The language the content is in.

cmuACL :: Lens' CreateMultipartUpload (Maybe ObjectCannedACL)

The canned ACL to apply to the object.

cmuContentDisposition :: Lens' CreateMultipartUpload (Maybe Text)

Specifies presentational information for the object.

cmuServerSideEncryption :: Lens' CreateMultipartUpload (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

cmuContentType :: Lens' CreateMultipartUpload (Maybe Text)

A standard MIME type describing the format of the object data.

cmuBucket :: Lens' CreateMultipartUpload BucketName

Undocumented member.

cmuKey :: Lens' CreateMultipartUpload ObjectKey

Undocumented member.

Destructuring the Response

createMultipartUploadResponse

Arguments

:: Int
cmursResponseStatus

-> CreateMultipartUploadResponse

Creates a value of CreateMultipartUploadResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

cmursRequestCharged - Undocumented member.

cmursBucket - Name of the bucket to which the multipart upload was initiated.

cmursSSECustomerAlgorithm - If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

cmursAbortDate - Date when multipart upload will become eligible for abort operation by lifecycle.

cmursAbortRuleId - Id of the lifecycle rule that makes a multipart upload eligible for abort operation.

cmursKey - Object key for which the multipart upload was initiated.

cmursSSECustomerKeyMD5 - If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

cmursSSEKMSKeyId - If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

cmursUploadId - ID for the initiated multipart upload.

cmursServerSideEncryption - The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

cmursResponseStatus - -- | The response status code.

data CreateMultipartUploadResponse

See: createMultipartUploadResponse smart constructor.

Response Lenses

cmursRequestCharged :: Lens' CreateMultipartUploadResponse (Maybe RequestCharged)

Undocumented member.

cmursBucket :: Lens' CreateMultipartUploadResponse (Maybe BucketName)

Name of the bucket to which the multipart upload was initiated.

cmursSSECustomerAlgorithm :: Lens' CreateMultipartUploadResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

cmursAbortDate :: Lens' CreateMultipartUploadResponse (Maybe UTCTime)

Date when multipart upload will become eligible for abort operation by lifecycle.

cmursAbortRuleId :: Lens' CreateMultipartUploadResponse (Maybe Text)

Id of the lifecycle rule that makes a multipart upload eligible for abort operation.

cmursKey :: Lens' CreateMultipartUploadResponse (Maybe ObjectKey)

Object key for which the multipart upload was initiated.

cmursSSECustomerKeyMD5 :: Lens' CreateMultipartUploadResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

cmursSSEKMSKeyId :: Lens' CreateMultipartUploadResponse (Maybe Text)

If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

cmursUploadId :: Lens' CreateMultipartUploadResponse (Maybe Text)

ID for the initiated multipart upload.

cmursServerSideEncryption :: Lens' CreateMultipartUploadResponse (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

cmursResponseStatus :: Lens' CreateMultipartUploadResponse Int

- | The response status code.