Network.AWS.S3.CreateMultipartUpload

Description

Initiates a multipart upload and returns an upload ID.

Note: After you initiate multipart upload and upload one or more parts, you must either complete or abort multipart upload in order to stop getting charged for storage of the uploaded parts. Only after you either complete or abort multipart upload, Amazon S3 frees up the parts storage and stops charging you for the parts storage.

http://docs.aws.amazon.com/AmazonS3/latest/API/CreateMultipartUpload.html

Request

data CreateMultipartUpload

Request constructor

createMultipartUpload

Arguments

:: Text
cmuBucket

-> Text
cmuKey

-> CreateMultipartUpload

CreateMultipartUpload constructor.

The fields accessible through corresponding lenses are:

cmuACL :: Maybe ObjectCannedACL

cmuBucket :: Text

cmuCacheControl :: Maybe Text

cmuContentDisposition :: Maybe Text

cmuContentEncoding :: Maybe Text

cmuContentLanguage :: Maybe Text

cmuContentType :: Maybe Text

cmuExpires :: Maybe UTCTime

cmuGrantFullControl :: Maybe Text

cmuGrantRead :: Maybe Text

cmuGrantReadACP :: Maybe Text

cmuGrantWriteACP :: Maybe Text

cmuKey :: Text

cmuMetadata :: HashMap (CI Text) Text

cmuRequestPayer :: Maybe RequestPayer

cmuSSECustomerAlgorithm :: Maybe Text

cmuSSECustomerKey :: Maybe Text

cmuSSECustomerKeyMD5 :: Maybe Text

cmuSSEKMSKeyId :: Maybe Text

cmuServerSideEncryption :: Maybe ServerSideEncryption

cmuStorageClass :: Maybe StorageClass

cmuWebsiteRedirectLocation :: Maybe Text

Request lenses

cmuACL :: Lens' CreateMultipartUpload (Maybe ObjectCannedACL)

The canned ACL to apply to the object.

cmuBucket :: Lens' CreateMultipartUpload Text

cmuCacheControl :: Lens' CreateMultipartUpload (Maybe Text)

Specifies caching behavior along the request/reply chain.

cmuContentDisposition :: Lens' CreateMultipartUpload (Maybe Text)

Specifies presentational information for the object.

cmuContentEncoding :: Lens' CreateMultipartUpload (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

cmuContentLanguage :: Lens' CreateMultipartUpload (Maybe Text)

The language the content is in.

cmuContentType :: Lens' CreateMultipartUpload (Maybe Text)

A standard MIME type describing the format of the object data.

cmuExpires :: Lens' CreateMultipartUpload (Maybe UTCTime)

The date and time at which the object is no longer cacheable.

cmuGrantFullControl :: Lens' CreateMultipartUpload (Maybe Text)

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

cmuGrantRead :: Lens' CreateMultipartUpload (Maybe Text)

Allows grantee to read the object data and its metadata.

cmuGrantReadACP :: Lens' CreateMultipartUpload (Maybe Text)

Allows grantee to read the object ACL.

cmuGrantWriteACP :: Lens' CreateMultipartUpload (Maybe Text)

Allows grantee to write the ACL for the applicable object.

cmuKey :: Lens' CreateMultipartUpload Text

cmuMetadata :: Lens' CreateMultipartUpload (HashMap (CI Text) Text)

A map of metadata to store with the object in S3.

cmuRequestPayer :: Lens' CreateMultipartUpload (Maybe RequestPayer)

cmuSSECustomerAlgorithm :: Lens' CreateMultipartUpload (Maybe Text)

Specifies the algorithm to use to when encrypting the object (e.g., AES256, aws:kms).

cmuSSECustomerKey :: Lens' CreateMultipartUpload (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

cmuSSECustomerKeyMD5 :: Lens' CreateMultipartUpload (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

cmuSSEKMSKeyId :: Lens' CreateMultipartUpload (Maybe Text)

Specifies the AWS KMS key ID to use for object encryption. All GET and PUT requests for an object protected by AWS KMS will fail if not made via SSL or using SigV4. Documentation on configuring any of the officially supported AWS SDKs and CLI can be found at http://docs.aws.amazon.com/AmazonS3/latest/dev/UsingAWSSDK.html#specify-signature-version

cmuServerSideEncryption :: Lens' CreateMultipartUpload (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

cmuStorageClass :: Lens' CreateMultipartUpload (Maybe StorageClass)

The type of storage to use for the object. Defaults to STANDARD.

cmuWebsiteRedirectLocation :: Lens' CreateMultipartUpload (Maybe Text)

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

Response

data CreateMultipartUploadResponse

Response constructor

createMultipartUploadResponse :: CreateMultipartUploadResponse

CreateMultipartUploadResponse constructor.

The fields accessible through corresponding lenses are:

cmurBucket :: Maybe Text

cmurKey :: Maybe Text

cmurRequestCharged :: Maybe RequestCharged

cmurSSECustomerAlgorithm :: Maybe Text

cmurSSECustomerKeyMD5 :: Maybe Text

cmurSSEKMSKeyId :: Maybe Text

cmurServerSideEncryption :: Maybe ServerSideEncryption

cmurUploadId :: Maybe Text

Response lenses

cmurBucket :: Lens' CreateMultipartUploadResponse (Maybe Text)

Name of the bucket to which the multipart upload was initiated.

cmurKey :: Lens' CreateMultipartUploadResponse (Maybe Text)

Object key for which the multipart upload was initiated.

cmurRequestCharged :: Lens' CreateMultipartUploadResponse (Maybe RequestCharged)

cmurSSECustomerAlgorithm :: Lens' CreateMultipartUploadResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

cmurSSECustomerKeyMD5 :: Lens' CreateMultipartUploadResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

cmurSSEKMSKeyId :: Lens' CreateMultipartUploadResponse (Maybe Text)

If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

cmurServerSideEncryption :: Lens' CreateMultipartUploadResponse (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

cmurUploadId :: Lens' CreateMultipartUploadResponse (Maybe Text)

ID for the initiated multipart upload.