Network.AWS.S3.DeleteObject

Description

Removes the null version (if there is one) of an object and inserts a delete marker, which becomes the latest version of the object. If there isn't a null version, Amazon S3 does not remove any objects.

Creating a Request

deleteObject

Arguments

:: BucketName
doBucket

-> ObjectKey
doKey

-> DeleteObject

Creates a value of DeleteObject with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

doVersionId - VersionId used to reference a specific version of the object.

doMFA - The concatenation of the authentication device's serial number, a space, and the value that is displayed on your authentication device.

doRequestPayer - Undocumented member.

doBucket - Undocumented member.

doKey - Undocumented member.

data DeleteObject

See: deleteObject smart constructor.

Request Lenses

doVersionId :: Lens' DeleteObject (Maybe ObjectVersionId)

VersionId used to reference a specific version of the object.

doMFA :: Lens' DeleteObject (Maybe Text)

The concatenation of the authentication device's serial number, a space, and the value that is displayed on your authentication device.

doRequestPayer :: Lens' DeleteObject (Maybe RequestPayer)

Undocumented member.

doBucket :: Lens' DeleteObject BucketName

Undocumented member.

doKey :: Lens' DeleteObject ObjectKey

Undocumented member.

Destructuring the Response

deleteObjectResponse

Arguments

:: Int
dorsResponseStatus

-> DeleteObjectResponse

Creates a value of DeleteObjectResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dorsRequestCharged - Undocumented member.

dorsVersionId - Returns the version ID of the delete marker created as a result of the DELETE operation.

dorsDeleteMarker - Specifies whether the versioned object that was permanently deleted was (true) or was not (false) a delete marker.

dorsResponseStatus - -- | The response status code.

data DeleteObjectResponse

See: deleteObjectResponse smart constructor.

Response Lenses

dorsRequestCharged :: Lens' DeleteObjectResponse (Maybe RequestCharged)

Undocumented member.

dorsVersionId :: Lens' DeleteObjectResponse (Maybe ObjectVersionId)

Returns the version ID of the delete marker created as a result of the DELETE operation.

dorsDeleteMarker :: Lens' DeleteObjectResponse (Maybe Bool)

Specifies whether the versioned object that was permanently deleted was (true) or was not (false) a delete marker.

dorsResponseStatus :: Lens' DeleteObjectResponse Int

- | The response status code.