Network.AWS.S3.DeleteObject

Description

Removes the null version (if there is one) of an object and inserts a delete marker, which becomes the latest version of the object. If there isn't a null version, Amazon S3 does not remove any objects.

Request

data DeleteObject

Request constructor

mkDeleteObject

Arguments

:: BucketName
doBucket

-> ObjectKey
doKey

-> DeleteObject

Smart constructor for the minimum required parameters to construct a valid DeleteObject request.

The fields accessible through corresponding lenses are:

Bucket :: BucketName

Key :: ObjectKey

MFA :: Maybe Text

VersionId :: Maybe ObjectVersionId

Request lenses

doBucket :: Lens' DeleteObject BucketName

doKey :: Lens' DeleteObject ObjectKey

doMFA :: Lens' DeleteObject (Maybe Text)

The concatenation of the authentication device's serial number, a space, and the value that is displayed on your authentication device.

doVersionId :: Lens' DeleteObject (Maybe ObjectVersionId)

VersionId used to reference a specific version of the object.

Response

data DeleteObjectResponse

Response constructor

mkDeleteObjectResponse :: DeleteObjectResponse

Smart constructor for the minimum required parameters to construct a valid DeleteObjectResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

DeleteMarker :: Maybe Bool

VersionId :: Maybe ObjectVersionId

Response lenses

dorDeleteMarker :: Lens' DeleteObjectResponse (Maybe Bool)

Specifies whether the versioned object that was permanently deleted was (true) or was not (false) a delete marker.

dorVersionId :: Lens' DeleteObjectResponse (Maybe ObjectVersionId)

Returns the version ID of the delete marker created as a result of the DELETE operation.