Network.AWS.S3.DeleteObject

Description

Removes the null version (if there is one) of an object and inserts a delete marker, which becomes the latest version of the object. If there isn't a null version, Amazon S3 does not remove any objects.

http://docs.aws.amazon.com/AmazonS3/latest/API/DeleteObject.html

Request

data DeleteObject

Request constructor

deleteObject

Arguments

:: Text
doBucket

-> Text
doKey

-> DeleteObject

DeleteObject constructor.

The fields accessible through corresponding lenses are:

doBucket :: Text

doKey :: Text

doMFA :: Maybe Text

doVersionId :: Maybe Text

Request lenses

doBucket :: Lens' DeleteObject Text

doKey :: Lens' DeleteObject Text

doMFA :: Lens' DeleteObject (Maybe Text)

The concatenation of the authentication device's serial number, a space, and the value that is displayed on your authentication device.

doVersionId :: Lens' DeleteObject (Maybe Text)

VersionId used to reference a specific version of the object.

Response

data DeleteObjectResponse

Response constructor

deleteObjectResponse :: DeleteObjectResponse

DeleteObjectResponse constructor.

The fields accessible through corresponding lenses are:

dorDeleteMarker :: Maybe Bool

dorVersionId :: Maybe Text

Response lenses

dorDeleteMarker :: Lens' DeleteObjectResponse (Maybe Bool)

Specifies whether the versioned object that was permanently deleted was (true) or was not (false) a delete marker.

dorVersionId :: Lens' DeleteObjectResponse (Maybe Text)

Returns the version ID of the delete marker created as a result of the DELETE operation.