Network.AWS.S3.DeleteObjectTagging

Description

Removes the tag-set from an existing object.

Creating a Request

deleteObjectTagging

Arguments

:: BucketName
dotBucket

-> ObjectKey
dotKey

-> DeleteObjectTagging

Creates a value of DeleteObjectTagging with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dotVersionId - The versionId of the object that the tag-set will be removed from.

dotBucket - Undocumented member.

dotKey - Undocumented member.

data DeleteObjectTagging

See: deleteObjectTagging smart constructor.

Request Lenses

dotVersionId :: Lens' DeleteObjectTagging (Maybe ObjectVersionId)

The versionId of the object that the tag-set will be removed from.

dotBucket :: Lens' DeleteObjectTagging BucketName

Undocumented member.

dotKey :: Lens' DeleteObjectTagging ObjectKey

Undocumented member.

Destructuring the Response

deleteObjectTaggingResponse

Arguments

:: Int
dotrsResponseStatus

-> DeleteObjectTaggingResponse

Creates a value of DeleteObjectTaggingResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dotrsVersionId - The versionId of the object the tag-set was removed from.

dotrsResponseStatus - -- | The response status code.

data DeleteObjectTaggingResponse

See: deleteObjectTaggingResponse smart constructor.

Response Lenses

dotrsVersionId :: Lens' DeleteObjectTaggingResponse (Maybe ObjectVersionId)

The versionId of the object the tag-set was removed from.

dotrsResponseStatus :: Lens' DeleteObjectTaggingResponse Int

- | The response status code.