Network.AWS.S3.DeleteObjects

Description

This operation enables you to delete multiple objects from a bucket using a single HTTP request. You may specify up to 1000 keys.

Creating a Request

deleteObjects

Arguments

:: BucketName
dosBucket

-> Delete
dosDelete

-> DeleteObjects

Creates a value of DeleteObjects with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

dosMFA - The concatenation of the authentication device's serial number, a space, and the value that is displayed on your authentication device.

dosRequestPayer - Undocumented member.

dosBucket - Undocumented member.

dosDelete - Undocumented member.

data DeleteObjects

See: deleteObjects smart constructor.

Request Lenses

dosMFA :: Lens' DeleteObjects (Maybe Text)

The concatenation of the authentication device's serial number, a space, and the value that is displayed on your authentication device.

dosRequestPayer :: Lens' DeleteObjects (Maybe RequestPayer)

Undocumented member.

dosBucket :: Lens' DeleteObjects BucketName

Undocumented member.

dosDelete :: Lens' DeleteObjects Delete

Undocumented member.

Destructuring the Response

deleteObjectsResponse

Arguments

:: Int
drsResponseStatus

-> DeleteObjectsResponse

Creates a value of DeleteObjectsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

drsRequestCharged - Undocumented member.

drsDeleted - Undocumented member.

drsErrors - Undocumented member.

drsResponseStatus - -- | The response status code.

data DeleteObjectsResponse

See: deleteObjectsResponse smart constructor.

Response Lenses

drsRequestCharged :: Lens' DeleteObjectsResponse (Maybe RequestCharged)

Undocumented member.

drsDeleted :: Lens' DeleteObjectsResponse [DeletedObject]

Undocumented member.

drsErrors :: Lens' DeleteObjectsResponse [S3ServiceError]

Undocumented member.

drsResponseStatus :: Lens' DeleteObjectsResponse Int

- | The response status code.