Network.AWS.S3.DeleteObjects

Description

This operation enables you to delete multiple objects from a bucket using a single HTTP request. You may specify up to 1000 keys.

http://docs.aws.amazon.com/AmazonS3/latest/API/DeleteObjects.html

Request

data DeleteObjects

Request constructor

deleteObjects

Arguments

:: Text
do1Bucket

-> Delete
do1Delete

-> DeleteObjects

DeleteObjects constructor.

The fields accessible through corresponding lenses are:

do1Bucket :: Text

do1Delete :: Delete

do1MFA :: Maybe Text

do1RequestPayer :: Maybe RequestPayer

Request lenses

do1Bucket :: Lens' DeleteObjects Text

do1Delete :: Lens' DeleteObjects Delete

do1MFA :: Lens' DeleteObjects (Maybe Text)

The concatenation of the authentication device's serial number, a space, and the value that is displayed on your authentication device.

do1RequestPayer :: Lens' DeleteObjects (Maybe RequestPayer)

Response

data DeleteObjectsResponse

Response constructor

deleteObjectsResponse :: DeleteObjectsResponse

DeleteObjectsResponse constructor.

The fields accessible through corresponding lenses are:

dor1Deleted :: [DeletedObject]

dor1Errors :: [S3ServiceError]

dor1RequestCharged :: Maybe RequestCharged

Response lenses

dor1Deleted :: Lens' DeleteObjectsResponse [DeletedObject]

dor1Errors :: Lens' DeleteObjectsResponse [S3ServiceError]

dor1RequestCharged :: Lens' DeleteObjectsResponse (Maybe RequestCharged)