Network.AWS.S3.GetBucketLogging

Description

Returns the logging status of a bucket and the permissions users have to view and modify that status. To use GET, you must be the bucket owner.

Creating a Request

getBucketLogging

Arguments

:: BucketName
gBucket

-> GetBucketLogging

Creates a value of GetBucketLogging with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gBucket - Undocumented member.

data GetBucketLogging

See: getBucketLogging smart constructor.

Request Lenses

gBucket :: Lens' GetBucketLogging BucketName

Undocumented member.

Destructuring the Response

getBucketLoggingResponse

Arguments

:: Int
gblrsResponseStatus

-> GetBucketLoggingResponse

Creates a value of GetBucketLoggingResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

gblrsLoggingEnabled - Undocumented member.

gblrsResponseStatus - -- | The response status code.

data GetBucketLoggingResponse

See: getBucketLoggingResponse smart constructor.

Response Lenses

gblrsLoggingEnabled :: Lens' GetBucketLoggingResponse (Maybe LoggingEnabled)

Undocumented member.

gblrsResponseStatus :: Lens' GetBucketLoggingResponse Int

- | The response status code.