Network.AWS.S3.GetObjectACL

Description

Returns the access control list (ACL) of an object.

Creating a Request

getObjectACL

Arguments

:: BucketName
goaBucket

-> ObjectKey
goaKey

-> GetObjectACL

Creates a value of GetObjectACL with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

goaVersionId - VersionId used to reference a specific version of the object.

goaRequestPayer - Undocumented member.

goaBucket - Undocumented member.

goaKey - Undocumented member.

data GetObjectACL

See: getObjectACL smart constructor.

Request Lenses

goaVersionId :: Lens' GetObjectACL (Maybe ObjectVersionId)

VersionId used to reference a specific version of the object.

goaRequestPayer :: Lens' GetObjectACL (Maybe RequestPayer)

Undocumented member.

goaBucket :: Lens' GetObjectACL BucketName

Undocumented member.

goaKey :: Lens' GetObjectACL ObjectKey

Undocumented member.

Destructuring the Response

getObjectACLResponse

Arguments

:: Int
goarsResponseStatus

-> GetObjectACLResponse

Creates a value of GetObjectACLResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

goarsRequestCharged - Undocumented member.

goarsGrants - A list of grants.

goarsOwner - Undocumented member.

goarsResponseStatus - -- | The response status code.

data GetObjectACLResponse

See: getObjectACLResponse smart constructor.

Response Lenses

goarsRequestCharged :: Lens' GetObjectACLResponse (Maybe RequestCharged)

Undocumented member.

goarsGrants :: Lens' GetObjectACLResponse [Grant]

A list of grants.

goarsOwner :: Lens' GetObjectACLResponse (Maybe Owner)

Undocumented member.

goarsResponseStatus :: Lens' GetObjectACLResponse Int

- | The response status code.