Network.AWS.S3.HeadBucket

Description

This operation is useful to determine if a bucket exists and you have permission to access it.

Creating a Request

headBucket

Arguments

:: BucketName
hbBucket

-> HeadBucket

Creates a value of HeadBucket with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

hbBucket - Undocumented member.

data HeadBucket

See: headBucket smart constructor.

Request Lenses

hbBucket :: Lens' HeadBucket BucketName

Undocumented member.

Destructuring the Response

headBucketResponse :: HeadBucketResponse

Creates a value of HeadBucketResponse with the minimum fields required to make a request.

data HeadBucketResponse

See: headBucketResponse smart constructor.