Network.AWS.S3.HeadObject

Description

The HEAD operation retrieves metadata from an object without returning the object itself. This operation is useful if you're only interested in an object's metadata. To use HEAD, you must have READ access to the object.

Creating a Request

headObject

Arguments

:: BucketName
hoBucket

-> ObjectKey
hoKey

-> HeadObject

Creates a value of HeadObject with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

hoIfMatch - Return the object only if its entity tag (ETag) is the same as the one specified, otherwise return a 412 (precondition failed).

hoVersionId - VersionId used to reference a specific version of the object.

hoSSECustomerAlgorithm - Specifies the algorithm to use to when encrypting the object (e.g., AES256).

hoSSECustomerKey - Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

hoRequestPayer - Undocumented member.

hoIfModifiedSince - Return the object only if it has been modified since the specified time, otherwise return a 304 (not modified).

hoPartNumber - Part number of the object being read. This is a positive integer between 1 and 10,000. Effectively performs a ranged HEAD request for the part specified. Useful querying about the size of the part and the number of parts in this object.

hoRange - Downloads the specified range bytes of an object. For more information about the HTTP Range header, go to http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.35.

hoIfUnmodifiedSince - Return the object only if it has not been modified since the specified time, otherwise return a 412 (precondition failed).

hoSSECustomerKeyMD5 - Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

hoIfNoneMatch - Return the object only if its entity tag (ETag) is different from the one specified, otherwise return a 304 (not modified).

hoBucket - Undocumented member.

hoKey - Undocumented member.

data HeadObject

See: headObject smart constructor.

Request Lenses

hoIfMatch :: Lens' HeadObject (Maybe Text)

Return the object only if its entity tag (ETag) is the same as the one specified, otherwise return a 412 (precondition failed).

hoVersionId :: Lens' HeadObject (Maybe ObjectVersionId)

VersionId used to reference a specific version of the object.

hoSSECustomerAlgorithm :: Lens' HeadObject (Maybe Text)

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

hoSSECustomerKey :: Lens' HeadObject (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

hoRequestPayer :: Lens' HeadObject (Maybe RequestPayer)

Undocumented member.

hoIfModifiedSince :: Lens' HeadObject (Maybe UTCTime)

Return the object only if it has been modified since the specified time, otherwise return a 304 (not modified).

hoPartNumber :: Lens' HeadObject (Maybe Int)

Part number of the object being read. This is a positive integer between 1 and 10,000. Effectively performs a ranged HEAD request for the part specified. Useful querying about the size of the part and the number of parts in this object.

hoRange :: Lens' HeadObject (Maybe Text)

Downloads the specified range bytes of an object. For more information about the HTTP Range header, go to http://www.w3.org/Protocols/rfc2616/rfc2616-sec14.html#sec14.35.

hoIfUnmodifiedSince :: Lens' HeadObject (Maybe UTCTime)

Return the object only if it has not been modified since the specified time, otherwise return a 412 (precondition failed).

hoSSECustomerKeyMD5 :: Lens' HeadObject (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

hoIfNoneMatch :: Lens' HeadObject (Maybe Text)

Return the object only if its entity tag (ETag) is different from the one specified, otherwise return a 304 (not modified).

hoBucket :: Lens' HeadObject BucketName

Undocumented member.

hoKey :: Lens' HeadObject ObjectKey

Undocumented member.

Destructuring the Response

headObjectResponse

Arguments

:: Int
horsResponseStatus

-> HeadObjectResponse

Creates a value of HeadObjectResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

horsRequestCharged - Undocumented member.

horsPartsCount - The count of parts this object has.

horsETag - An ETag is an opaque identifier assigned by a web server to a specific version of a resource found at a URL

horsVersionId - Version of the object.

horsContentLength - Size of the body in bytes.

horsExpires - The date and time at which the object is no longer cacheable.

horsRestore - Provides information about object restoration operation and expiration time of the restored object copy.

horsExpiration - If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key value pairs providing object expiration information. The value of the rule-id is URL encoded.

horsDeleteMarker - Specifies whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

horsSSECustomerAlgorithm - If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

horsMissingMeta - This is set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

horsWebsiteRedirectLocation - If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

horsAcceptRanges - Undocumented member.

horsStorageClass - Undocumented member.

horsSSECustomerKeyMD5 - If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

horsSSEKMSKeyId - If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

horsContentEncoding - Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

horsMetadata - A map of metadata to store with the object in S3.

horsReplicationStatus - Undocumented member.

horsCacheControl - Specifies caching behavior along the request/reply chain.

horsContentLanguage - The language the content is in.

horsLastModified - Last modified date of the object

horsContentDisposition - Specifies presentational information for the object.

horsServerSideEncryption - The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

horsContentType - A standard MIME type describing the format of the object data.

horsResponseStatus - -- | The response status code.

data HeadObjectResponse

See: headObjectResponse smart constructor.

Response Lenses

horsRequestCharged :: Lens' HeadObjectResponse (Maybe RequestCharged)

Undocumented member.

horsPartsCount :: Lens' HeadObjectResponse (Maybe Int)

The count of parts this object has.

horsETag :: Lens' HeadObjectResponse (Maybe ETag)

An ETag is an opaque identifier assigned by a web server to a specific version of a resource found at a URL

horsVersionId :: Lens' HeadObjectResponse (Maybe ObjectVersionId)

Version of the object.

horsContentLength :: Lens' HeadObjectResponse (Maybe Integer)

Size of the body in bytes.

horsExpires :: Lens' HeadObjectResponse (Maybe UTCTime)

The date and time at which the object is no longer cacheable.

horsRestore :: Lens' HeadObjectResponse (Maybe Text)

Provides information about object restoration operation and expiration time of the restored object copy.

horsExpiration :: Lens' HeadObjectResponse (Maybe Text)

If the object expiration is configured (see PUT Bucket lifecycle), the response includes this header. It includes the expiry-date and rule-id key value pairs providing object expiration information. The value of the rule-id is URL encoded.

horsDeleteMarker :: Lens' HeadObjectResponse (Maybe Bool)

Specifies whether the object retrieved was (true) or was not (false) a Delete Marker. If false, this response header does not appear in the response.

horsSSECustomerAlgorithm :: Lens' HeadObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

horsMissingMeta :: Lens' HeadObjectResponse (Maybe Int)

This is set to the number of metadata entries not returned in x-amz-meta headers. This can happen if you create metadata using an API like SOAP that supports more flexible metadata than the REST API. For example, using SOAP, you can create metadata whose values are not legal HTTP headers.

horsWebsiteRedirectLocation :: Lens' HeadObjectResponse (Maybe Text)

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

horsAcceptRanges :: Lens' HeadObjectResponse (Maybe Text)

Undocumented member.

horsStorageClass :: Lens' HeadObjectResponse (Maybe StorageClass)

Undocumented member.

horsSSECustomerKeyMD5 :: Lens' HeadObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

horsSSEKMSKeyId :: Lens' HeadObjectResponse (Maybe Text)

If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

horsContentEncoding :: Lens' HeadObjectResponse (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

horsMetadata :: Lens' HeadObjectResponse (HashMap Text Text)

A map of metadata to store with the object in S3.

horsReplicationStatus :: Lens' HeadObjectResponse (Maybe ReplicationStatus)

Undocumented member.

horsCacheControl :: Lens' HeadObjectResponse (Maybe Text)

Specifies caching behavior along the request/reply chain.

horsContentLanguage :: Lens' HeadObjectResponse (Maybe Text)

The language the content is in.

horsLastModified :: Lens' HeadObjectResponse (Maybe UTCTime)

Last modified date of the object

horsContentDisposition :: Lens' HeadObjectResponse (Maybe Text)

Specifies presentational information for the object.

horsServerSideEncryption :: Lens' HeadObjectResponse (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

horsContentType :: Lens' HeadObjectResponse (Maybe Text)

A standard MIME type describing the format of the object data.

horsResponseStatus :: Lens' HeadObjectResponse Int

- | The response status code.