Network.AWS.S3.ListBucketAnalyticsConfigurations

Description

Lists the analytics configurations for the bucket.

Creating a Request

listBucketAnalyticsConfigurations

Arguments

:: BucketName
lbacBucket

-> ListBucketAnalyticsConfigurations

Creates a value of ListBucketAnalyticsConfigurations with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbacContinuationToken - The ContinuationToken that represents a placeholder from where this request should begin.

lbacBucket - The name of the bucket from which analytics configurations are retrieved.

data ListBucketAnalyticsConfigurations

See: listBucketAnalyticsConfigurations smart constructor.

Request Lenses

lbacContinuationToken :: Lens' ListBucketAnalyticsConfigurations (Maybe Text)

The ContinuationToken that represents a placeholder from where this request should begin.

lbacBucket :: Lens' ListBucketAnalyticsConfigurations BucketName

The name of the bucket from which analytics configurations are retrieved.

Destructuring the Response

listBucketAnalyticsConfigurationsResponse

Arguments

:: Int
lbacrsResponseStatus

-> ListBucketAnalyticsConfigurationsResponse

Creates a value of ListBucketAnalyticsConfigurationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbacrsAnalyticsConfigurationList - The list of analytics configurations for a bucket.

lbacrsContinuationToken - The ContinuationToken that represents where this request began.

lbacrsNextContinuationToken - NextContinuationToken is sent when isTruncated is true, which indicates that there are more analytics configurations to list. The next request must include this NextContinuationToken. The token is obfuscated and is not a usable value.

lbacrsIsTruncated - Indicates whether the returned list of analytics configurations is complete. A value of true indicates that the list is not complete and the NextContinuationToken will be provided for a subsequent request.

lbacrsResponseStatus - -- | The response status code.

data ListBucketAnalyticsConfigurationsResponse

See: listBucketAnalyticsConfigurationsResponse smart constructor.

Response Lenses

lbacrsAnalyticsConfigurationList :: Lens' ListBucketAnalyticsConfigurationsResponse [AnalyticsConfiguration]

The list of analytics configurations for a bucket.

lbacrsContinuationToken :: Lens' ListBucketAnalyticsConfigurationsResponse (Maybe Text)

The ContinuationToken that represents where this request began.

lbacrsNextContinuationToken :: Lens' ListBucketAnalyticsConfigurationsResponse (Maybe Text)

NextContinuationToken is sent when isTruncated is true, which indicates that there are more analytics configurations to list. The next request must include this NextContinuationToken. The token is obfuscated and is not a usable value.

lbacrsIsTruncated :: Lens' ListBucketAnalyticsConfigurationsResponse (Maybe Bool)

Indicates whether the returned list of analytics configurations is complete. A value of true indicates that the list is not complete and the NextContinuationToken will be provided for a subsequent request.

lbacrsResponseStatus :: Lens' ListBucketAnalyticsConfigurationsResponse Int

- | The response status code.