Network.AWS.S3.ListBucketInventoryConfigurations

Description

Returns a list of inventory configurations for the bucket.

Creating a Request

listBucketInventoryConfigurations

Arguments

:: BucketName
lbicBucket

-> ListBucketInventoryConfigurations

Creates a value of ListBucketInventoryConfigurations with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbicContinuationToken - The marker used to continue an inventory configuration listing that has been truncated. Use the NextContinuationToken from a previously truncated list response to continue the listing. The continuation token is an opaque value that Amazon S3 understands.

lbicBucket - The name of the bucket containing the inventory configurations to retrieve.

data ListBucketInventoryConfigurations

See: listBucketInventoryConfigurations smart constructor.

Request Lenses

lbicContinuationToken :: Lens' ListBucketInventoryConfigurations (Maybe Text)

The marker used to continue an inventory configuration listing that has been truncated. Use the NextContinuationToken from a previously truncated list response to continue the listing. The continuation token is an opaque value that Amazon S3 understands.

lbicBucket :: Lens' ListBucketInventoryConfigurations BucketName

The name of the bucket containing the inventory configurations to retrieve.

Destructuring the Response

listBucketInventoryConfigurationsResponse

Arguments

:: Int
lbicrsResponseStatus

-> ListBucketInventoryConfigurationsResponse

Creates a value of ListBucketInventoryConfigurationsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lbicrsContinuationToken - If sent in the request, the marker that is used as a starting point for this inventory configuration list response.

lbicrsInventoryConfigurationList - The list of inventory configurations for a bucket.

lbicrsNextContinuationToken - The marker used to continue this inventory configuration listing. Use the NextContinuationToken from this response to continue the listing in a subsequent request. The continuation token is an opaque value that Amazon S3 understands.

lbicrsIsTruncated - Indicates whether the returned list of inventory configurations is truncated in this response. A value of true indicates that the list is truncated.

lbicrsResponseStatus - -- | The response status code.

data ListBucketInventoryConfigurationsResponse

See: listBucketInventoryConfigurationsResponse smart constructor.

Response Lenses

lbicrsContinuationToken :: Lens' ListBucketInventoryConfigurationsResponse (Maybe Text)

If sent in the request, the marker that is used as a starting point for this inventory configuration list response.

lbicrsInventoryConfigurationList :: Lens' ListBucketInventoryConfigurationsResponse [InventoryConfiguration]

The list of inventory configurations for a bucket.

lbicrsNextContinuationToken :: Lens' ListBucketInventoryConfigurationsResponse (Maybe Text)

The marker used to continue this inventory configuration listing. Use the NextContinuationToken from this response to continue the listing in a subsequent request. The continuation token is an opaque value that Amazon S3 understands.

lbicrsIsTruncated :: Lens' ListBucketInventoryConfigurationsResponse (Maybe Bool)

Indicates whether the returned list of inventory configurations is truncated in this response. A value of true indicates that the list is truncated.

lbicrsResponseStatus :: Lens' ListBucketInventoryConfigurationsResponse Int

- | The response status code.