Network.AWS.S3.ListMultipartUploads

Description

This operation lists in-progress multipart uploads.

This operation returns paginated results.

Creating a Request

listMultipartUploads

Arguments

:: BucketName
lmuBucket

-> ListMultipartUploads

Creates a value of ListMultipartUploads with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lmuKeyMarker - Together with upload-id-marker, this parameter specifies the multipart upload after which listing should begin.

lmuPrefix - Lists in-progress uploads only for those keys that begin with the specified prefix.

lmuEncodingType - Undocumented member.

lmuUploadIdMarker - Together with key-marker, specifies the multipart upload after which listing should begin. If key-marker is not specified, the upload-id-marker parameter is ignored.

lmuMaxUploads - Sets the maximum number of multipart uploads, from 1 to 1,000, to return in the response body. 1,000 is the maximum number of uploads that can be returned in a response.

lmuDelimiter - Character you use to group keys.

lmuBucket - Undocumented member.

data ListMultipartUploads

See: listMultipartUploads smart constructor.

Request Lenses

lmuKeyMarker :: Lens' ListMultipartUploads (Maybe Text)

Together with upload-id-marker, this parameter specifies the multipart upload after which listing should begin.

lmuPrefix :: Lens' ListMultipartUploads (Maybe Text)

Lists in-progress uploads only for those keys that begin with the specified prefix.

lmuEncodingType :: Lens' ListMultipartUploads (Maybe EncodingType)

Undocumented member.

lmuUploadIdMarker :: Lens' ListMultipartUploads (Maybe Text)

Together with key-marker, specifies the multipart upload after which listing should begin. If key-marker is not specified, the upload-id-marker parameter is ignored.

lmuMaxUploads :: Lens' ListMultipartUploads (Maybe Int)

Sets the maximum number of multipart uploads, from 1 to 1,000, to return in the response body. 1,000 is the maximum number of uploads that can be returned in a response.

lmuDelimiter :: Lens' ListMultipartUploads (Maybe Delimiter)

Character you use to group keys.

lmuBucket :: Lens' ListMultipartUploads BucketName

Undocumented member.

Destructuring the Response

listMultipartUploadsResponse

Arguments

:: Int
lmursResponseStatus

-> ListMultipartUploadsResponse

Creates a value of ListMultipartUploadsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lmursKeyMarker - The key at or after which the listing began.

lmursPrefix - When a prefix is provided in the request, this field contains the specified prefix. The result contains only keys starting with the specified prefix.

lmursCommonPrefixes - Undocumented member.

lmursEncodingType - Encoding type used by Amazon S3 to encode object keys in the response.

lmursBucket - Name of the bucket to which the multipart upload was initiated.

lmursUploadIdMarker - Upload ID after which listing began.

lmursMaxUploads - Maximum number of multipart uploads that could have been included in the response.

lmursNextKeyMarker - When a list is truncated, this element specifies the value that should be used for the key-marker request parameter in a subsequent request.

lmursUploads - Undocumented member.

lmursIsTruncated - Indicates whether the returned list of multipart uploads is truncated. A value of true indicates that the list was truncated. The list can be truncated if the number of multipart uploads exceeds the limit allowed or specified by max uploads.

lmursNextUploadIdMarker - When a list is truncated, this element specifies the value that should be used for the upload-id-marker request parameter in a subsequent request.

lmursDelimiter - Undocumented member.

lmursResponseStatus - -- | The response status code.

data ListMultipartUploadsResponse

See: listMultipartUploadsResponse smart constructor.

Response Lenses

lmursKeyMarker :: Lens' ListMultipartUploadsResponse (Maybe Text)

The key at or after which the listing began.

lmursPrefix :: Lens' ListMultipartUploadsResponse (Maybe Text)

When a prefix is provided in the request, this field contains the specified prefix. The result contains only keys starting with the specified prefix.

lmursCommonPrefixes :: Lens' ListMultipartUploadsResponse [CommonPrefix]

Undocumented member.

lmursEncodingType :: Lens' ListMultipartUploadsResponse (Maybe EncodingType)

Encoding type used by Amazon S3 to encode object keys in the response.

lmursBucket :: Lens' ListMultipartUploadsResponse (Maybe BucketName)

Name of the bucket to which the multipart upload was initiated.

lmursUploadIdMarker :: Lens' ListMultipartUploadsResponse (Maybe Text)

Upload ID after which listing began.

lmursMaxUploads :: Lens' ListMultipartUploadsResponse (Maybe Int)

Maximum number of multipart uploads that could have been included in the response.

lmursNextKeyMarker :: Lens' ListMultipartUploadsResponse (Maybe Text)

When a list is truncated, this element specifies the value that should be used for the key-marker request parameter in a subsequent request.

lmursUploads :: Lens' ListMultipartUploadsResponse [MultipartUpload]

Undocumented member.

lmursIsTruncated :: Lens' ListMultipartUploadsResponse (Maybe Bool)

Indicates whether the returned list of multipart uploads is truncated. A value of true indicates that the list was truncated. The list can be truncated if the number of multipart uploads exceeds the limit allowed or specified by max uploads.

lmursNextUploadIdMarker :: Lens' ListMultipartUploadsResponse (Maybe Text)

When a list is truncated, this element specifies the value that should be used for the upload-id-marker request parameter in a subsequent request.

lmursDelimiter :: Lens' ListMultipartUploadsResponse (Maybe Delimiter)

Undocumented member.

lmursResponseStatus :: Lens' ListMultipartUploadsResponse Int

- | The response status code.