Network.AWS.S3.ListMultipartUploads

Description

This operation lists in-progress multipart uploads.

http://docs.aws.amazon.com/AmazonS3/latest/API/ListMultipartUploads.html

Request

data ListMultipartUploads

Request constructor

listMultipartUploads

Arguments

:: Text
lmuBucket

-> ListMultipartUploads

ListMultipartUploads constructor.

The fields accessible through corresponding lenses are:

lmuBucket :: Text

lmuDelimiter :: Maybe Text

lmuEncodingType :: Maybe EncodingType

lmuKeyMarker :: Maybe Text

lmuMaxUploads :: Maybe Int

lmuPrefix :: Maybe Text

lmuUploadIdMarker :: Maybe Text

Request lenses

lmuBucket :: Lens' ListMultipartUploads Text

lmuDelimiter :: Lens' ListMultipartUploads (Maybe Text)

Character you use to group keys.

lmuEncodingType :: Lens' ListMultipartUploads (Maybe EncodingType)

lmuKeyMarker :: Lens' ListMultipartUploads (Maybe Text)

Together with upload-id-marker, this parameter specifies the multipart upload after which listing should begin.

lmuMaxUploads :: Lens' ListMultipartUploads (Maybe Int)

Sets the maximum number of multipart uploads, from 1 to 1,000, to return in the response body. 1,000 is the maximum number of uploads that can be returned in a response.

lmuPrefix :: Lens' ListMultipartUploads (Maybe Text)

Lists in-progress uploads only for those keys that begin with the specified prefix.

lmuUploadIdMarker :: Lens' ListMultipartUploads (Maybe Text)

Together with key-marker, specifies the multipart upload after which listing should begin. If key-marker is not specified, the upload-id-marker parameter is ignored.

Response

data ListMultipartUploadsResponse

Response constructor

listMultipartUploadsResponse :: ListMultipartUploadsResponse

ListMultipartUploadsResponse constructor.

The fields accessible through corresponding lenses are:

lmurBucket :: Maybe Text

lmurCommonPrefixes :: [CommonPrefix]

lmurDelimiter :: Maybe Text

lmurEncodingType :: Maybe EncodingType

lmurIsTruncated :: Maybe Bool

lmurKeyMarker :: Maybe Text

lmurMaxUploads :: Maybe Int

lmurNextKeyMarker :: Maybe Text

lmurNextUploadIdMarker :: Maybe Text

lmurPrefix :: Maybe Text

lmurUploadIdMarker :: Maybe Text

lmurUploads :: [MultipartUpload]

Response lenses

lmurBucket :: Lens' ListMultipartUploadsResponse (Maybe Text)

Name of the bucket to which the multipart upload was initiated.

lmurCommonPrefixes :: Lens' ListMultipartUploadsResponse [CommonPrefix]

lmurDelimiter :: Lens' ListMultipartUploadsResponse (Maybe Text)

lmurEncodingType :: Lens' ListMultipartUploadsResponse (Maybe EncodingType)

Encoding type used by Amazon S3 to encode object keys in the response.

lmurIsTruncated :: Lens' ListMultipartUploadsResponse (Maybe Bool)

Indicates whether the returned list of multipart uploads is truncated. A value of true indicates that the list was truncated. The list can be truncated if the number of multipart uploads exceeds the limit allowed or specified by max uploads.

lmurKeyMarker :: Lens' ListMultipartUploadsResponse (Maybe Text)

The key at or after which the listing began.

lmurMaxUploads :: Lens' ListMultipartUploadsResponse (Maybe Int)

Maximum number of multipart uploads that could have been included in the response.

lmurNextKeyMarker :: Lens' ListMultipartUploadsResponse (Maybe Text)

When a list is truncated, this element specifies the value that should be used for the key-marker request parameter in a subsequent request.

lmurNextUploadIdMarker :: Lens' ListMultipartUploadsResponse (Maybe Text)

When a list is truncated, this element specifies the value that should be used for the upload-id-marker request parameter in a subsequent request.

lmurPrefix :: Lens' ListMultipartUploadsResponse (Maybe Text)

When a prefix is provided in the request, this field contains the specified prefix. The result contains only keys starting with the specified prefix.

lmurUploadIdMarker :: Lens' ListMultipartUploadsResponse (Maybe Text)

Upload ID after which listing began.

lmurUploads :: Lens' ListMultipartUploadsResponse [MultipartUpload]