Network.AWS.S3.ListObjectVersions

Description

Returns metadata about all of the versions of objects in a bucket.

This operation returns paginated results.

Creating a Request

listObjectVersions

Arguments

:: BucketName
lovBucket

-> ListObjectVersions

Creates a value of ListObjectVersions with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lovKeyMarker - Specifies the key to start with when listing objects in a bucket.

lovPrefix - Limits the response to keys that begin with the specified prefix.

lovEncodingType - Undocumented member.

lovVersionIdMarker - Specifies the object version you want to start listing from.

lovMaxKeys - Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.

lovDelimiter - A delimiter is a character you use to group keys.

lovBucket - Undocumented member.

data ListObjectVersions

See: listObjectVersions smart constructor.

Request Lenses

lovKeyMarker :: Lens' ListObjectVersions (Maybe Text)

Specifies the key to start with when listing objects in a bucket.

lovPrefix :: Lens' ListObjectVersions (Maybe Text)

Limits the response to keys that begin with the specified prefix.

lovEncodingType :: Lens' ListObjectVersions (Maybe EncodingType)

Undocumented member.

lovVersionIdMarker :: Lens' ListObjectVersions (Maybe Text)

Specifies the object version you want to start listing from.

lovMaxKeys :: Lens' ListObjectVersions (Maybe Int)

Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.

lovDelimiter :: Lens' ListObjectVersions (Maybe Delimiter)

A delimiter is a character you use to group keys.

lovBucket :: Lens' ListObjectVersions BucketName

Undocumented member.

Destructuring the Response

listObjectVersionsResponse

Arguments

:: Int
lovrsResponseStatus

-> ListObjectVersionsResponse

Creates a value of ListObjectVersionsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lovrsNextVersionIdMarker - Use this value for the next version id marker parameter in a subsequent request.

lovrsKeyMarker - Marks the last Key returned in a truncated response.

lovrsDeleteMarkers - Undocumented member.

lovrsPrefix - Undocumented member.

lovrsCommonPrefixes - Undocumented member.

lovrsEncodingType - Encoding type used by Amazon S3 to encode object keys in the response.

lovrsVersions - Undocumented member.

lovrsName - Undocumented member.

lovrsNextKeyMarker - Use this value for the key marker request parameter in a subsequent request.

lovrsVersionIdMarker - Undocumented member.

lovrsMaxKeys - Undocumented member.

lovrsIsTruncated - A flag that indicates whether or not Amazon S3 returned all of the results that satisfied the search criteria. If your results were truncated, you can make a follow-up paginated request using the NextKeyMarker and NextVersionIdMarker response parameters as a starting place in another request to return the rest of the results.

lovrsDelimiter - Undocumented member.

lovrsResponseStatus - -- | The response status code.

data ListObjectVersionsResponse

See: listObjectVersionsResponse smart constructor.

Response Lenses

lovrsNextVersionIdMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

Use this value for the next version id marker parameter in a subsequent request.

lovrsKeyMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

Marks the last Key returned in a truncated response.

lovrsDeleteMarkers :: Lens' ListObjectVersionsResponse [DeleteMarkerEntry]

Undocumented member.

lovrsPrefix :: Lens' ListObjectVersionsResponse (Maybe Text)

Undocumented member.

lovrsCommonPrefixes :: Lens' ListObjectVersionsResponse [CommonPrefix]

Undocumented member.

lovrsEncodingType :: Lens' ListObjectVersionsResponse (Maybe EncodingType)

Encoding type used by Amazon S3 to encode object keys in the response.

lovrsVersions :: Lens' ListObjectVersionsResponse [ObjectVersion]

Undocumented member.

lovrsName :: Lens' ListObjectVersionsResponse (Maybe BucketName)

Undocumented member.

lovrsNextKeyMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

Use this value for the key marker request parameter in a subsequent request.

lovrsVersionIdMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

Undocumented member.

lovrsMaxKeys :: Lens' ListObjectVersionsResponse (Maybe Int)

Undocumented member.

lovrsIsTruncated :: Lens' ListObjectVersionsResponse (Maybe Bool)

A flag that indicates whether or not Amazon S3 returned all of the results that satisfied the search criteria. If your results were truncated, you can make a follow-up paginated request using the NextKeyMarker and NextVersionIdMarker response parameters as a starting place in another request to return the rest of the results.

lovrsDelimiter :: Lens' ListObjectVersionsResponse (Maybe Delimiter)

Undocumented member.

lovrsResponseStatus :: Lens' ListObjectVersionsResponse Int

- | The response status code.