Network.AWS.S3.ListObjectVersions

Description

Returns metadata about all of the versions of objects in a bucket.

Request

data ListObjectVersions

Request constructor

listObjectVersions

Arguments

:: Text
lovrBucket

-> ListObjectVersions

ListObjectVersions constructor.

The fields accessible through corresponding lenses are:

lovrBucket :: Text

lovrDelimiter :: Maybe Text

lovrEncodingType :: Maybe Text

lovrKeyMarker :: Maybe Text

lovrMaxKeys :: Maybe Int

lovrPrefix :: Maybe Text

lovrVersionIdMarker :: Maybe Text

Request lenses

lovrBucket :: Lens' ListObjectVersions Text

lovrDelimiter :: Lens' ListObjectVersions (Maybe Text)

A delimiter is a character you use to group keys.

lovrEncodingType :: Lens' ListObjectVersions (Maybe Text)

lovrKeyMarker :: Lens' ListObjectVersions (Maybe Text)

Specifies the key to start with when listing objects in a bucket.

lovrMaxKeys :: Lens' ListObjectVersions (Maybe Int)

Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.

lovrPrefix :: Lens' ListObjectVersions (Maybe Text)

Limits the response to keys that begin with the specified prefix.

lovrVersionIdMarker :: Lens' ListObjectVersions (Maybe Text)

Specifies the object version you want to start listing from.

Response

data ListObjectVersionsOutput

Response constructor

listObjectVersionsOutput :: ListObjectVersionsOutput

ListObjectVersionsOutput constructor.

The fields accessible through corresponding lenses are:

lovoCommonPrefixes :: [CommonPrefix]

lovoDeleteMarkers :: [DeleteMarkerEntry]

lovoEncodingType :: Maybe Text

lovoIsTruncated :: Maybe Bool

lovoKeyMarker :: Maybe Text

lovoMaxKeys :: Maybe Int

lovoName :: Maybe Text

lovoNextKeyMarker :: Maybe Text

lovoNextVersionIdMarker :: Maybe Text

lovoPrefix :: Maybe Text

lovoVersionIdMarker :: Maybe Text

lovoVersions :: [ObjectVersion]

Response lenses

lovoCommonPrefixes :: Lens' ListObjectVersionsOutput [CommonPrefix]

lovoDeleteMarkers :: Lens' ListObjectVersionsOutput [DeleteMarkerEntry]

lovoEncodingType :: Lens' ListObjectVersionsOutput (Maybe Text)

Encoding type used by Amazon S3 to encode object keys in the response.

lovoIsTruncated :: Lens' ListObjectVersionsOutput (Maybe Bool)

A flag that indicates whether or not Amazon S3 returned all of the results that satisfied the search criteria. If your results were truncated, you can make a follow-up paginated request using the NextKeyMarker and NextVersionIdMarker response parameters as a starting place in another request to return the rest of the results.

lovoKeyMarker :: Lens' ListObjectVersionsOutput (Maybe Text)

Marks the last Key returned in a truncated response.

lovoMaxKeys :: Lens' ListObjectVersionsOutput (Maybe Int)

lovoName :: Lens' ListObjectVersionsOutput (Maybe Text)

lovoNextKeyMarker :: Lens' ListObjectVersionsOutput (Maybe Text)

Use this value for the key marker request parameter in a subsequent request.

lovoNextVersionIdMarker :: Lens' ListObjectVersionsOutput (Maybe Text)

Use this value for the next version id marker parameter in a subsequent request.

lovoPrefix :: Lens' ListObjectVersionsOutput (Maybe Text)

lovoVersionIdMarker :: Lens' ListObjectVersionsOutput (Maybe Text)

lovoVersions :: Lens' ListObjectVersionsOutput [ObjectVersion]