Network.AWS.S3.ListObjectVersions

Description

Returns metadata about all of the versions of objects in a bucket.

http://docs.aws.amazon.com/AmazonS3/latest/API/ListObjectVersions.html

Request

data ListObjectVersions

Request constructor

listObjectVersions

Arguments

:: Text
lovBucket

-> ListObjectVersions

ListObjectVersions constructor.

The fields accessible through corresponding lenses are:

lovBucket :: Text

lovDelimiter :: Maybe Text

lovEncodingType :: Maybe EncodingType

lovKeyMarker :: Maybe Text

lovMaxKeys :: Maybe Int

lovPrefix :: Maybe Text

lovVersionIdMarker :: Maybe Text

Request lenses

lovBucket :: Lens' ListObjectVersions Text

lovDelimiter :: Lens' ListObjectVersions (Maybe Text)

A delimiter is a character you use to group keys.

lovEncodingType :: Lens' ListObjectVersions (Maybe EncodingType)

lovKeyMarker :: Lens' ListObjectVersions (Maybe Text)

Specifies the key to start with when listing objects in a bucket.

lovMaxKeys :: Lens' ListObjectVersions (Maybe Int)

Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.

lovPrefix :: Lens' ListObjectVersions (Maybe Text)

Limits the response to keys that begin with the specified prefix.

lovVersionIdMarker :: Lens' ListObjectVersions (Maybe Text)

Specifies the object version you want to start listing from.

Response

data ListObjectVersionsResponse

Response constructor

listObjectVersionsResponse :: ListObjectVersionsResponse

ListObjectVersionsResponse constructor.

The fields accessible through corresponding lenses are:

lovrCommonPrefixes :: [CommonPrefix]

lovrDeleteMarkers :: [DeleteMarkerEntry]

lovrDelimiter :: Maybe Text

lovrEncodingType :: Maybe EncodingType

lovrIsTruncated :: Maybe Bool

lovrKeyMarker :: Maybe Text

lovrMaxKeys :: Maybe Int

lovrName :: Maybe Text

lovrNextKeyMarker :: Maybe Text

lovrNextVersionIdMarker :: Maybe Text

lovrPrefix :: Maybe Text

lovrVersionIdMarker :: Maybe Text

lovrVersions :: [ObjectVersion]

Response lenses

lovrCommonPrefixes :: Lens' ListObjectVersionsResponse [CommonPrefix]

lovrDeleteMarkers :: Lens' ListObjectVersionsResponse [DeleteMarkerEntry]

lovrDelimiter :: Lens' ListObjectVersionsResponse (Maybe Text)

lovrEncodingType :: Lens' ListObjectVersionsResponse (Maybe EncodingType)

Encoding type used by Amazon S3 to encode object keys in the response.

lovrIsTruncated :: Lens' ListObjectVersionsResponse (Maybe Bool)

A flag that indicates whether or not Amazon S3 returned all of the results that satisfied the search criteria. If your results were truncated, you can make a follow-up paginated request using the NextKeyMarker and NextVersionIdMarker response parameters as a starting place in another request to return the rest of the results.

lovrKeyMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

Marks the last Key returned in a truncated response.

lovrMaxKeys :: Lens' ListObjectVersionsResponse (Maybe Int)

lovrName :: Lens' ListObjectVersionsResponse (Maybe Text)

lovrNextKeyMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

Use this value for the key marker request parameter in a subsequent request.

lovrNextVersionIdMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

Use this value for the next version id marker parameter in a subsequent request.

lovrPrefix :: Lens' ListObjectVersionsResponse (Maybe Text)

lovrVersionIdMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

lovrVersions :: Lens' ListObjectVersionsResponse [ObjectVersion]