Network.AWS.S3.ListObjects

Description

Returns some or all (up to 1000) of the objects in a bucket. You can use the request parameters as selection criteria to return a subset of the objects in a bucket.

This operation returns paginated results.

Creating a Request

listObjects

Arguments

:: BucketName
loBucket

-> ListObjects

Creates a value of ListObjects with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

loPrefix - Limits the response to keys that begin with the specified prefix.

loEncodingType - Undocumented member.

loRequestPayer - Confirms that the requester knows that she or he will be charged for the list objects request. Bucket owners need not specify this parameter in their requests.

loMarker - Specifies the key to start with when listing objects in a bucket.

loMaxKeys - Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.

loDelimiter - A delimiter is a character you use to group keys.

loBucket - Undocumented member.

data ListObjects

See: listObjects smart constructor.

Request Lenses

loPrefix :: Lens' ListObjects (Maybe Text)

Limits the response to keys that begin with the specified prefix.

loEncodingType :: Lens' ListObjects (Maybe EncodingType)

Undocumented member.

loRequestPayer :: Lens' ListObjects (Maybe RequestPayer)

Confirms that the requester knows that she or he will be charged for the list objects request. Bucket owners need not specify this parameter in their requests.

loMarker :: Lens' ListObjects (Maybe Text)

Specifies the key to start with when listing objects in a bucket.

loMaxKeys :: Lens' ListObjects (Maybe Int)

Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.

loDelimiter :: Lens' ListObjects (Maybe Delimiter)

A delimiter is a character you use to group keys.

loBucket :: Lens' ListObjects BucketName

Undocumented member.

Destructuring the Response

listObjectsResponse

Arguments

:: Int
lorsResponseStatus

-> ListObjectsResponse

Creates a value of ListObjectsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lorsContents - Undocumented member.

lorsPrefix - Undocumented member.

lorsCommonPrefixes - Undocumented member.

lorsEncodingType - Encoding type used by Amazon S3 to encode object keys in the response.

lorsName - Undocumented member.

lorsMarker - Undocumented member.

lorsNextMarker - When response is truncated (the IsTruncated element value in the response is true), you can use the key name in this field as marker in the subsequent request to get next set of objects. Amazon S3 lists objects in alphabetical order Note: This element is returned only if you have delimiter request parameter specified. If response does not include the NextMaker and it is truncated, you can use the value of the last Key in the response as the marker in the subsequent request to get the next set of object keys.

lorsMaxKeys - Undocumented member.

lorsIsTruncated - A flag that indicates whether or not Amazon S3 returned all of the results that satisfied the search criteria.

lorsDelimiter - Undocumented member.

lorsResponseStatus - -- | The response status code.

data ListObjectsResponse

See: listObjectsResponse smart constructor.

Response Lenses

lorsContents :: Lens' ListObjectsResponse [Object]

Undocumented member.

lorsPrefix :: Lens' ListObjectsResponse (Maybe Text)

Undocumented member.

lorsCommonPrefixes :: Lens' ListObjectsResponse [CommonPrefix]

Undocumented member.

lorsEncodingType :: Lens' ListObjectsResponse (Maybe EncodingType)

Encoding type used by Amazon S3 to encode object keys in the response.

lorsName :: Lens' ListObjectsResponse (Maybe BucketName)

Undocumented member.

lorsMarker :: Lens' ListObjectsResponse (Maybe Text)

Undocumented member.

lorsNextMarker :: Lens' ListObjectsResponse (Maybe Text)

When response is truncated (the IsTruncated element value in the response is true), you can use the key name in this field as marker in the subsequent request to get next set of objects. Amazon S3 lists objects in alphabetical order Note: This element is returned only if you have delimiter request parameter specified. If response does not include the NextMaker and it is truncated, you can use the value of the last Key in the response as the marker in the subsequent request to get the next set of object keys.

lorsMaxKeys :: Lens' ListObjectsResponse (Maybe Int)

Undocumented member.

lorsIsTruncated :: Lens' ListObjectsResponse (Maybe Bool)

A flag that indicates whether or not Amazon S3 returned all of the results that satisfied the search criteria.

lorsDelimiter :: Lens' ListObjectsResponse (Maybe Delimiter)

Undocumented member.

lorsResponseStatus :: Lens' ListObjectsResponse Int

- | The response status code.