Network.AWS.S3.ListObjects

Description

Returns some or all (up to 1000) of the objects in a bucket. You can use the request parameters as selection criteria to return a subset of the objects in a bucket.

http://docs.aws.amazon.com/AmazonS3/latest/API/ListObjects.html

Request

data ListObjects

Request constructor

listObjects

Arguments

:: Text
loBucket

-> ListObjects

ListObjects constructor.

The fields accessible through corresponding lenses are:

loBucket :: Text

loDelimiter :: Maybe Text

loEncodingType :: Maybe EncodingType

loMarker :: Maybe Text

loMaxKeys :: Maybe Int

loPrefix :: Maybe Text

Request lenses

loBucket :: Lens' ListObjects Text

loDelimiter :: Lens' ListObjects (Maybe Text)

A delimiter is a character you use to group keys.

loEncodingType :: Lens' ListObjects (Maybe EncodingType)

loMarker :: Lens' ListObjects (Maybe Text)

Specifies the key to start with when listing objects in a bucket.

loMaxKeys :: Lens' ListObjects (Maybe Int)

Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.

loPrefix :: Lens' ListObjects (Maybe Text)

Limits the response to keys that begin with the specified prefix.

Response

data ListObjectsResponse

Response constructor

listObjectsResponse :: ListObjectsResponse

ListObjectsResponse constructor.

The fields accessible through corresponding lenses are:

lorCommonPrefixes :: [CommonPrefix]

lorContents :: [Object]

lorDelimiter :: Maybe Text

lorEncodingType :: Maybe EncodingType

lorIsTruncated :: Maybe Bool

lorMarker :: Maybe Text

lorMaxKeys :: Maybe Int

lorName :: Maybe Text

lorNextMarker :: Maybe Text

lorPrefix :: Maybe Text

Response lenses

lorCommonPrefixes :: Lens' ListObjectsResponse [CommonPrefix]

lorContents :: Lens' ListObjectsResponse [Object]

lorDelimiter :: Lens' ListObjectsResponse (Maybe Text)

lorEncodingType :: Lens' ListObjectsResponse (Maybe EncodingType)

Encoding type used by Amazon S3 to encode object keys in the response.

lorIsTruncated :: Lens' ListObjectsResponse (Maybe Bool)

A flag that indicates whether or not Amazon S3 returned all of the results that satisfied the search criteria.

lorMarker :: Lens' ListObjectsResponse (Maybe Text)

lorMaxKeys :: Lens' ListObjectsResponse (Maybe Int)

lorName :: Lens' ListObjectsResponse (Maybe Text)

lorNextMarker :: Lens' ListObjectsResponse (Maybe Text)

When response is truncated (the IsTruncated element value in the response is true), you can use the key name in this field as marker in the subsequent request to get next set of objects. Amazon S3 lists objects in alphabetical order Note: This element is returned only if you have delimiter request parameter specified. If response does not include the NextMaker and it is truncated, you can use the value of the last Key in the response as the marker in the subsequent request to get the next set of object keys.

lorPrefix :: Lens' ListObjectsResponse (Maybe Text)