Network.AWS.S3.ListParts

Description

Lists the parts that have been uploaded for a specific multipart upload.

This operation returns paginated results.

Creating a Request

listParts

Arguments

:: BucketName
lpBucket

-> ObjectKey
lpKey

-> Text
lpUploadId

-> ListParts

Creates a value of ListParts with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lpMaxParts - Sets the maximum number of parts to return.

lpRequestPayer - Undocumented member.

lpPartNumberMarker - Specifies the part after which listing should begin. Only parts with higher part numbers will be listed.

lpBucket - Undocumented member.

lpKey - Undocumented member.

lpUploadId - Upload ID identifying the multipart upload whose parts are being listed.

data ListParts

See: listParts smart constructor.

Request Lenses

lpMaxParts :: Lens' ListParts (Maybe Int)

Sets the maximum number of parts to return.

lpRequestPayer :: Lens' ListParts (Maybe RequestPayer)

Undocumented member.

lpPartNumberMarker :: Lens' ListParts (Maybe Int)

Specifies the part after which listing should begin. Only parts with higher part numbers will be listed.

lpBucket :: Lens' ListParts BucketName

Undocumented member.

lpKey :: Lens' ListParts ObjectKey

Undocumented member.

lpUploadId :: Lens' ListParts Text

Upload ID identifying the multipart upload whose parts are being listed.

Destructuring the Response

listPartsResponse

Arguments

:: Int
lprsResponseStatus

-> ListPartsResponse

Creates a value of ListPartsResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

lprsParts - Undocumented member.

lprsRequestCharged - Undocumented member.

lprsMaxParts - Maximum number of parts that were allowed in the response.

lprsInitiator - Identifies who initiated the multipart upload.

lprsBucket - Name of the bucket to which the multipart upload was initiated.

lprsAbortDate - Date when multipart upload will become eligible for abort operation by lifecycle.

lprsNextPartNumberMarker - When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

lprsAbortRuleId - Id of the lifecycle rule that makes a multipart upload eligible for abort operation.

lprsOwner - Undocumented member.

lprsKey - Object key for which the multipart upload was initiated.

lprsStorageClass - The class of storage used to store the object.

lprsIsTruncated - Indicates whether the returned list of parts is truncated.

lprsPartNumberMarker - Part number after which listing begins.

lprsUploadId - Upload ID identifying the multipart upload whose parts are being listed.

lprsResponseStatus - -- | The response status code.

data ListPartsResponse

See: listPartsResponse smart constructor.

Response Lenses

lprsParts :: Lens' ListPartsResponse [Part]

Undocumented member.

lprsRequestCharged :: Lens' ListPartsResponse (Maybe RequestCharged)

Undocumented member.

lprsMaxParts :: Lens' ListPartsResponse (Maybe Int)

Maximum number of parts that were allowed in the response.

lprsInitiator :: Lens' ListPartsResponse (Maybe Initiator)

Identifies who initiated the multipart upload.

lprsBucket :: Lens' ListPartsResponse (Maybe BucketName)

Name of the bucket to which the multipart upload was initiated.

lprsAbortDate :: Lens' ListPartsResponse (Maybe UTCTime)

Date when multipart upload will become eligible for abort operation by lifecycle.

lprsNextPartNumberMarker :: Lens' ListPartsResponse (Maybe Int)

When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

lprsAbortRuleId :: Lens' ListPartsResponse (Maybe Text)

Id of the lifecycle rule that makes a multipart upload eligible for abort operation.

lprsOwner :: Lens' ListPartsResponse (Maybe Owner)

Undocumented member.

lprsKey :: Lens' ListPartsResponse (Maybe ObjectKey)

Object key for which the multipart upload was initiated.

lprsStorageClass :: Lens' ListPartsResponse (Maybe StorageClass)

The class of storage used to store the object.

lprsIsTruncated :: Lens' ListPartsResponse (Maybe Bool)

Indicates whether the returned list of parts is truncated.

lprsPartNumberMarker :: Lens' ListPartsResponse (Maybe Int)

Part number after which listing begins.

lprsUploadId :: Lens' ListPartsResponse (Maybe Text)

Upload ID identifying the multipart upload whose parts are being listed.

lprsResponseStatus :: Lens' ListPartsResponse Int

- | The response status code.