Network.AWS.S3.ListParts

Description

Lists the parts that have been uploaded for a specific multipart upload.

http://docs.aws.amazon.com/AmazonS3/latest/API/ListParts.html

Request

data ListParts

Request constructor

listParts

Arguments

:: Text
lpBucket

-> Text
lpKey

-> Text
lpUploadId

-> ListParts

ListParts constructor.

The fields accessible through corresponding lenses are:

lpBucket :: Text

lpKey :: Text

lpMaxParts :: Maybe Int

lpPartNumberMarker :: Maybe Int

lpUploadId :: Text

Request lenses

lpBucket :: Lens' ListParts Text

lpKey :: Lens' ListParts Text

lpMaxParts :: Lens' ListParts (Maybe Int)

Sets the maximum number of parts to return.

lpPartNumberMarker :: Lens' ListParts (Maybe Int)

Specifies the part after which listing should begin. Only parts with higher part numbers will be listed.

lpUploadId :: Lens' ListParts Text

Upload ID identifying the multipart upload whose parts are being listed.

Response

data ListPartsResponse

Response constructor

listPartsResponse :: ListPartsResponse

ListPartsResponse constructor.

The fields accessible through corresponding lenses are:

lprBucket :: Maybe Text

lprInitiator :: Maybe Initiator

lprIsTruncated :: Maybe Bool

lprKey :: Maybe Text

lprMaxParts :: Maybe Int

lprNextPartNumberMarker :: Maybe Int

lprOwner :: Maybe Owner

lprPartNumberMarker :: Maybe Int

lprParts :: [Part]

lprStorageClass :: Maybe StorageClass

lprUploadId :: Maybe Text

Response lenses

lprBucket :: Lens' ListPartsResponse (Maybe Text)

Name of the bucket to which the multipart upload was initiated.

lprInitiator :: Lens' ListPartsResponse (Maybe Initiator)

Identifies who initiated the multipart upload.

lprIsTruncated :: Lens' ListPartsResponse (Maybe Bool)

Indicates whether the returned list of parts is truncated.

lprKey :: Lens' ListPartsResponse (Maybe Text)

Object key for which the multipart upload was initiated.

lprMaxParts :: Lens' ListPartsResponse (Maybe Int)

Maximum number of parts that were allowed in the response.

lprNextPartNumberMarker :: Lens' ListPartsResponse (Maybe Int)

When a list is truncated, this element specifies the last part in the list, as well as the value to use for the part-number-marker request parameter in a subsequent request.

lprOwner :: Lens' ListPartsResponse (Maybe Owner)

lprPartNumberMarker :: Lens' ListPartsResponse (Maybe Int)

Part number after which listing begins.

lprParts :: Lens' ListPartsResponse [Part]

lprStorageClass :: Lens' ListPartsResponse (Maybe StorageClass)

The class of storage used to store the object.

lprUploadId :: Lens' ListPartsResponse (Maybe Text)

Upload ID identifying the multipart upload whose parts are being listed.