Network.AWS.S3.PutBucketAcl

Description

Sets the permissions on a bucket using access control lists (ACL).

http://docs.aws.amazon.com/AmazonS3/latest/API/PutBucketAcl.html

Request

data PutBucketAcl

Request constructor

putBucketAcl

Arguments

:: Text
pbaBucket

-> PutBucketAcl

PutBucketAcl constructor.

The fields accessible through corresponding lenses are:

pbaACL :: Maybe BucketCannedACL

pbaAccessControlPolicy :: Maybe AccessControlPolicy

pbaBucket :: Text

pbaContentMD5 :: Maybe Text

pbaGrantFullControl :: Maybe Text

pbaGrantRead :: Maybe Text

pbaGrantReadACP :: Maybe Text

pbaGrantWrite :: Maybe Text

pbaGrantWriteACP :: Maybe Text

Request lenses

pbaACL :: Lens' PutBucketAcl (Maybe BucketCannedACL)

The canned ACL to apply to the bucket.

pbaAccessControlPolicy :: Lens' PutBucketAcl (Maybe AccessControlPolicy)

pbaBucket :: Lens' PutBucketAcl Text

pbaContentMD5 :: Lens' PutBucketAcl (Maybe Text)

pbaGrantFullControl :: Lens' PutBucketAcl (Maybe Text)

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

pbaGrantRead :: Lens' PutBucketAcl (Maybe Text)

Allows grantee to list the objects in the bucket.

pbaGrantReadACP :: Lens' PutBucketAcl (Maybe Text)

Allows grantee to read the bucket ACL.

pbaGrantWrite :: Lens' PutBucketAcl (Maybe Text)

Allows grantee to create, overwrite, and delete any object in the bucket.

pbaGrantWriteACP :: Lens' PutBucketAcl (Maybe Text)

Allows grantee to write the ACL for the applicable bucket.

Response

data PutBucketAclResponse

Response constructor

putBucketAclResponse :: PutBucketAclResponse

PutBucketAclResponse constructor.