Network.AWS.S3.PutBucketCORS

Description

Sets the cors configuration for a bucket.

Creating a Request

putBucketCORS

Arguments

:: BucketName
pbcBucket

-> CORSConfiguration
pbcCORSConfiguration

-> PutBucketCORS

Creates a value of PutBucketCORS with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pbcContentMD5 - Undocumented member.

pbcBucket - Undocumented member.

pbcCORSConfiguration - Undocumented member.

data PutBucketCORS

See: putBucketCORS smart constructor.

Request Lenses

pbcContentMD5 :: Lens' PutBucketCORS (Maybe Text)

Undocumented member.

pbcBucket :: Lens' PutBucketCORS BucketName

Undocumented member.

pbcCORSConfiguration :: Lens' PutBucketCORS CORSConfiguration

Undocumented member.

Destructuring the Response

putBucketCORSResponse :: PutBucketCORSResponse

Creates a value of PutBucketCORSResponse with the minimum fields required to make a request.

data PutBucketCORSResponse

See: putBucketCORSResponse smart constructor.