Network.AWS.S3.PutBucketInventoryConfiguration

Description

Adds an inventory configuration (identified by the inventory ID) from the bucket.

Creating a Request

putBucketInventoryConfiguration

Arguments

:: BucketName
pbicBucket

-> Text
pbicId

-> InventoryConfiguration
pbicInventoryConfiguration

-> PutBucketInventoryConfiguration

Creates a value of PutBucketInventoryConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pbicBucket - The name of the bucket where the inventory configuration will be stored.

pbicId - The ID used to identify the inventory configuration.

pbicInventoryConfiguration - Specifies the inventory configuration.

data PutBucketInventoryConfiguration

See: putBucketInventoryConfiguration smart constructor.

Request Lenses

pbicBucket :: Lens' PutBucketInventoryConfiguration BucketName

The name of the bucket where the inventory configuration will be stored.

pbicId :: Lens' PutBucketInventoryConfiguration Text

The ID used to identify the inventory configuration.

pbicInventoryConfiguration :: Lens' PutBucketInventoryConfiguration InventoryConfiguration

Specifies the inventory configuration.

Destructuring the Response

putBucketInventoryConfigurationResponse :: PutBucketInventoryConfigurationResponse

Creates a value of PutBucketInventoryConfigurationResponse with the minimum fields required to make a request.

data PutBucketInventoryConfigurationResponse

See: putBucketInventoryConfigurationResponse smart constructor.