Network.AWS.S3.PutBucketMetricsConfiguration

Description

Sets a metrics configuration (specified by the metrics configuration ID) for the bucket.

Creating a Request

putBucketMetricsConfiguration

Arguments

:: BucketName
pbmcBucket

-> Text
pbmcId

-> MetricsConfiguration
pbmcMetricsConfiguration

-> PutBucketMetricsConfiguration

Creates a value of PutBucketMetricsConfiguration with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pbmcBucket - The name of the bucket for which the metrics configuration is set.

pbmcId - The ID used to identify the metrics configuration.

pbmcMetricsConfiguration - Specifies the metrics configuration.

data PutBucketMetricsConfiguration

See: putBucketMetricsConfiguration smart constructor.

Request Lenses

pbmcBucket :: Lens' PutBucketMetricsConfiguration BucketName

The name of the bucket for which the metrics configuration is set.

pbmcId :: Lens' PutBucketMetricsConfiguration Text

The ID used to identify the metrics configuration.

pbmcMetricsConfiguration :: Lens' PutBucketMetricsConfiguration MetricsConfiguration

Specifies the metrics configuration.

Destructuring the Response

putBucketMetricsConfigurationResponse :: PutBucketMetricsConfigurationResponse

Creates a value of PutBucketMetricsConfigurationResponse with the minimum fields required to make a request.

data PutBucketMetricsConfigurationResponse

See: putBucketMetricsConfigurationResponse smart constructor.