Network.AWS.S3.PutBucketRequestPayment

Description

Sets the request payment configuration for a bucket. By default, the bucket owner pays for downloads from the bucket. This configuration parameter enables the bucket owner (only) to specify that the person requesting the download will be charged for the download. Documentation on requester pays buckets can be found at http://docs.aws.amazon.com/AmazonS3/latest/dev/RequesterPaysBuckets.html

Creating a Request

putBucketRequestPayment

Arguments

:: BucketName
pbrpBucket

-> RequestPaymentConfiguration
pbrpRequestPaymentConfiguration

-> PutBucketRequestPayment

Creates a value of PutBucketRequestPayment with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

pbrpContentMD5 - Undocumented member.

pbrpBucket - Undocumented member.

pbrpRequestPaymentConfiguration - Undocumented member.

data PutBucketRequestPayment

See: putBucketRequestPayment smart constructor.

Request Lenses

pbrpContentMD5 :: Lens' PutBucketRequestPayment (Maybe Text)

Undocumented member.

pbrpBucket :: Lens' PutBucketRequestPayment BucketName

Undocumented member.

pbrpRequestPaymentConfiguration :: Lens' PutBucketRequestPayment RequestPaymentConfiguration

Undocumented member.

Destructuring the Response

putBucketRequestPaymentResponse :: PutBucketRequestPaymentResponse

Creates a value of PutBucketRequestPaymentResponse with the minimum fields required to make a request.

data PutBucketRequestPaymentResponse

See: putBucketRequestPaymentResponse smart constructor.