Network.AWS.S3.PutObject

Description

Adds an object to a bucket.

Creating a Request

putObject

Arguments

:: BucketName
poBucket

-> ObjectKey
poKey

-> RqBody
poBody

-> PutObject

Creates a value of PutObject with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

poContentLength - Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically.

poExpires - The date and time at which the object is no longer cacheable.

poGrantReadACP - Allows grantee to read the object ACL.

poSSECustomerAlgorithm - Specifies the algorithm to use to when encrypting the object (e.g., AES256).

poSSECustomerKey - Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

poRequestPayer - Undocumented member.

poGrantWriteACP - Allows grantee to write the ACL for the applicable object.

poWebsiteRedirectLocation - If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

poGrantRead - Allows grantee to read the object data and its metadata.

poStorageClass - The type of storage to use for the object. Defaults to STANDARD.

poSSECustomerKeyMD5 - Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

poSSEKMSKeyId - Specifies the AWS KMS key ID to use for object encryption. All GET and PUT requests for an object protected by AWS KMS will fail if not made via SSL or using SigV4. Documentation on configuring any of the officially supported AWS SDKs and CLI can be found at http://docs.aws.amazon.com/AmazonS3/latest/dev/UsingAWSSDK.html#specify-signature-version

poGrantFullControl - Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

poContentEncoding - Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

poTagging - The tag-set for the object. The tag-set must be encoded as URL Query parameters

poContentMD5 - The base64-encoded 128-bit MD5 digest of the part data.

poMetadata - A map of metadata to store with the object in S3.

poCacheControl - Specifies caching behavior along the request/reply chain.

poContentLanguage - The language the content is in.

poACL - The canned ACL to apply to the object.

poContentDisposition - Specifies presentational information for the object.

poServerSideEncryption - The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

poContentType - A standard MIME type describing the format of the object data.

poBucket - Name of the bucket to which the PUT operation was initiated.

poKey - Object key for which the PUT operation was initiated.

poBody - Object data.

data PutObject

See: putObject smart constructor.

Request Lenses

poContentLength :: Lens' PutObject (Maybe Integer)

Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically.

poExpires :: Lens' PutObject (Maybe UTCTime)

The date and time at which the object is no longer cacheable.

poGrantReadACP :: Lens' PutObject (Maybe Text)

Allows grantee to read the object ACL.

poSSECustomerAlgorithm :: Lens' PutObject (Maybe Text)

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

poSSECustomerKey :: Lens' PutObject (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

poRequestPayer :: Lens' PutObject (Maybe RequestPayer)

Undocumented member.

poGrantWriteACP :: Lens' PutObject (Maybe Text)

Allows grantee to write the ACL for the applicable object.

poWebsiteRedirectLocation :: Lens' PutObject (Maybe Text)

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

poGrantRead :: Lens' PutObject (Maybe Text)

Allows grantee to read the object data and its metadata.

poStorageClass :: Lens' PutObject (Maybe StorageClass)

The type of storage to use for the object. Defaults to STANDARD.

poSSECustomerKeyMD5 :: Lens' PutObject (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

poSSEKMSKeyId :: Lens' PutObject (Maybe Text)

Specifies the AWS KMS key ID to use for object encryption. All GET and PUT requests for an object protected by AWS KMS will fail if not made via SSL or using SigV4. Documentation on configuring any of the officially supported AWS SDKs and CLI can be found at http://docs.aws.amazon.com/AmazonS3/latest/dev/UsingAWSSDK.html#specify-signature-version

poGrantFullControl :: Lens' PutObject (Maybe Text)

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

poContentEncoding :: Lens' PutObject (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

poTagging :: Lens' PutObject (Maybe Text)

The tag-set for the object. The tag-set must be encoded as URL Query parameters

poContentMD5 :: Lens' PutObject (Maybe Text)

The base64-encoded 128-bit MD5 digest of the part data.

poMetadata :: Lens' PutObject (HashMap Text Text)

A map of metadata to store with the object in S3.

poCacheControl :: Lens' PutObject (Maybe Text)

Specifies caching behavior along the request/reply chain.

poContentLanguage :: Lens' PutObject (Maybe Text)

The language the content is in.

poACL :: Lens' PutObject (Maybe ObjectCannedACL)

The canned ACL to apply to the object.

poContentDisposition :: Lens' PutObject (Maybe Text)

Specifies presentational information for the object.

poServerSideEncryption :: Lens' PutObject (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

poContentType :: Lens' PutObject (Maybe Text)

A standard MIME type describing the format of the object data.

poBucket :: Lens' PutObject BucketName

Name of the bucket to which the PUT operation was initiated.

poKey :: Lens' PutObject ObjectKey

Object key for which the PUT operation was initiated.

poBody :: Lens' PutObject RqBody

Object data.

Destructuring the Response

putObjectResponse

Arguments

:: Int
porsResponseStatus

-> PutObjectResponse

Creates a value of PutObjectResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

porsRequestCharged - Undocumented member.

porsETag - Entity tag for the uploaded object.

porsVersionId - Version of the object.

porsExpiration - If the object expiration is configured, this will contain the expiration date (expiry-date) and rule ID (rule-id). The value of rule-id is URL encoded.

porsSSECustomerAlgorithm - If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

porsSSECustomerKeyMD5 - If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

porsSSEKMSKeyId - If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

porsServerSideEncryption - The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

porsResponseStatus - -- | The response status code.

data PutObjectResponse

See: putObjectResponse smart constructor.

Response Lenses

porsRequestCharged :: Lens' PutObjectResponse (Maybe RequestCharged)

Undocumented member.

porsETag :: Lens' PutObjectResponse (Maybe ETag)

Entity tag for the uploaded object.

porsVersionId :: Lens' PutObjectResponse (Maybe ObjectVersionId)

Version of the object.

porsExpiration :: Lens' PutObjectResponse (Maybe Text)

If the object expiration is configured, this will contain the expiration date (expiry-date) and rule ID (rule-id). The value of rule-id is URL encoded.

porsSSECustomerAlgorithm :: Lens' PutObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

porsSSECustomerKeyMD5 :: Lens' PutObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

porsSSEKMSKeyId :: Lens' PutObjectResponse (Maybe Text)

If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

porsServerSideEncryption :: Lens' PutObjectResponse (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

porsResponseStatus :: Lens' PutObjectResponse Int

- | The response status code.