Network.AWS.S3.PutObject

Description

Adds an object to a bucket.

Request

data PutObject

Request constructor

putObject

Arguments

:: ObjectKey
poKey

-> RqBody
poBody

-> BucketName
poBucket

-> PutObject

Smart constructor for the minimum required parameters to construct a valid PutObject request.

The fields accessible through corresponding lenses are:

ACL :: Maybe ObjectCannedACL

Body :: RqBody

Bucket :: BucketName

CacheControl :: Maybe Text

ContentDisposition :: Maybe Text

ContentEncoding :: Maybe Text

ContentLanguage :: Maybe Text

ContentLength :: Maybe Integer

ContentMD5 :: Maybe Text

ContentType :: Maybe Text

Expires :: Maybe RFC822

GrantFullControl :: Maybe Text

GrantRead :: Maybe Text

GrantReadACP :: Maybe Text

GrantWriteACP :: Maybe Text

Key :: ObjectKey

Metadata :: Map Text Text

ServerSideEncryption :: Maybe ServerSideEncryption

StorageClass :: Maybe StorageClass

WebsiteRedirectLocation :: Maybe Text

SSECustomerAlgorithm :: Maybe Text

SSECustomerKey :: Maybe Text

SSECustomerKeyMD5 :: Maybe Text

Request lenses

poACL :: Lens' PutObject (Maybe ObjectCannedACL)

The canned ACL to apply to the object.

poBody :: Lens' PutObject RqBody

poBucket :: Lens' PutObject BucketName

poCacheControl :: Lens' PutObject (Maybe Text)

Specifies caching behavior along the request/reply chain.

poContentDisposition :: Lens' PutObject (Maybe Text)

Specifies presentational information for the object.

poContentEncoding :: Lens' PutObject (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

poContentLanguage :: Lens' PutObject (Maybe Text)

The language the content is in.

poContentLength :: Lens' PutObject (Maybe Integer)

Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically.

poContentMD5 :: Lens' PutObject (Maybe Text)

poContentType :: Lens' PutObject (Maybe Text)

A standard MIME type describing the format of the object data.

poExpires :: Lens' PutObject (Maybe RFC822)

The date and time at which the object is no longer cacheable.

poGrantFullControl :: Lens' PutObject (Maybe Text)

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

poGrantRead :: Lens' PutObject (Maybe Text)

Allows grantee to read the object data and its metadata.

poGrantReadACP :: Lens' PutObject (Maybe Text)

Allows grantee to read the object ACL.

poGrantWriteACP :: Lens' PutObject (Maybe Text)

Allows grantee to write the ACL for the applicable object.

poKey :: Lens' PutObject ObjectKey

poMetadata :: Lens' PutObject (Map Text Text)

A map of metadata to store with the object in S3.

poServerSideEncryption :: Lens' PutObject (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3.

poStorageClass :: Lens' PutObject (Maybe StorageClass)

The type of storage to use for the object. Defaults to STANDARD.

poWebsiteRedirectLocation :: Lens' PutObject (Maybe Text)

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

poSSECustomerAlgorithm :: Lens' PutObject (Maybe Text)

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

poSSECustomerKey :: Lens' PutObject (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

poSSECustomerKeyMD5 :: Lens' PutObject (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

Response

data PutObjectResponse

Response constructor

putObjectResponse :: PutObjectResponse

Smart constructor for the minimum required parameters to construct a valid PutObjectResponse response.

This constructor is provided for convenience and testing purposes.

The fields accessible through corresponding lenses are:

Expiration :: Maybe RFC822

ETag :: Maybe ETag

ServerSideEncryption :: Maybe ServerSideEncryption

VersionId :: Maybe ObjectVersionId

SSECustomerAlgorithm :: Maybe Text

SSECustomerKeyMD5 :: Maybe Text

Response lenses

porExpiration :: Lens' PutObjectResponse (Maybe RFC822)

If the object expiration is configured, this will contain the expiration date (expiry-date) and rule ID (rule-id). The value of rule-id is URL encoded.

porETag :: Lens' PutObjectResponse (Maybe ETag)

Entity tag for the uploaded object.

porServerSideEncryption :: Lens' PutObjectResponse (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3.

porVersionId :: Lens' PutObjectResponse (Maybe ObjectVersionId)

Version of the object.

porSSECustomerAlgorithm :: Lens' PutObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

porSSECustomerKeyMD5 :: Lens' PutObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.