Network.AWS.S3.PutObject

Description

Adds an object to a bucket.

http://docs.aws.amazon.com/AmazonS3/latest/API/PutObject.html

Request

data PutObject

Request constructor

putObject

Arguments

:: RqBody
poBody

-> Text
poBucket

-> Text
poKey

-> PutObject

PutObject constructor.

The fields accessible through corresponding lenses are:

poACL :: Maybe ObjectCannedACL

poBody :: RqBody

poBucket :: Text

poCacheControl :: Maybe Text

poContentDisposition :: Maybe Text

poContentEncoding :: Maybe Text

poContentLanguage :: Maybe Text

poContentLength :: Maybe Int

poContentMD5 :: Maybe Text

poContentType :: Maybe Text

poExpires :: Maybe UTCTime

poGrantFullControl :: Maybe Text

poGrantRead :: Maybe Text

poGrantReadACP :: Maybe Text

poGrantWriteACP :: Maybe Text

poKey :: Text

poMetadata :: HashMap (CI Text) Text

poRequestPayer :: Maybe RequestPayer

poSSECustomerAlgorithm :: Maybe Text

poSSECustomerKey :: Maybe Text

poSSECustomerKeyMD5 :: Maybe Text

poSSEKMSKeyId :: Maybe Text

poServerSideEncryption :: Maybe ServerSideEncryption

poStorageClass :: Maybe StorageClass

poWebsiteRedirectLocation :: Maybe Text

Request lenses

poACL :: Lens' PutObject (Maybe ObjectCannedACL)

The canned ACL to apply to the object.

poBody :: Lens' PutObject RqBody

Object data.

poBucket :: Lens' PutObject Text

poCacheControl :: Lens' PutObject (Maybe Text)

Specifies caching behavior along the request/reply chain.

poContentDisposition :: Lens' PutObject (Maybe Text)

Specifies presentational information for the object.

poContentEncoding :: Lens' PutObject (Maybe Text)

Specifies what content encodings have been applied to the object and thus what decoding mechanisms must be applied to obtain the media-type referenced by the Content-Type header field.

poContentLanguage :: Lens' PutObject (Maybe Text)

The language the content is in.

poContentLength :: Lens' PutObject (Maybe Int)

Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically.

poContentMD5 :: Lens' PutObject (Maybe Text)

poContentType :: Lens' PutObject (Maybe Text)

A standard MIME type describing the format of the object data.

poExpires :: Lens' PutObject (Maybe UTCTime)

The date and time at which the object is no longer cacheable.

poGrantFullControl :: Lens' PutObject (Maybe Text)

Gives the grantee READ, READ_ACP, and WRITE_ACP permissions on the object.

poGrantRead :: Lens' PutObject (Maybe Text)

Allows grantee to read the object data and its metadata.

poGrantReadACP :: Lens' PutObject (Maybe Text)

Allows grantee to read the object ACL.

poGrantWriteACP :: Lens' PutObject (Maybe Text)

Allows grantee to write the ACL for the applicable object.

poKey :: Lens' PutObject Text

poMetadata :: Lens' PutObject (HashMap (CI Text) Text)

A map of metadata to store with the object in S3.

poRequestPayer :: Lens' PutObject (Maybe RequestPayer)

poSSECustomerAlgorithm :: Lens' PutObject (Maybe Text)

Specifies the algorithm to use to when encrypting the object (e.g., AES256, aws:kms).

poSSECustomerKey :: Lens' PutObject (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header.

poSSECustomerKeyMD5 :: Lens' PutObject (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

poSSEKMSKeyId :: Lens' PutObject (Maybe Text)

Specifies the AWS KMS key ID to use for object encryption. All GET and PUT requests for an object protected by AWS KMS will fail if not made via SSL or using SigV4. Documentation on configuring any of the officially supported AWS SDKs and CLI can be found at http://docs.aws.amazon.com/AmazonS3/latest/dev/UsingAWSSDK.html#specify-signature-version

poServerSideEncryption :: Lens' PutObject (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

poStorageClass :: Lens' PutObject (Maybe StorageClass)

The type of storage to use for the object. Defaults to STANDARD.

poWebsiteRedirectLocation :: Lens' PutObject (Maybe Text)

If the bucket is configured as a website, redirects requests for this object to another object in the same bucket or to an external URL. Amazon S3 stores the value of this header in the object metadata.

Response

data PutObjectResponse

Response constructor

putObjectResponse :: PutObjectResponse

PutObjectResponse constructor.

The fields accessible through corresponding lenses are:

porETag :: Maybe Text

porExpiration :: Maybe Text

porRequestCharged :: Maybe RequestCharged

porSSECustomerAlgorithm :: Maybe Text

porSSECustomerKeyMD5 :: Maybe Text

porSSEKMSKeyId :: Maybe Text

porServerSideEncryption :: Maybe ServerSideEncryption

porVersionId :: Maybe Text

Response lenses

porETag :: Lens' PutObjectResponse (Maybe Text)

Entity tag for the uploaded object.

porExpiration :: Lens' PutObjectResponse (Maybe Text)

If the object expiration is configured, this will contain the expiration date (expiry-date) and rule ID (rule-id). The value of rule-id is URL encoded.

porRequestCharged :: Lens' PutObjectResponse (Maybe RequestCharged)

porSSECustomerAlgorithm :: Lens' PutObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

porSSECustomerKeyMD5 :: Lens' PutObjectResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

porSSEKMSKeyId :: Lens' PutObjectResponse (Maybe Text)

If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

porServerSideEncryption :: Lens' PutObjectResponse (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

porVersionId :: Lens' PutObjectResponse (Maybe Text)

Version of the object.