Network.AWS.S3.PutObjectAcl

Description

uses the acl subresource to set the access control list (ACL) permissions for an object that already exists in a bucket

http://docs.aws.amazon.com/AmazonS3/latest/API/PutObjectAcl.html

Request

data PutObjectAcl

Request constructor

putObjectAcl

Arguments

:: Text
poaBucket

-> Text
poaKey

-> PutObjectAcl

PutObjectAcl constructor.

The fields accessible through corresponding lenses are:

poaACL :: Maybe ObjectCannedACL

poaAccessControlPolicy :: Maybe AccessControlPolicy

poaBucket :: Text

poaContentMD5 :: Maybe Text

poaGrantFullControl :: Maybe Text

poaGrantRead :: Maybe Text

poaGrantReadACP :: Maybe Text

poaGrantWrite :: Maybe Text

poaGrantWriteACP :: Maybe Text

poaKey :: Text

poaRequestPayer :: Maybe RequestPayer

Request lenses

poaACL :: Lens' PutObjectAcl (Maybe ObjectCannedACL)

The canned ACL to apply to the object.

poaAccessControlPolicy :: Lens' PutObjectAcl (Maybe AccessControlPolicy)

poaBucket :: Lens' PutObjectAcl Text

poaContentMD5 :: Lens' PutObjectAcl (Maybe Text)

poaGrantFullControl :: Lens' PutObjectAcl (Maybe Text)

Allows grantee the read, write, read ACP, and write ACP permissions on the bucket.

poaGrantRead :: Lens' PutObjectAcl (Maybe Text)

Allows grantee to list the objects in the bucket.

poaGrantReadACP :: Lens' PutObjectAcl (Maybe Text)

Allows grantee to read the bucket ACL.

poaGrantWrite :: Lens' PutObjectAcl (Maybe Text)

Allows grantee to create, overwrite, and delete any object in the bucket.

poaGrantWriteACP :: Lens' PutObjectAcl (Maybe Text)

Allows grantee to write the ACL for the applicable bucket.

poaKey :: Lens' PutObjectAcl Text

poaRequestPayer :: Lens' PutObjectAcl (Maybe RequestPayer)

Response

data PutObjectAclResponse

Response constructor

putObjectAclResponse :: PutObjectAclResponse

PutObjectAclResponse constructor.

The fields accessible through corresponding lenses are:

poarRequestCharged :: Maybe RequestCharged

Response lenses

poarRequestCharged :: Lens' PutObjectAclResponse (Maybe RequestCharged)