Network.AWS.S3.RestoreObject

Description

Restores an archived copy of an object back into Amazon S3

Creating a Request

restoreObject

Arguments

:: BucketName
roBucket

-> ObjectKey
roKey

-> RestoreObject

Creates a value of RestoreObject with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

roVersionId - Undocumented member.

roRequestPayer - Undocumented member.

roRestoreRequest - Undocumented member.

roBucket - Undocumented member.

roKey - Undocumented member.

data RestoreObject

See: restoreObject smart constructor.

Request Lenses

roVersionId :: Lens' RestoreObject (Maybe ObjectVersionId)

Undocumented member.

roRequestPayer :: Lens' RestoreObject (Maybe RequestPayer)

Undocumented member.

roRestoreRequest :: Lens' RestoreObject (Maybe RestoreRequest)

Undocumented member.

roBucket :: Lens' RestoreObject BucketName

Undocumented member.

roKey :: Lens' RestoreObject ObjectKey

Undocumented member.

Destructuring the Response

restoreObjectResponse

Arguments

:: Int
rorsResponseStatus

-> RestoreObjectResponse

Creates a value of RestoreObjectResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

rorsRequestCharged - Undocumented member.

rorsResponseStatus - -- | The response status code.

data RestoreObjectResponse

See: restoreObjectResponse smart constructor.

Response Lenses

rorsRequestCharged :: Lens' RestoreObjectResponse (Maybe RequestCharged)

Undocumented member.

rorsResponseStatus :: Lens' RestoreObjectResponse Int

- | The response status code.