Network.AWS.S3.Types

Service

data S3

Version 2006-03-01 of the Amazon Simple Storage Service service.

Error

data RESTError :: *

XML

ns :: Text

GetBucketNotificationConfigurationRequest

data GetBucketNotificationConfigurationRequest

getBucketNotificationConfigurationRequest

Arguments

:: Text
gbncBucket

-> GetBucketNotificationConfigurationRequest

GetBucketNotificationConfigurationRequest constructor.

The fields accessible through corresponding lenses are:

gbncBucket :: Text

gbncBucket :: Lens' GetBucketNotificationConfigurationRequest Text

Name of the buket to get the notification configuration for.

ReplicationConfiguration

data ReplicationConfiguration

replicationConfiguration

Arguments

:: Text
rcRole

-> ReplicationConfiguration

ReplicationConfiguration constructor.

The fields accessible through corresponding lenses are:

rcRole :: Text

rcRules :: [ReplicationRule]

rcRole :: Lens' ReplicationConfiguration Text

Amazon Resource Name (ARN) of an IAM role for Amazon S3 to assume when replicating the objects.

rcRules :: Lens' ReplicationConfiguration [ReplicationRule]

Container for information about a particular replication rule. Replication configuration must have at least one rule and can contain up to 1,000 rules.

Destination

data Destination

destination

Arguments

:: Text
dBucket

-> Destination

Destination constructor.

The fields accessible through corresponding lenses are:

dBucket :: Text

dBucket :: Lens' Destination Text

Amazon resource name (ARN) of the bucket where you want Amazon S3 to store replicas of the object identified by the rule.

Event

data Event

Constructors

S3ObjectCreatedCompleteMultipartUpload
s3:ObjectCreated:CompleteMultipartUpload

S3ObjectCreatedCopy
s3:ObjectCreated:Copy

S3ObjectCreatedPost
s3:ObjectCreated:Post

S3ObjectCreatedPut
s3:ObjectCreated:Put

S3ReducedRedundancyLostObject
s3:ReducedRedundancyLostObject

RequestCharged

data RequestCharged

Constructors

Requester
requester

NoncurrentVersionExpiration

data NoncurrentVersionExpiration

noncurrentVersionExpiration

Arguments

:: Int
nveNoncurrentDays

-> NoncurrentVersionExpiration

NoncurrentVersionExpiration constructor.

The fields accessible through corresponding lenses are:

nveNoncurrentDays :: Int

nveNoncurrentDays :: Lens' NoncurrentVersionExpiration Int

Specifies the number of days an object is noncurrent before Amazon S3 can perform the associated action. For information about the noncurrent days calculations, see How Amazon S3 Calculates When an Object Became Noncurrent in the Amazon Simple Storage Service Developer Guide.

Transition

data Transition

transition :: Transition

Transition constructor.

The fields accessible through corresponding lenses are:

tDate :: Maybe UTCTime

tDays :: Maybe Int

tStorageClass :: Maybe TransitionStorageClass

tDate :: Lens' Transition (Maybe UTCTime)

Indicates at what date the object is to be moved or deleted. Should be in GMT ISO 8601 Format.

tDays :: Lens' Transition (Maybe Int)

Indicates the lifetime, in days, of the objects that are subject to the rule. The value must be a non-zero positive integer.

tStorageClass :: Lens' Transition (Maybe TransitionStorageClass)

The class of storage used to store the object.

DeleteMarkerEntry

data DeleteMarkerEntry

deleteMarkerEntry :: DeleteMarkerEntry

DeleteMarkerEntry constructor.

The fields accessible through corresponding lenses are:

dmeIsLatest :: Maybe Bool

dmeKey :: Maybe Text

dmeLastModified :: Maybe UTCTime

dmeOwner :: Maybe Owner

dmeVersionId :: Maybe Text

dmeIsLatest :: Lens' DeleteMarkerEntry (Maybe Bool)

Specifies whether the object is (true) or is not (false) the latest version of an object.

dmeKey :: Lens' DeleteMarkerEntry (Maybe Text)

The object key.

dmeLastModified :: Lens' DeleteMarkerEntry (Maybe UTCTime)

Date and time the object was last modified.

dmeOwner :: Lens' DeleteMarkerEntry (Maybe Owner)

dmeVersionId :: Lens' DeleteMarkerEntry (Maybe Text)

Version ID of an object.

ExpirationStatus

data ExpirationStatus

Constructors

Disabled
Disabled

Enabled
Enabled

Part

data Part

part :: Part

Part constructor.

The fields accessible through corresponding lenses are:

pETag :: Maybe Text

pLastModified :: Maybe UTCTime

pPartNumber :: Maybe Int

pSize :: Maybe Int

pETag :: Lens' Part (Maybe Text)

Entity tag returned when the part was uploaded.

pLastModified :: Lens' Part (Maybe UTCTime)

Date and time at which the part was uploaded.

pPartNumber :: Lens' Part (Maybe Int)

Part number identifying the part.

pSize :: Lens' Part (Maybe Int)

Size of the uploaded part data.

VersioningConfiguration

data VersioningConfiguration

versioningConfiguration :: VersioningConfiguration

VersioningConfiguration constructor.

The fields accessible through corresponding lenses are:

vcMFADelete :: Maybe MFADelete

vcStatus :: Maybe BucketVersioningStatus

vcMFADelete :: Lens' VersioningConfiguration (Maybe MFADelete)

Specifies whether MFA delete is enabled in the bucket versioning configuration. This element is only returned if the bucket has been configured with MFA delete. If the bucket has never been so configured, this element is not returned.

vcStatus :: Lens' VersioningConfiguration (Maybe BucketVersioningStatus)

The versioning state of the bucket.

Tag

data Tag

tag

Arguments

:: Text
tagKey

-> Text
tagValue

-> Tag

Tag constructor.

The fields accessible through corresponding lenses are:

tagKey :: Text

tagValue :: Text

tagKey :: Lens' Tag Text

Name of the tag.

tagValue :: Lens' Tag Text

Value of the tag.

ObjectStorageClass

data ObjectStorageClass

Constructors

Glacier
GLACIER

ReducedRedundancy
REDUCED_REDUNDANCY

Standard
STANDARD

MetadataDirective

data MetadataDirective

Constructors

Copy
COPY

Replace
REPLACE

ReplicationRule

data ReplicationRule

replicationRule

Arguments

:: Text
rrPrefix

-> ReplicationRuleStatus
rrStatus

-> Destination
rrDestination

-> ReplicationRule

ReplicationRule constructor.

The fields accessible through corresponding lenses are:

rrDestination :: Destination

rrID :: Maybe Text

rrPrefix :: Text

rrStatus :: ReplicationRuleStatus

rrDestination :: Lens' ReplicationRule Destination

rrID :: Lens' ReplicationRule (Maybe Text)

Unique identifier for the rule. The value cannot be longer than 255 characters.

rrPrefix :: Lens' ReplicationRule Text

Object keyname prefix identifying one or more objects to which the rule applies. Maximum prefix length can be up to 1,024 characters. Overlapping prefixes are not supported.

rrStatus :: Lens' ReplicationRule ReplicationRuleStatus

The rule is ignored if status is not Enabled.

RedirectAllRequestsTo

data RedirectAllRequestsTo

redirectAllRequestsTo

Arguments

:: Text
rartHostName

-> RedirectAllRequestsTo

RedirectAllRequestsTo constructor.

The fields accessible through corresponding lenses are:

rartHostName :: Text

rartProtocol :: Maybe Protocol

rartHostName :: Lens' RedirectAllRequestsTo Text

Name of the host where requests will be redirected.

rartProtocol :: Lens' RedirectAllRequestsTo (Maybe Protocol)

Protocol to use (http, https) when redirecting requests. The default is the protocol that is used in the original request.

RoutingRule

data RoutingRule

routingRule

Arguments

:: Redirect
rrRedirect

-> RoutingRule

RoutingRule constructor.

The fields accessible through corresponding lenses are:

rrCondition :: Maybe Condition

rrRedirect :: Redirect

rrCondition :: Lens' RoutingRule (Maybe Condition)

A container for describing a condition that must be met for the specified redirect to apply. For example, 1. If request is for pages in the /docs folder, redirect to the /documents folder. 2. If request results in HTTP error 4xx, redirect request to another host where you might process the error.

rrRedirect :: Lens' RoutingRule Redirect

Container for redirect information. You can redirect requests to another host, to another page, or with another protocol. In the event of an error, you can can specify a different error code to return.

NotificationConfiguration

data NotificationConfiguration

notificationConfiguration :: NotificationConfiguration

NotificationConfiguration constructor.

The fields accessible through corresponding lenses are:

ncLambdaFunctionConfigurations :: [LambdaFunctionConfiguration]

ncQueueConfigurations :: [QueueConfiguration]

ncTopicConfigurations :: [TopicConfiguration]

ncLambdaFunctionConfigurations :: Lens' NotificationConfiguration [LambdaFunctionConfiguration]

ncQueueConfigurations :: Lens' NotificationConfiguration [QueueConfiguration]

ncTopicConfigurations :: Lens' NotificationConfiguration [TopicConfiguration]

S3ServiceError

data S3ServiceError

s3ServiceError :: S3ServiceError

S3ServiceError constructor.

The fields accessible through corresponding lenses are:

sseCode :: Maybe Text

sseKey :: Maybe Text

sseMessage :: Maybe Text

sseVersionId :: Maybe Text

sseCode :: Lens' S3ServiceError (Maybe Text)

sseKey :: Lens' S3ServiceError (Maybe Text)

sseMessage :: Lens' S3ServiceError (Maybe Text)

sseVersionId :: Lens' S3ServiceError (Maybe Text)

ObjectCannedACL

data ObjectCannedACL

Constructors

AuthenticatedRead
authenticated-read

BucketOwnerFullControl
bucket-owner-full-control

BucketOwnerRead
bucket-owner-read

Private
private

PublicRead
public-read

PublicReadWrite
public-read-write

BucketVersioningStatus

data BucketVersioningStatus

Constructors

BVSEnabled
Enabled

BVSSuspended
Suspended

DeletedObject

data DeletedObject

deletedObject :: DeletedObject

DeletedObject constructor.

The fields accessible through corresponding lenses are:

do1DeleteMarker :: Maybe Bool

do1DeleteMarkerVersionId :: Maybe Text

do1Key :: Maybe Text

do1VersionId :: Maybe Text

do1DeleteMarker :: Lens' DeletedObject (Maybe Bool)

do1DeleteMarkerVersionId :: Lens' DeletedObject (Maybe Text)

do1Key :: Lens' DeletedObject (Maybe Text)

do1VersionId :: Lens' DeletedObject (Maybe Text)

ObjectVersionStorageClass

data ObjectVersionStorageClass

Constructors

OVSCStandard
STANDARD

CopyPartResult

data CopyPartResult

copyPartResult :: CopyPartResult

CopyPartResult constructor.

The fields accessible through corresponding lenses are:

cprETag :: Maybe Text

cprLastModified :: Maybe UTCTime

cprETag :: Lens' CopyPartResult (Maybe Text)

Entity tag of the object.

cprLastModified :: Lens' CopyPartResult (Maybe UTCTime)

Date and time at which the object was uploaded.

EncodingType

data EncodingType

Constructors

Url
url

RequestPaymentConfiguration

data RequestPaymentConfiguration

requestPaymentConfiguration

Arguments

:: Payer
rpcPayer

-> RequestPaymentConfiguration

RequestPaymentConfiguration constructor.

The fields accessible through corresponding lenses are:

rpcPayer :: Payer

rpcPayer :: Lens' RequestPaymentConfiguration Payer

Specifies who pays for the download and request fees.

CORSRule

data CORSRule

corsrule :: CORSRule

CORSRule constructor.

The fields accessible through corresponding lenses are:

corsrAllowedHeaders :: [Text]

corsrAllowedMethods :: [Text]

corsrAllowedOrigins :: [Text]

corsrExposeHeaders :: [Text]

corsrMaxAgeSeconds :: Maybe Int

corsrAllowedHeaders :: Lens' CORSRule [Text]

Specifies which headers are allowed in a pre-flight OPTIONS request.

corsrAllowedMethods :: Lens' CORSRule [Text]

Identifies HTTP methods that the domain/origin specified in the rule is allowed to execute.

corsrAllowedOrigins :: Lens' CORSRule [Text]

One or more origins you want customers to be able to access the bucket from.

corsrExposeHeaders :: Lens' CORSRule [Text]

One or more headers in the response that you want customers to be able to access from their applications (for example, from a JavaScript XMLHttpRequest object).

corsrMaxAgeSeconds :: Lens' CORSRule (Maybe Int)

The time in seconds that your browser is to cache the preflight response for the specified resource.

WebsiteConfiguration

data WebsiteConfiguration

websiteConfiguration :: WebsiteConfiguration

WebsiteConfiguration constructor.

The fields accessible through corresponding lenses are:

wcErrorDocument :: Maybe ErrorDocument

wcIndexDocument :: Maybe IndexDocument

wcRedirectAllRequestsTo :: Maybe RedirectAllRequestsTo

wcRoutingRules :: [RoutingRule]

wcErrorDocument :: Lens' WebsiteConfiguration (Maybe ErrorDocument)

wcIndexDocument :: Lens' WebsiteConfiguration (Maybe IndexDocument)

wcRedirectAllRequestsTo :: Lens' WebsiteConfiguration (Maybe RedirectAllRequestsTo)

wcRoutingRules :: Lens' WebsiteConfiguration [RoutingRule]

NoncurrentVersionTransition

data NoncurrentVersionTransition

noncurrentVersionTransition

Arguments

:: Int
nvtNoncurrentDays

-> TransitionStorageClass
nvtStorageClass

-> NoncurrentVersionTransition

NoncurrentVersionTransition constructor.

The fields accessible through corresponding lenses are:

nvtNoncurrentDays :: Int

nvtStorageClass :: TransitionStorageClass

nvtNoncurrentDays :: Lens' NoncurrentVersionTransition Int

Specifies the number of days an object is noncurrent before Amazon S3 can perform the associated action. For information about the noncurrent days calculations, see How Amazon S3 Calculates When an Object Became Noncurrent in the Amazon Simple Storage Service Developer Guide.

nvtStorageClass :: Lens' NoncurrentVersionTransition TransitionStorageClass

The class of storage used to store the object.

Initiator

data Initiator

initiator :: Initiator

Initiator constructor.

The fields accessible through corresponding lenses are:

iDisplayName :: Maybe Text

iID :: Maybe Text

iDisplayName :: Lens' Initiator (Maybe Text)

Name of the Principal.

iID :: Lens' Initiator (Maybe Text)

If the principal is an AWS account, it provides the Canonical User ID. If the principal is an IAM User, it provides a user ARN value.

ObjectIdentifier

data ObjectIdentifier

objectIdentifier

Arguments

:: Text
oiKey

-> ObjectIdentifier

ObjectIdentifier constructor.

The fields accessible through corresponding lenses are:

oiKey :: Text

oiVersionId :: Maybe Text

oiKey :: Lens' ObjectIdentifier Text

Key name of the object to delete.

oiVersionId :: Lens' ObjectIdentifier (Maybe Text)

VersionId for the specific version of the object to delete.

Bucket

data Bucket

bucket

Arguments

:: Text
bName

-> UTCTime
bCreationDate

-> Bucket

Bucket constructor.

The fields accessible through corresponding lenses are:

bCreationDate :: UTCTime

bName :: Text

bCreationDate :: Lens' Bucket UTCTime

Date the bucket was created.

bName :: Lens' Bucket Text

The name of the bucket.

Protocol

data Protocol

Constructors

Http
http

Https
https

Grant

data Grant

grant :: Grant

Grant constructor.

The fields accessible through corresponding lenses are:

gGrantee :: Maybe Grantee

gPermission :: Maybe Permission

gGrantee :: Lens' Grant (Maybe Grantee)

gPermission :: Lens' Grant (Maybe Permission)

Specifies the permission given to the grantee.

Rule

data Rule

rule

Arguments

:: Text
rPrefix

-> ExpirationStatus
rStatus

-> Rule

Rule constructor.

The fields accessible through corresponding lenses are:

rExpiration :: Maybe LifecycleExpiration

rID :: Maybe Text

rNoncurrentVersionExpiration :: Maybe NoncurrentVersionExpiration

rNoncurrentVersionTransition :: Maybe NoncurrentVersionTransition

rPrefix :: Text

rStatus :: ExpirationStatus

rTransition :: Maybe Transition

rExpiration :: Lens' Rule (Maybe LifecycleExpiration)

rID :: Lens' Rule (Maybe Text)

Unique identifier for the rule. The value cannot be longer than 255 characters.

rNoncurrentVersionExpiration :: Lens' Rule (Maybe NoncurrentVersionExpiration)

rNoncurrentVersionTransition :: Lens' Rule (Maybe NoncurrentVersionTransition)

rPrefix :: Lens' Rule Text

Prefix identifying one or more objects to which the rule applies.

rStatus :: Lens' Rule ExpirationStatus

If Enabled, the rule is currently being applied. If Disabled, the rule is not currently being applied.

rTransition :: Lens' Rule (Maybe Transition)

RequestPayer

data RequestPayer

Constructors

RPRequester
requester

TopicConfiguration

data TopicConfiguration

topicConfiguration

Arguments

:: Text
tcTopicArn

-> TopicConfiguration

TopicConfiguration constructor.

The fields accessible through corresponding lenses are:

tcEvents :: [Event]

tcId :: Maybe Text

tcTopicArn :: Text

tcEvents :: Lens' TopicConfiguration [Event]

tcId :: Lens' TopicConfiguration (Maybe Text)

tcTopicArn :: Lens' TopicConfiguration Text

Amazon SNS topic ARN to which Amazon S3 will publish a message when it detects events of specified type.

QueueConfiguration

data QueueConfiguration

queueConfiguration

Arguments

:: Text
qcQueueArn

-> QueueConfiguration

QueueConfiguration constructor.

The fields accessible through corresponding lenses are:

qcEvents :: [Event]

qcId :: Maybe Text

qcQueueArn :: Text

qcEvents :: Lens' QueueConfiguration [Event]

qcId :: Lens' QueueConfiguration (Maybe Text)

qcQueueArn :: Lens' QueueConfiguration Text

Amazon SQS queue ARN to which Amazon S3 will publish a message when it detects events of specified type.

Owner

data Owner

owner :: Owner

Owner constructor.

The fields accessible through corresponding lenses are:

oDisplayName :: Maybe Text

oID :: Maybe Text

oDisplayName :: Lens' Owner (Maybe Text)

oID :: Lens' Owner (Maybe Text)

BucketLoggingStatus

data BucketLoggingStatus

bucketLoggingStatus :: BucketLoggingStatus

BucketLoggingStatus constructor.

The fields accessible through corresponding lenses are:

blsLoggingEnabled :: Maybe LoggingEnabled

blsLoggingEnabled :: Lens' BucketLoggingStatus (Maybe LoggingEnabled)

NotificationConfigurationDeprecated

data NotificationConfigurationDeprecated

notificationConfigurationDeprecated :: NotificationConfigurationDeprecated

NotificationConfigurationDeprecated constructor.

The fields accessible through corresponding lenses are:

ncdCloudFunctionConfiguration :: Maybe CloudFunctionConfiguration

ncdQueueConfiguration :: Maybe QueueConfigurationDeprecated

ncdTopicConfiguration :: Maybe TopicConfigurationDeprecated

ncdCloudFunctionConfiguration :: Lens' NotificationConfigurationDeprecated (Maybe CloudFunctionConfiguration)

ncdQueueConfiguration :: Lens' NotificationConfigurationDeprecated (Maybe QueueConfigurationDeprecated)

ncdTopicConfiguration :: Lens' NotificationConfigurationDeprecated (Maybe TopicConfigurationDeprecated)

ErrorDocument

data ErrorDocument

errorDocument

Arguments

:: Text
edKey

-> ErrorDocument

ErrorDocument constructor.

The fields accessible through corresponding lenses are:

edKey :: Text

edKey :: Lens' ErrorDocument Text

The object key name to use when a 4XX class error occurs.

StorageClass

data StorageClass

Constructors

SCReducedRedundancy
REDUCED_REDUNDANCY

SCStandard
STANDARD

ObjectVersion

data ObjectVersion

objectVersion :: ObjectVersion

ObjectVersion constructor.

The fields accessible through corresponding lenses are:

ovETag :: Maybe Text

ovIsLatest :: Maybe Bool

ovKey :: Maybe Text

ovLastModified :: Maybe UTCTime

ovOwner :: Maybe Owner

ovSize :: Maybe Int

ovStorageClass :: Maybe ObjectVersionStorageClass

ovVersionId :: Maybe Text

ovETag :: Lens' ObjectVersion (Maybe Text)

ovIsLatest :: Lens' ObjectVersion (Maybe Bool)

Specifies whether the object is (true) or is not (false) the latest version of an object.

ovKey :: Lens' ObjectVersion (Maybe Text)

The object key.

ovLastModified :: Lens' ObjectVersion (Maybe UTCTime)

Date and time the object was last modified.

ovOwner :: Lens' ObjectVersion (Maybe Owner)

ovSize :: Lens' ObjectVersion (Maybe Int)

Size in bytes of the object.

ovStorageClass :: Lens' ObjectVersion (Maybe ObjectVersionStorageClass)

The class of storage used to store the object.

ovVersionId :: Lens' ObjectVersion (Maybe Text)

Version ID of an object.

TargetGrant

data TargetGrant

targetGrant :: TargetGrant

TargetGrant constructor.

The fields accessible through corresponding lenses are:

tgGrantee :: Maybe Grantee

tgPermission :: Maybe BucketLogsPermission

tgGrantee :: Lens' TargetGrant (Maybe Grantee)

tgPermission :: Lens' TargetGrant (Maybe BucketLogsPermission)

Logging permissions assigned to the Grantee for the bucket.

MFADeleteStatus

data MFADeleteStatus

Constructors

MFADSDisabled
Disabled

MFADSEnabled
Enabled

Payer

data Payer

Constructors

PBucketOwner
BucketOwner

PRequester
Requester

Redirect

data Redirect

redirect :: Redirect

Redirect constructor.

The fields accessible through corresponding lenses are:

rHostName :: Maybe Text

rHttpRedirectCode :: Maybe Text

rProtocol :: Maybe Protocol

rReplaceKeyPrefixWith :: Maybe Text

rReplaceKeyWith :: Maybe Text

rHostName :: Lens' Redirect (Maybe Text)

The host name to use in the redirect request.

rHttpRedirectCode :: Lens' Redirect (Maybe Text)

The HTTP redirect code to use on the response. Not required if one of the siblings is present.

rProtocol :: Lens' Redirect (Maybe Protocol)

Protocol to use (http, https) when redirecting requests. The default is the protocol that is used in the original request.

rReplaceKeyPrefixWith :: Lens' Redirect (Maybe Text)

The object key prefix to use in the redirect request. For example, to redirect requests for all pages with prefix docs (objects in the docs folder) to documents/, you can set a condition block with KeyPrefixEquals set to docs and in the Redirect set ReplaceKeyPrefixWith to documents. Not required if one of the siblings is present. Can be present only if ReplaceKeyWith is not provided.

rReplaceKeyWith :: Lens' Redirect (Maybe Text)

The specific object key to use in the redirect request. For example, redirect request to error.html. Not required if one of the sibling is present. Can be present only if ReplaceKeyPrefixWith is not provided.

BucketLogsPermission

data BucketLogsPermission

Constructors

FullControl
FULL_CONTROL

Read
READ

Write
WRITE

CompletedPart

data CompletedPart

completedPart :: CompletedPart

CompletedPart constructor.

The fields accessible through corresponding lenses are:

cpETag :: Maybe Text

cpPartNumber :: Maybe Int

cpETag :: Lens' CompletedPart (Maybe Text)

Entity tag returned when the part was uploaded.

cpPartNumber :: Lens' CompletedPart (Maybe Int)

Part number that identifies the part.

CreateBucketConfiguration

data CreateBucketConfiguration

createBucketConfiguration :: CreateBucketConfiguration

CreateBucketConfiguration constructor.

The fields accessible through corresponding lenses are:

cbcLocationConstraint :: Maybe Region

cbcLocationConstraint :: Lens' CreateBucketConfiguration (Maybe Region)

Specifies the region where the bucket will be created. If you don't specify a region, the bucket will be created in US Standard.

Tagging

data Tagging

tagging :: Tagging

Tagging constructor.

The fields accessible through corresponding lenses are:

tTagSet :: [Tag]

tTagSet :: Lens' Tagging [Tag]

LifecycleExpiration

data LifecycleExpiration

lifecycleExpiration :: LifecycleExpiration

LifecycleExpiration constructor.

The fields accessible through corresponding lenses are:

leDate :: Maybe UTCTime

leDays :: Maybe Int

leDate :: Lens' LifecycleExpiration (Maybe UTCTime)

Indicates at what date the object is to be moved or deleted. Should be in GMT ISO 8601 Format.

leDays :: Lens' LifecycleExpiration (Maybe Int)

Indicates the lifetime, in days, of the objects that are subject to the rule. The value must be a non-zero positive integer.

CORSConfiguration

data CORSConfiguration

corsconfiguration :: CORSConfiguration

CORSConfiguration constructor.

The fields accessible through corresponding lenses are:

corscCORSRules :: [CORSRule]

corscCORSRules :: Lens' CORSConfiguration [CORSRule]

Object

data Object

object'

Arguments

:: Text
oKey

-> UTCTime
oLastModified

-> Text
oETag

-> Int
oSize

-> ObjectStorageClass
oStorageClass

-> Owner
oOwner

-> Object

Object constructor.

The fields accessible through corresponding lenses are:

oETag :: Text

oKey :: Text

oLastModified :: UTCTime

oOwner :: Owner

oSize :: Int

oStorageClass :: ObjectStorageClass

oETag :: Lens' Object Text

oKey :: Lens' Object Text

oLastModified :: Lens' Object UTCTime

oOwner :: Lens' Object Owner

oSize :: Lens' Object Int

oStorageClass :: Lens' Object ObjectStorageClass

The class of storage used to store the object.

CommonPrefix

data CommonPrefix

commonPrefix :: CommonPrefix

CommonPrefix constructor.

The fields accessible through corresponding lenses are:

cpPrefix :: Maybe Text

cpPrefix :: Lens' CommonPrefix (Maybe Text)

MultipartUpload

data MultipartUpload

multipartUpload :: MultipartUpload

MultipartUpload constructor.

The fields accessible through corresponding lenses are:

muInitiated :: Maybe UTCTime

muInitiator :: Maybe Initiator

muKey :: Maybe Text

muOwner :: Maybe Owner

muStorageClass :: Maybe StorageClass

muUploadId :: Maybe Text

muInitiated :: Lens' MultipartUpload (Maybe UTCTime)

Date and time at which the multipart upload was initiated.

muInitiator :: Lens' MultipartUpload (Maybe Initiator)

Identifies who initiated the multipart upload.

muKey :: Lens' MultipartUpload (Maybe Text)

Key of the object for which the multipart upload was initiated.

muOwner :: Lens' MultipartUpload (Maybe Owner)

muStorageClass :: Lens' MultipartUpload (Maybe StorageClass)

The class of storage used to store the object.

muUploadId :: Lens' MultipartUpload (Maybe Text)

Upload ID that identifies the multipart upload.

LambdaFunctionConfiguration

data LambdaFunctionConfiguration

lambdaFunctionConfiguration

Arguments

:: Text
lfcLambdaFunctionArn

-> LambdaFunctionConfiguration

LambdaFunctionConfiguration constructor.

The fields accessible through corresponding lenses are:

lfcEvents :: [Event]

lfcId :: Maybe Text

lfcLambdaFunctionArn :: Text

lfcEvents :: Lens' LambdaFunctionConfiguration [Event]

lfcId :: Lens' LambdaFunctionConfiguration (Maybe Text)

lfcLambdaFunctionArn :: Lens' LambdaFunctionConfiguration Text

Lambda cloud function ARN that Amazon S3 can invoke when it detects events of the specified type.

QueueConfigurationDeprecated

data QueueConfigurationDeprecated

queueConfigurationDeprecated :: QueueConfigurationDeprecated

QueueConfigurationDeprecated constructor.

The fields accessible through corresponding lenses are:

qcdEvent :: Maybe Event

qcdEvents :: [Event]

qcdId :: Maybe Text

qcdQueue :: Maybe Text

qcdEvent :: Lens' QueueConfigurationDeprecated (Maybe Event)

qcdEvents :: Lens' QueueConfigurationDeprecated [Event]

qcdId :: Lens' QueueConfigurationDeprecated (Maybe Text)

qcdQueue :: Lens' QueueConfigurationDeprecated (Maybe Text)

TopicConfigurationDeprecated

data TopicConfigurationDeprecated

topicConfigurationDeprecated :: TopicConfigurationDeprecated

TopicConfigurationDeprecated constructor.

The fields accessible through corresponding lenses are:

tcdEvent :: Maybe Event

tcdEvents :: [Event]

tcdId :: Maybe Text

tcdTopic :: Maybe Text

tcdEvent :: Lens' TopicConfigurationDeprecated (Maybe Event)

Bucket event for which to send notifications.

tcdEvents :: Lens' TopicConfigurationDeprecated [Event]

tcdId :: Lens' TopicConfigurationDeprecated (Maybe Text)

tcdTopic :: Lens' TopicConfigurationDeprecated (Maybe Text)

Amazon SNS topic to which Amazon S3 will publish a message to report the specified events for the bucket.

Type

data Type

Constructors

AmazonCustomerByEmail
AmazonCustomerByEmail

CanonicalUser
CanonicalUser

Group
Group

ReplicationStatus

data ReplicationStatus

Constructors

Complete
COMPLETE

Failed
FAILED

Pending
PENDING

Replica
REPLICA

TransitionStorageClass

data TransitionStorageClass

Constructors

TSCGlacier
GLACIER

CompletedMultipartUpload

data CompletedMultipartUpload

completedMultipartUpload :: CompletedMultipartUpload

CompletedMultipartUpload constructor.

The fields accessible through corresponding lenses are:

cmuParts :: [CompletedPart]

cmuParts :: Lens' CompletedMultipartUpload [CompletedPart]

Condition

data Condition

condition :: Condition

Condition constructor.

The fields accessible through corresponding lenses are:

cHttpErrorCodeReturnedEquals :: Maybe Text

cKeyPrefixEquals :: Maybe Text

cHttpErrorCodeReturnedEquals :: Lens' Condition (Maybe Text)

The HTTP error code when the redirect is applied. In the event of an error, if the error code equals this value, then the specified redirect is applied. Required when parent element Condition is specified and sibling KeyPrefixEquals is not specified. If both are specified, then both must be true for the redirect to be applied.

cKeyPrefixEquals :: Lens' Condition (Maybe Text)

The object key name prefix when the redirect is applied. For example, to redirect requests for ExamplePage.html, the key prefix will be ExamplePage.html. To redirect request for all pages with the prefix docs/, the key prefix will be docs, which identifies all objects in the docs folder. Required when the parent element Condition is specified and sibling HttpErrorCodeReturnedEquals is not specified. If both conditions are specified, both must be true for the redirect to be applied.

Permission

data Permission

Constructors

PFullControl
FULL_CONTROL

PRead
READ

PReadAcp
READ_ACP

PWrite
WRITE

PWriteAcp
WRITE_ACP

AccessControlPolicy

data AccessControlPolicy

accessControlPolicy :: AccessControlPolicy

AccessControlPolicy constructor.

The fields accessible through corresponding lenses are:

acpGrants :: [Grant]

acpOwner :: Maybe Owner

acpGrants :: Lens' AccessControlPolicy [Grant]

A list of grants.

acpOwner :: Lens' AccessControlPolicy (Maybe Owner)

BucketCannedACL

data BucketCannedACL

Constructors

CannedAuthenticatedRead
authenticated-read

CannedPrivate
private

CannedPublicRead
public-read

CannedPublicReadWrite
public-read-write

MFADelete

data MFADelete

Constructors

MFADDisabled
Disabled

MFADEnabled
Enabled

CloudFunctionConfiguration

data CloudFunctionConfiguration

cloudFunctionConfiguration :: CloudFunctionConfiguration

CloudFunctionConfiguration constructor.

The fields accessible through corresponding lenses are:

cfcCloudFunction :: Maybe Text

cfcEvent :: Maybe Event

cfcEvents :: [Event]

cfcId :: Maybe Text

cfcInvocationRole :: Maybe Text

cfcCloudFunction :: Lens' CloudFunctionConfiguration (Maybe Text)

cfcEvent :: Lens' CloudFunctionConfiguration (Maybe Event)

cfcEvents :: Lens' CloudFunctionConfiguration [Event]

cfcId :: Lens' CloudFunctionConfiguration (Maybe Text)

cfcInvocationRole :: Lens' CloudFunctionConfiguration (Maybe Text)

Grantee

data Grantee

grantee

Arguments

:: Type
gType

-> Grantee

Grantee constructor.

The fields accessible through corresponding lenses are:

gDisplayName :: Maybe Text

gEmailAddress :: Maybe Text

gID :: Maybe Text

gType :: Type

gURI :: Maybe Text

gDisplayName :: Lens' Grantee (Maybe Text)

Screen name of the grantee.

gEmailAddress :: Lens' Grantee (Maybe Text)

Email address of the grantee.

gID :: Lens' Grantee (Maybe Text)

The canonical user ID of the grantee.

gType :: Lens' Grantee Type

Type of grantee

gURI :: Lens' Grantee (Maybe Text)

URI of the grantee group.

LifecycleConfiguration

data LifecycleConfiguration

lifecycleConfiguration :: LifecycleConfiguration

LifecycleConfiguration constructor.

The fields accessible through corresponding lenses are:

lcRules :: [Rule]

lcRules :: Lens' LifecycleConfiguration [Rule]

LoggingEnabled

data LoggingEnabled

loggingEnabled :: LoggingEnabled

LoggingEnabled constructor.

The fields accessible through corresponding lenses are:

leTargetBucket :: Maybe Text

leTargetGrants :: [TargetGrant]

leTargetPrefix :: Maybe Text

leTargetBucket :: Lens' LoggingEnabled (Maybe Text)

Specifies the bucket where you want Amazon S3 to store server access logs. You can have your logs delivered to any bucket that you own, including the same bucket that is being logged. You can also configure multiple buckets to deliver their logs to the same target bucket. In this case you should choose a different TargetPrefix for each source bucket so that the delivered log files can be distinguished by key.

leTargetGrants :: Lens' LoggingEnabled [TargetGrant]

leTargetPrefix :: Lens' LoggingEnabled (Maybe Text)

This element lets you specify a prefix for the keys that the log files will be stored under.

ReplicationRuleStatus

data ReplicationRuleStatus

Constructors

RRSDisabled
Disabled

RRSEnabled
Enabled

ServerSideEncryption

data ServerSideEncryption

Constructors

AES256
AES256

IndexDocument

data IndexDocument

indexDocument

Arguments

:: Text
idSuffix

-> IndexDocument

IndexDocument constructor.

The fields accessible through corresponding lenses are:

idSuffix :: Text

idSuffix :: Lens' IndexDocument Text

A suffix that is appended to a request that is for a directory on the website endpoint (e.g. if the suffix is index.html and you make a request to samplebucketimages the data that is returned will be for the object with the key name images/index.html) The suffix must not be empty and must not include a slash character.

CopyObjectResult

data CopyObjectResult

copyObjectResult :: CopyObjectResult

CopyObjectResult constructor.

The fields accessible through corresponding lenses are:

corETag :: Maybe Text

corLastModified :: Maybe UTCTime

corETag :: Lens' CopyObjectResult (Maybe Text)

corLastModified :: Lens' CopyObjectResult (Maybe UTCTime)

Delete

data Delete

delete' :: Delete

Delete constructor.

The fields accessible through corresponding lenses are:

dObjects :: [ObjectIdentifier]

dQuiet :: Maybe Bool

dObjects :: Lens' Delete [ObjectIdentifier]

dQuiet :: Lens' Delete (Maybe Bool)

Element to enable quiet mode for the request. When you add this element, you must set its value to true.

RestoreRequest

data RestoreRequest

restoreRequest

Arguments

:: Int
rDays

-> RestoreRequest

RestoreRequest constructor.

The fields accessible through corresponding lenses are:

rDays :: Int

rDays :: Lens' RestoreRequest Int

Lifetime of the active copy in days

Common

newtype BucketName

Constructors

BucketName Text

newtype ObjectKey

Constructors

ObjectKey Text

newtype ObjectVersionId

Constructors

ObjectVersionId Text

newtype ETag

Constructors

ETag Text

data Region :: *