Network.AWS.S3.UploadPart

Description

Uploads a part in a multipart upload.

Note: After you initiate multipart upload and upload one or more parts, you must either complete or abort multipart upload in order to stop getting charged for storage of the uploaded parts. Only after you either complete or abort multipart upload, Amazon S3 frees up the parts storage and stops charging you for the parts storage.

Creating a Request

uploadPart

Arguments

:: BucketName
upBucket

-> ObjectKey
upKey

-> Int
upPartNumber

-> Text
upUploadId

-> RqBody
upBody

-> UploadPart

Creates a value of UploadPart with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

upContentLength - Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically.

upSSECustomerAlgorithm - Specifies the algorithm to use to when encrypting the object (e.g., AES256).

upSSECustomerKey - Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

upRequestPayer - Undocumented member.

upSSECustomerKeyMD5 - Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

upContentMD5 - The base64-encoded 128-bit MD5 digest of the part data.

upBucket - Name of the bucket to which the multipart upload was initiated.

upKey - Object key for which the multipart upload was initiated.

upPartNumber - Part number of part being uploaded. This is a positive integer between 1 and 10,000.

upUploadId - Upload ID identifying the multipart upload whose part is being uploaded.

upBody - Object data.

data UploadPart

See: uploadPart smart constructor.

Request Lenses

upContentLength :: Lens' UploadPart (Maybe Integer)

Size of the body in bytes. This parameter is useful when the size of the body cannot be determined automatically.

upSSECustomerAlgorithm :: Lens' UploadPart (Maybe Text)

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

upSSECustomerKey :: Lens' UploadPart (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

upRequestPayer :: Lens' UploadPart (Maybe RequestPayer)

Undocumented member.

upSSECustomerKeyMD5 :: Lens' UploadPart (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

upContentMD5 :: Lens' UploadPart (Maybe Text)

The base64-encoded 128-bit MD5 digest of the part data.

upBucket :: Lens' UploadPart BucketName

Name of the bucket to which the multipart upload was initiated.

upKey :: Lens' UploadPart ObjectKey

Object key for which the multipart upload was initiated.

upPartNumber :: Lens' UploadPart Int

Part number of part being uploaded. This is a positive integer between 1 and 10,000.

upUploadId :: Lens' UploadPart Text

Upload ID identifying the multipart upload whose part is being uploaded.

upBody :: Lens' UploadPart RqBody

Object data.

Destructuring the Response

uploadPartResponse

Arguments

:: Int
uprsResponseStatus

-> UploadPartResponse

Creates a value of UploadPartResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

uprsRequestCharged - Undocumented member.

uprsETag - Entity tag for the uploaded object.

uprsSSECustomerAlgorithm - If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

uprsSSECustomerKeyMD5 - If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

uprsSSEKMSKeyId - If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

uprsServerSideEncryption - The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

uprsResponseStatus - -- | The response status code.

data UploadPartResponse

See: uploadPartResponse smart constructor.

Response Lenses

uprsRequestCharged :: Lens' UploadPartResponse (Maybe RequestCharged)

Undocumented member.

uprsETag :: Lens' UploadPartResponse (Maybe ETag)

Entity tag for the uploaded object.

uprsSSECustomerAlgorithm :: Lens' UploadPartResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

uprsSSECustomerKeyMD5 :: Lens' UploadPartResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

uprsSSEKMSKeyId :: Lens' UploadPartResponse (Maybe Text)

If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

uprsServerSideEncryption :: Lens' UploadPartResponse (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

uprsResponseStatus :: Lens' UploadPartResponse Int

- | The response status code.