Network.AWS.S3.UploadPartCopy

Description

Uploads a part by copying data from an existing object as data source.

Creating a Request

uploadPartCopy

Arguments

:: BucketName
upcBucket

-> Text
upcCopySource

-> ObjectKey
upcKey

-> Int
upcPartNumber

-> Text
upcUploadId

-> UploadPartCopy

Creates a value of UploadPartCopy with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

upcCopySourceIfModifiedSince - Copies the object if it has been modified since the specified time.

upcCopySourceIfUnmodifiedSince - Copies the object if it hasn't been modified since the specified time.

upcCopySourceRange - The range of bytes to copy from the source object. The range value must use the form bytes=first-last, where the first and last are the zero-based byte offsets to copy. For example, bytes=0-9 indicates that you want to copy the first ten bytes of the source. You can copy a range only if the source object is greater than 5 GB.

upcCopySourceSSECustomerKeyMD5 - Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

upcCopySourceIfNoneMatch - Copies the object if its entity tag (ETag) is different than the specified ETag.

upcSSECustomerAlgorithm - Specifies the algorithm to use to when encrypting the object (e.g., AES256).

upcSSECustomerKey - Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

upcRequestPayer - Undocumented member.

upcCopySourceIfMatch - Copies the object if its entity tag (ETag) matches the specified tag.

upcSSECustomerKeyMD5 - Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

upcCopySourceSSECustomerKey - Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

upcCopySourceSSECustomerAlgorithm - Specifies the algorithm to use when decrypting the source object (e.g., AES256).

upcBucket - Undocumented member.

upcCopySource - The name of the source bucket and key name of the source object, separated by a slash (/). Must be URL-encoded.

upcKey - Undocumented member.

upcPartNumber - Part number of part being copied. This is a positive integer between 1 and 10,000.

upcUploadId - Upload ID identifying the multipart upload whose part is being copied.

data UploadPartCopy

See: uploadPartCopy smart constructor.

Request Lenses

upcCopySourceIfModifiedSince :: Lens' UploadPartCopy (Maybe UTCTime)

Copies the object if it has been modified since the specified time.

upcCopySourceIfUnmodifiedSince :: Lens' UploadPartCopy (Maybe UTCTime)

Copies the object if it hasn't been modified since the specified time.

upcCopySourceRange :: Lens' UploadPartCopy (Maybe Text)

The range of bytes to copy from the source object. The range value must use the form bytes=first-last, where the first and last are the zero-based byte offsets to copy. For example, bytes=0-9 indicates that you want to copy the first ten bytes of the source. You can copy a range only if the source object is greater than 5 GB.

upcCopySourceSSECustomerKeyMD5 :: Lens' UploadPartCopy (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

upcCopySourceIfNoneMatch :: Lens' UploadPartCopy (Maybe Text)

Copies the object if its entity tag (ETag) is different than the specified ETag.

upcSSECustomerAlgorithm :: Lens' UploadPartCopy (Maybe Text)

Specifies the algorithm to use to when encrypting the object (e.g., AES256).

upcSSECustomerKey :: Lens' UploadPartCopy (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

upcRequestPayer :: Lens' UploadPartCopy (Maybe RequestPayer)

Undocumented member.

upcCopySourceIfMatch :: Lens' UploadPartCopy (Maybe Text)

Copies the object if its entity tag (ETag) matches the specified tag.

upcSSECustomerKeyMD5 :: Lens' UploadPartCopy (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

upcCopySourceSSECustomerKey :: Lens' UploadPartCopy (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

upcCopySourceSSECustomerAlgorithm :: Lens' UploadPartCopy (Maybe Text)

Specifies the algorithm to use when decrypting the source object (e.g., AES256).

upcBucket :: Lens' UploadPartCopy BucketName

Undocumented member.

upcCopySource :: Lens' UploadPartCopy Text

The name of the source bucket and key name of the source object, separated by a slash (/). Must be URL-encoded.

upcKey :: Lens' UploadPartCopy ObjectKey

Undocumented member.

upcPartNumber :: Lens' UploadPartCopy Int

Part number of part being copied. This is a positive integer between 1 and 10,000.

upcUploadId :: Lens' UploadPartCopy Text

Upload ID identifying the multipart upload whose part is being copied.

Destructuring the Response

uploadPartCopyResponse

Arguments

:: Int
upcrsResponseStatus

-> UploadPartCopyResponse

Creates a value of UploadPartCopyResponse with the minimum fields required to make a request.

Use one of the following lenses to modify other fields as desired:

upcrsRequestCharged - Undocumented member.

upcrsCopyPartResult - Undocumented member.

upcrsSSECustomerAlgorithm - If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

upcrsCopySourceVersionId - The version of the source object that was copied, if you have enabled versioning on the source bucket.

upcrsSSECustomerKeyMD5 - If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

upcrsSSEKMSKeyId - If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

upcrsServerSideEncryption - The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

upcrsResponseStatus - -- | The response status code.

data UploadPartCopyResponse

See: uploadPartCopyResponse smart constructor.

Response Lenses

upcrsRequestCharged :: Lens' UploadPartCopyResponse (Maybe RequestCharged)

Undocumented member.

upcrsCopyPartResult :: Lens' UploadPartCopyResponse (Maybe CopyPartResult)

Undocumented member.

upcrsSSECustomerAlgorithm :: Lens' UploadPartCopyResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

upcrsCopySourceVersionId :: Lens' UploadPartCopyResponse (Maybe Text)

The version of the source object that was copied, if you have enabled versioning on the source bucket.

upcrsSSECustomerKeyMD5 :: Lens' UploadPartCopyResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

upcrsSSEKMSKeyId :: Lens' UploadPartCopyResponse (Maybe Text)

If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

upcrsServerSideEncryption :: Lens' UploadPartCopyResponse (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).

upcrsResponseStatus :: Lens' UploadPartCopyResponse Int

- | The response status code.