Network.AWS.S3.UploadPartCopy

Description

Uploads a part by copying data from an existing object as data source.

http://docs.aws.amazon.com/AmazonS3/latest/API/UploadPartCopy.html

Request

data UploadPartCopy

Request constructor

uploadPartCopy

Arguments

:: Text
upcBucket

-> Text
upcCopySource

-> Text
upcKey

-> Int
upcPartNumber

-> Text
upcUploadId

-> UploadPartCopy

UploadPartCopy constructor.

The fields accessible through corresponding lenses are:

upcBucket :: Text

upcCopySource :: Text

upcCopySourceIfMatch :: Maybe Text

upcCopySourceIfModifiedSince :: Maybe UTCTime

upcCopySourceIfNoneMatch :: Maybe Text

upcCopySourceIfUnmodifiedSince :: Maybe UTCTime

upcCopySourceRange :: Maybe Text

upcCopySourceSSECustomerAlgorithm :: Maybe Text

upcCopySourceSSECustomerKey :: Maybe Text

upcCopySourceSSECustomerKeyMD5 :: Maybe Text

upcKey :: Text

upcPartNumber :: Int

upcRequestPayer :: Maybe RequestPayer

upcSSECustomerAlgorithm :: Maybe Text

upcSSECustomerKey :: Maybe Text

upcSSECustomerKeyMD5 :: Maybe Text

upcUploadId :: Text

Request lenses

upcBucket :: Lens' UploadPartCopy Text

upcCopySource :: Lens' UploadPartCopy Text

The name of the source bucket and key name of the source object, separated by a slash (/). Must be URL-encoded.

upcCopySourceIfMatch :: Lens' UploadPartCopy (Maybe Text)

Copies the object if its entity tag (ETag) matches the specified tag.

upcCopySourceIfModifiedSince :: Lens' UploadPartCopy (Maybe UTCTime)

Copies the object if it has been modified since the specified time.

upcCopySourceIfNoneMatch :: Lens' UploadPartCopy (Maybe Text)

Copies the object if its entity tag (ETag) is different than the specified ETag.

upcCopySourceIfUnmodifiedSince :: Lens' UploadPartCopy (Maybe UTCTime)

Copies the object if it hasn't been modified since the specified time.

upcCopySourceRange :: Lens' UploadPartCopy (Maybe Text)

The range of bytes to copy from the source object. The range value must use the form bytes=first-last, where the first and last are the zero-based byte offsets to copy. For example, bytes=0-9 indicates that you want to copy the first ten bytes of the source. You can copy a range only if the source object is greater than 5 GB.

upcCopySourceSSECustomerAlgorithm :: Lens' UploadPartCopy (Maybe Text)

Specifies the algorithm to use when decrypting the source object (e.g., AES256).

upcCopySourceSSECustomerKey :: Lens' UploadPartCopy (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use to decrypt the source object. The encryption key provided in this header must be one that was used when the source object was created.

upcCopySourceSSECustomerKeyMD5 :: Lens' UploadPartCopy (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

upcKey :: Lens' UploadPartCopy Text

upcPartNumber :: Lens' UploadPartCopy Int

Part number of part being copied.

upcRequestPayer :: Lens' UploadPartCopy (Maybe RequestPayer)

upcSSECustomerAlgorithm :: Lens' UploadPartCopy (Maybe Text)

Specifies the algorithm to use to when encrypting the object (e.g., AES256, aws:kms).

upcSSECustomerKey :: Lens' UploadPartCopy (Maybe Text)

Specifies the customer-provided encryption key for Amazon S3 to use in encrypting data. This value is used to store the object and then it is discarded; Amazon does not store the encryption key. The key must be appropriate for use with the algorithm specified in the x-amz-server-side​-encryption​-customer-algorithm header. This must be the same encryption key specified in the initiate multipart upload request.

upcSSECustomerKeyMD5 :: Lens' UploadPartCopy (Maybe Text)

Specifies the 128-bit MD5 digest of the encryption key according to RFC 1321. Amazon S3 uses this header for a message integrity check to ensure the encryption key was transmitted without error.

upcUploadId :: Lens' UploadPartCopy Text

Upload ID identifying the multipart upload whose part is being copied.

Response

data UploadPartCopyResponse

Response constructor

uploadPartCopyResponse :: UploadPartCopyResponse

UploadPartCopyResponse constructor.

The fields accessible through corresponding lenses are:

upcrCopyPartResult :: Maybe CopyPartResult

upcrCopySourceVersionId :: Maybe Text

upcrRequestCharged :: Maybe RequestCharged

upcrSSECustomerAlgorithm :: Maybe Text

upcrSSECustomerKeyMD5 :: Maybe Text

upcrSSEKMSKeyId :: Maybe Text

upcrServerSideEncryption :: Maybe ServerSideEncryption

Response lenses

upcrCopyPartResult :: Lens' UploadPartCopyResponse (Maybe CopyPartResult)

upcrCopySourceVersionId :: Lens' UploadPartCopyResponse (Maybe Text)

The version of the source object that was copied, if you have enabled versioning on the source bucket.

upcrRequestCharged :: Lens' UploadPartCopyResponse (Maybe RequestCharged)

upcrSSECustomerAlgorithm :: Lens' UploadPartCopyResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header confirming the encryption algorithm used.

upcrSSECustomerKeyMD5 :: Lens' UploadPartCopyResponse (Maybe Text)

If server-side encryption with a customer-provided encryption key was requested, the response will include this header to provide round trip message integrity verification of the customer-provided encryption key.

upcrSSEKMSKeyId :: Lens' UploadPartCopyResponse (Maybe Text)

If present, specifies the ID of the AWS Key Management Service (KMS) master encryption key that was used for the object.

upcrServerSideEncryption :: Lens' UploadPartCopyResponse (Maybe ServerSideEncryption)

The Server-side encryption algorithm used when storing this object in S3 (e.g., AES256, aws:kms).