Network.AWS.S3.V2006_03_01.ListObjectVersions

Description

Returns metadata about all of the versions of objects in a bucket.

Documentation

type GetBucketObjectVersions = ListObjectVersions

listObjectVersions

Arguments

:: BucketName
_lovrBucket

-> ListObjectVersions

Minimum specification for a ListObjectVersions request.

data ListObjectVersions

Constructors

ListObjectVersions

Fields

_lovrBucket :: BucketName

_lovrDelimiter :: Maybe Char

A delimiter is a character you use to group keys.

_lovrEncodingType :: Maybe EncodingType

Requests Amazon S3 to encode the object keys in the response and specifies the encoding method to use. An object key may contain any Unicode character; however, XML 1.0 parser cannot parse some characters, such as characters with an ASCII value from 0 to 10. For characters that are not supported in XML 1.0, you can add this parameter to request that Amazon S3 encode the keys in the response.

_lovrKeyMarker :: Maybe Text

Specifies the key to start with when listing objects in a bucket.

_lovrMaxKeys :: Maybe Integer

Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.

_lovrPrefix :: Maybe Text

Limits the response to keys that begin with the specified prefix.

_lovrVersionIdMarker :: Maybe Text

Specifies the object version you want to start listing from.

lovrVersionIdMarker :: Lens' ListObjectVersions (Maybe Text)

lovrPrefix :: Lens' ListObjectVersions (Maybe Text)

lovrMaxKeys :: Lens' ListObjectVersions (Maybe Integer)

lovrKeyMarker :: Lens' ListObjectVersions (Maybe Text)

lovrEncodingType :: Lens' ListObjectVersions (Maybe EncodingType)

lovrDelimiter :: Lens' ListObjectVersions (Maybe Char)

lovrBucket :: Lens' ListObjectVersions BucketName

data ListObjectVersionsResponse

Constructors

ListObjectVersionsResponse

Fields

_lovoIsTruncated :: Bool

A flag that indicates whether or not Amazon S3 returned all of the results that satisfied the search criteria. If your results were truncated, you can make a follow-up paginated request using the NextKeyMarker and NextVersionIdMarker response parameters as a starting place in another request to return the rest of the results.

_lovoName :: Maybe BucketName

_lovoCommonPrefixes :: [CommonPrefix]

_lovoDeleteMarkers :: [DeleteMarkerEntry]

_lovoEncodingType :: Maybe EncodingType

Encoding type used by Amazon S3 to encode object keys in the response.

_lovoKeyMarker :: Maybe Text

Marks the last Key returned in a truncated response.

_lovoMaxKeys :: Maybe Integer

_lovoNextKeyMarker :: Maybe Text

Use this value for the key marker request parameter in a subsequent request.

_lovoNextVersionIdMarker :: Maybe Text

Use this value for the next version id marker parameter in a subsequent request.

_lovoVersions :: [ObjectVersion]

_lovoPrefix :: Maybe Text

_lovoVersionIdMarker :: Maybe Text

lovoVersions :: Lens' ListObjectVersionsResponse [ObjectVersion]

lovoVersionIdMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

lovoPrefix :: Lens' ListObjectVersionsResponse (Maybe Text)

lovoNextVersionIdMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

lovoNextKeyMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

lovoName :: Lens' ListObjectVersionsResponse (Maybe BucketName)

lovoMaxKeys :: Lens' ListObjectVersionsResponse (Maybe Integer)

lovoKeyMarker :: Lens' ListObjectVersionsResponse (Maybe Text)

lovoIsTruncated :: Lens' ListObjectVersionsResponse Bool

lovoEncodingType :: Lens' ListObjectVersionsResponse (Maybe EncodingType)

lovoDeleteMarkers :: Lens' ListObjectVersionsResponse [DeleteMarkerEntry]

lovoCommonPrefixes :: Lens' ListObjectVersionsResponse [CommonPrefix]