Network.AWS.S3.V2006_03_01.ListObjects

Description

Returns some or all (up to 1000) of the objects in a bucket. You can use the request parameters as selection criteria to return a subset of the objects in a bucket.

Request

data ListObjects

Request alias

type GetBucket = ListObjects

Request constructor

mkListObjects

Arguments

:: BucketName
loBucket

-> ListObjects

Smart constructor for the minimum required parameters to construct a valid ListObjects request.

Request lenses

loBucket :: Lens' ListObjects BucketName

loDelimiter :: Lens' ListObjects (Maybe Char)

A delimiter is a character you use to group keys.

loEncodingType :: Lens' ListObjects (Maybe EncodingType)

Requests Amazon S3 to encode the object keys in the response and specifies the encoding method to use. An object key may contain any Unicode character; however, XML 1.0 parser cannot parse some characters, such as characters with an ASCII value from 0 to 10. For characters that are not supported in XML 1.0, you can add this parameter to request that Amazon S3 encode the keys in the response.

loMarker :: Lens' ListObjects (Maybe Text)

Specifies the key to start with when listing objects in a bucket.

loMaxKeys :: Lens' ListObjects (Maybe Integer)

Sets the maximum number of keys returned in the response. The response might contain fewer keys but will never contain more.

loPrefix :: Lens' ListObjects (Maybe Text)

Limits the response to keys that begin with the specified prefix.

Response

data ListObjectsResponse

Response constructor

mkListObjectsResponse

Arguments

:: Bool
lorIsTruncated

-> BucketName
lorName

-> ListObjectsResponse

Smart constructor for the minimum required parameters to construct a valid ListObjectsResponse response.

This constructor is provided for convenience and testing purposes.

Response lenses

lorIsTruncated :: Lens' ListObjectsResponse Bool

A flag that indicates whether or not Amazon S3 returned all of the results that satisfied the search criteria.

lorMarker :: Lens' ListObjectsResponse (Maybe Text)

lorNextMarker :: Lens' ListObjectsResponse (Maybe Text)

When response is truncated (the IsTruncated element value in the response is true), you can use the key name in this field as marker in the subsequent request to get next set of objects. Amazon S3 lists objects in alphabetical order Note: This element is returned only if you have delimiter request parameter specified. If response does not include the NextMaker and it is truncated, you can use the value of the last Key in the response as the marker in the subsequent request to get the next set of object keys.

lorContents :: Lens' ListObjectsResponse [Object]

lorName :: Lens' ListObjectsResponse BucketName

lorPrefix :: Lens' ListObjectsResponse (Maybe Text)

lorMaxKeys :: Lens' ListObjectsResponse (Maybe Integer)

lorCommonPrefixes :: Lens' ListObjectsResponse [CommonPrefix]

lorEncodingType :: Lens' ListObjectsResponse (Maybe EncodingType)

Encoding type used by Amazon S3 to encode object keys in the response.